% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{AR1}
\alias{AR1}
\title{AR1 prior}
\usage{
AR1(min = 0.8, max = 0.98, s = 1)
}
\arguments{
\item{min, max}{Minimum and maximum values
for autocorrelation parameter (\eqn{\phi}).
Default to 0.8 and 0.98.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\sigma}).
Defaults to 1.}
}
\value{
An object of class \code{bage_prior_ar1}.
}
\description{
Autoregression prior.
}
\details{
The model is
\deqn{x_0 \sim \text{N}(0, \sigma^2)}
\deqn{x_i = \phi x_{i-1} + \sqrt{1 - \phi^2}\epsilon_i}
\deqn{\epsilon \sim \text{N}(0, \sigma^2)}

\eqn{\sigma} is drawn from a half-normal distribition
with scale set by the \code{s} argument.

Correlation parameter \eqn{\phi} is constrained
to lie in the interval \verb{(a, b)},
where \eqn{a} = \code{min} and \eqn{b} = \code{max}.
The prior distribution is for \eqn{\phi}
is
\deqn{\phi = (b - a) \phi' - a}
where
\deqn{\phi' \sim \text{beta}(2, 2)}.
}
\examples{
AR1()
AR1(min = 0, max = 1, s = 2.4)
}
\references{
TMB documentation of
\href{http://kaskr.github.io/adcomp/classdensity_1_1AR1__t.html#details}{AR1}
}
\seealso{
\code{\link[=N]{N()}}, \code{\link[=RW]{RW()}}, \code{\link[=RW2]{RW2()}}, \code{\link[=Known]{Known()}}.
The values for \code{min} and \code{max} are based on the
defaults for function \code{forecast::ets()}.
}
