% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidier.R
\name{augment}
\alias{augment}
\title{Data and values from a fitted model}
\usage{
augment(x, interval, ...)
}
\arguments{
\item{x}{A fitted \code{bage_mod} object.}

\item{interval}{Width of credible interval.}

\item{...}{Unused. Included for generic consistency only.}
}
\value{
A \link[tibble:tibble-package]{tibble},
consisting of the data, plus extra columns.
}
\description{
Data and values from a fitted model
}
\section{Warning}{

The tidymodels
\href{https://www.tidymodels.org/learn/develop/broom/}{website}
states that the augment function may change soon.
The method here will be updated to match the new interface.
}

\examples{
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)
mod <- fit(mod)
tidy(mod)
}
