% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-methods.R
\name{components.bage_mod}
\alias{components.bage_mod}
\title{Extract components from a fitted model}
\usage{
\method{components}{bage_mod}(object, type = c("par", "hyper"), ...)
}
\arguments{
\item{object}{An fitted model.}

\item{type}{Type of component to extract.
\code{"par"} or \code{"hyper"}.}

\item{...}{Not currently used.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Extract components froma fitted object
of class \code{bage_mod}.
}
\details{
The component extracted depends on the \code{type} argument:
\itemize{
\item \code{"par"} Intercept, main effects, and interactions
\item \code{"hyper"} Hyper-parameters from priors for intercept,
main effects, and interactions
}
}
\examples{
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn) |>
  fit()
mod |> components("par")
mod |> components("hyper")
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[=tidy]{tidy()}}
}
