% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_ssvd-methods.R
\name{components.bage_ssvd}
\alias{components.bage_ssvd}
\title{Extract Components used by SVD Prior}
\usage{
\method{components}{bage_ssvd}(object, n_comp = NULL, joint = NULL, age_labels = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"bage_ssvd"}.}

\item{n_comp}{The number of components.
The default is half the total number of
components of \code{object}.}

\item{joint}{Whether to use joint or
independent SVDs for each sex/gender. If
no value is supplied, an SVD for all sexes/genders
combined is used.}

\item{age_labels}{Age labels for the
desired profile. If no labels are supplied, the
most detailed profile available is used.}

\item{...}{Not currently used.}
}
\value{
A named list with elements
\itemize{
\item \code{matrix} A matrix with \code{n_comp} columns.
\item `offset A numeric vector.
}
}
\description{
Extract the matrix and offset used by an SVD prior.
}
\section{Mathematical details}{


If \eqn{\pmb{F}} is the matrix and \eqn{\pmb{g}}
is the offset returned by \code{components()},
then an \code{\link[=SVD]{SVD()}} or \code{\link[=SVDS]{SVDS()}} prior
is generated using
\deqn{\pmb{beta} = \pmb{F} \pmb{z} + \pmb{g}}
where \eqn{\pmb{z}} is vector of standard normal
variates. More complicated SVD-based priors use
more complicated versions of \eqn{\pmb{z}}.
}

\section{Scaled SVDs currently implemented in bage}{

\itemize{
\item \code{\link{HMD}} Mortality rates from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item \code{\link{LFP}} Labour forcce participation
rates from the \href{https://data-explorer.oecd.org}{OECD}.
}
}

\examples{
## components for females and males combined
comp_total <- components(LFP, n_comp = 3)
comp_total$matrix
comp_total$offset

## joint model for females and males
comp_joint <- components(LFP, joint = TRUE, n_comp = 3)
comp_joint$matrix
comp_joint$offset

## females and males modelled independently
comp_indep <- components(LFP, joint = FALSE, n_comp = 3)
comp_indep$matrix
comp_indep$offset

## specify age groups
labels <- poputils::age_labels(type = "five", min = 15, max = 60)
comp_labels <- components(LFP, age_labels = labels)
comp_labels$matrix
comp_labels$offset
}
\seealso{
\itemize{
\item \link[=generate]{generate()} Randomly generate
age or age-sex profiles.
\item \code{\link[=SVD]{SVD()}} SVD prior for age main effect.
\item \code{\link[=SVDS]{SVDS()}} SVD prior for interaction between
age and sex or gender.
\item \code{\link[poputils:age_labels]{poputils::age_labels()}} Generate age labels.
}
}
