% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidential.R
\name{confidential}
\alias{confidential}
\title{Confidentialization}
\description{
The models for rates, probabilities, or means
created with functions \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}},
and \code{\link[=mod_norm]{mod_norm()}} can be extended by adding
descriptions of confidentalization procedures applied
to the outcome variable.
}
\details{
\strong{Data models for outcome variable}\tabular{lllll}{
   \strong{Function} \tab \strong{Confidentialization procedure} \tab \strong{pois} \tab \strong{binom} \tab \strong{norm} \cr
   \code{\link[=set_confidential_rr3]{set_confidential_rr3()}} \tab Outcome randomly rounded to multiple of 3 \tab Yes \tab Yes \tab No \cr
}
}
