% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible_intervals.R
\name{credible_intervals}
\alias{credible_intervals}
\title{Calculate credible intervals and medians.}
\usage{
credible_intervals(classif_vars, measure_var, width = 0.95, na_rm = FALSE)
}
\arguments{
\item{classif_vars}{A data frame with classification
variable(s).}

\item{measure_var}{A numeric vector with measurements.}

\item{width}{Width(s) of credible interval(s).
A numeric vector, with values greater than 0
and less than or equal to 1.}

\item{na_rm}{Whether to remove \code{NA}s
from \code{measure_var} before calculating
quantiles.}
}
\value{
A data frame with \code{nrow(unique(classif_vars))}
rows and \code{ncol(classif_vars)+4} columns.
}
\description{
Calculate credible intervals and medians
for a measure variable,
within every combination of  the
classification variables. Output from \code{credible_intervals}
follows the 'simple names' format of function
\code{point_interval} in package \code{ggdist},
and should work with \code{ggdist} plotting functions.
\code{credible_intervals} is less flexible but faster
than \code{point_interval}.
}
\details{
\code{length(measure_var)} should equal
\code{nrow(classif_vars)}. 
\code{classif_vars} should contain multiple
copies of each combination of the classifying
variables. Each copy typically represents a
draw from a simulation.

\code{credible_intervals} does not assume that
\code{classif_vars} and \code{measure_var}
have been sorted.
}
\seealso{
Quantiles are calculated using
function \code{\link[matrixStats]{rowQuantiles}},
which is very fast.
}
