% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CSA}
\alias{CSA}
\title{Scaled SVD Components from Census School Attendance Data}
\format{
Object of class \code{"bage_ssvd"}.
}
\source{
Derived from data in the
"Population 5 to 24 years of age by school
attendance, sex and urban/rural residence" table in
from the "Population Censuses' Datasets"
database assembled by the United Nations
Statistics Division.
Code to create \code{CSA}
is in folder 'data-raw/ssvd_csa'
in the source code for the \pkg{bage}
package.
}
\usage{
CSA
}
\description{
An object of class \code{"bage_ssvd"}
holding scaled SVD components derived from
census data on school attendance.
The attendance data is assembed by
by the United Nations Statistics Division,
and presented in the
"Population 5 to 24 years of age by school
attendance, sex and urban/rural residence" table in
from the
\href{https://unstats.un.org/unsd/demographic-social/products/dyb/index.cshtml#censusdatasets}{Population Censuses' Datasets}.
\code{CSA} holds 5 components.
}
\details{
\strong{Versions:}
\itemize{
\item \code{"v2025"} (default) version of data
downloadd on 5 November 2025
}
}
\section{Warning}{


Compared other demographic processes, such as
mortality, age-sex patterns in school attendance
have more variation across populations. More
components may be needed to obtain satisfactory
representations of actual age-sex patterns
for school attendance than for other processes,
and even then the accuracy may not be as high.
}

\seealso{
\itemize{
\item \link[=svds]{Scaled SVDs} Overview of scaled SVDs
implemented in \pkg{bage}
}
}
\keyword{datasets}
