% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{Datasets}
\description{
Datasets in \pkg{bage}
}
\details{
\tabular{llll}{
   \strong{dataset} \tab \strong{Outcome} \tab \strong{Variables} \tab \strong{Country} \cr
   \link{isl_deaths} \tab Deaths \tab \code{age}, \code{sex}, \code{time}, \code{deaths}, \code{popn} \tab Iceland \cr
   \link{kor_births} \tab Births \tab \code{age}, \code{region}, \code{time}, \code{popn}, \code{gdp_pc_2023}, \code{dens2020} \tab South Korea \cr
   \link{nld_expenditure} \tab Health expenditure \tab \code{diag}, \code{age}, \code{year}, \code{value} \tab Netherlands \cr
   \link{nzl_divorces} \tab Divorces \tab \code{age}, \code{sex}, \code{time}, \code{divorces}, \code{population} \tab New Zealand \cr
   \link{nzl_households} \tab One-person households \tab \code{age}, \code{region}, \code{year}, \code{oneperson}, \code{total} \tab New Zealand \cr
   \link{nzl_injuries} \tab Accidental deaths \tab \code{age}, \code{sex}, \code{ethnicity}, \code{year}, \code{injuries}, \code{popn} \tab New Zealand \cr
   \link{prt_deaths} \tab Deaths \tab \code{age}, \code{time}, \code{deaths}, \code{exposure} \tab Portugal \cr
   \link{swe_infant} \tab Infant deaths \tab \code{county}, \code{time}, \code{births}, \code{deaths} \tab Sweden \cr
   \link{usa_deaths} \tab Accidental deaths \tab \code{month}, \code{deaths} \tab United States \cr
}
}
