% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{EAR}
\alias{EAR}
\title{Exchangeable Autoregressive Prior}
\usage{
EAR(n = 2, s = 1, along = NULL)
}
\arguments{
\item{n}{The order of the model.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\sigma}).
Defaults to 1.}

\item{along}{Name of one of the dimensions
in the interaction. Optional, provided
the data contain a time or age dimension.}
}
\value{
An object of class \code{bage_prior_ear}.
}
\description{
Prior for an interaction,
where an autoregression model
of order \code{n} is applied
to the "along" variable, within each
combination of values of the "by" variable.
The damping coefficients are shared across
different combinations of the "by" variables.
The time series within each combination of the
'by' variables are, however, treated as exchangeable.
}
\section{'Along' and 'by' variables}{


Multivariate priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the by variables are likely to be region and sex.

If no \code{along} argument is supplied, then:
\itemize{
\item if the interaction contains a time variable, then
it is assumed to be the 'along' variable;
\item otherwise, if the interaction contains an
age variable, then it is assumed to be
the 'along' variable;
\item otherwise, an error is raised.
}
}

\section{Mathematical description}{


The model is

\deqn{x_{u,v} = \phi_1 x_{u,i-1} + \cdots + \phi_k x_{u,i-k} + \epsilon_{u,v}}
\deqn{\epsilon_{u,v} \sim \text{N}(0, \omega^2)}

where \eqn{\omega} is chosen so that each \eqn{x_{u,v}} has
marginal variance \eqn{\sigma^2}. The value of
\eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

The \eqn{\phi_j} are constrained to the interval
between -1 and 1.
}

\examples{
EAR(n = 3)
EAR(n = 3, s = 2.4)
EAR(along = "cohort")
}
\references{
TMB documentation for
\href{http://kaskr.github.io/adcomp/classdensity_1_1ARk__t.html#details}{ARk}
}
\seealso{
\code{\link[=N]{N()}}, \code{\link[=RW]{RW()}}, \code{\link[=RW2]{RW2()}}, \code{\link[=Known]{Known()}}.
The values for \code{min} and \code{max} are based on the
defaults for function \code{forecast::ets()}.
\code{\link[=AR1]{AR1()}}
}
