% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{EAR1}
\alias{EAR1}
\title{Exchangeable AR1 Prior}
\usage{
EAR1(min = 0.8, max = 0.98, s = 1, along = NULL)
}
\arguments{
\item{min, max}{Minimum and maximum values
for autocorrelation parameter (\eqn{\phi}).
Defaults are \code{0.8} and \code{0.98}.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\tau}).
Defaults to 1.}

\item{along}{Name of one of the dimensions
in the interaction. Optional, provided
the data contain a time or age dimension.}
}
\value{
An object of class \code{"bage_prior_ear"}.
}
\description{
An exchangeable AR1 prior
is used with interactions. An autoregressive
model of order \eqn{k} is applied to the 'along'
variable, within each combination of values of the 'by' variables.
The damping coefficients are shared across
different combinations of the 'by' variables.
The series within each combination of the
'by' variables are treated as exchangeable.
}
\section{Mathematical details}{


If \eqn{\beta_{u,v}} is the \eqn{v}th element of the interaction
within the \eqn{u}th combination of the
'by' variables, then

\deqn{x_{u,v} = \phi x_{u,v-1} + \epsilon_{u,v}}
\deqn{\epsilon_{u,v} \sim \text{N}(0, \omega^2)}

where \eqn{\omega} is chosen so that each \eqn{x_{u,v}} has
marginal variance \eqn{\tau^2}. The value of
\eqn{\tau} has prior

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

Correlation parameter \eqn{\phi} is constrained
to lie in the interval \verb{(\\text\{min\}, \\text\{max\})}.
The prior distribution for \eqn{\phi} is

\deqn{\phi = (\text{max} - \text{min}) \phi' - \text{min}}

where

\deqn{\phi' \sim \text{beta}(2, 2)}.
}

\section{'Along' and 'by' variables}{


Some priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the 'along' variable is likely to be time,
and the 'by' variables are likely to be region and sex.

If no \code{along} argument is supplied, a default value
is chosen as follows:
\itemize{
\item if the interaction contains a time variable, then
the 'along' variable is assumed to be time;
\item otherwise, if the interaction contains an
age variable, then the 'along' variable
is assumed to be age;
\item otherwise, an error is raised.
}
}

\examples{
EAR1()
EAR1(min = 0, max = 1, s = 2.4)
EAR1(along = "cohort")
}
\references{
TMB documentation for
\href{http://kaskr.github.io/adcomp/classdensity_1_1ARk__t.html#details}{ARk}
}
\seealso{
\itemize{
\item \code{\link[=EAR]{EAR()}} More general exchangeable AR model.
\item \code{\link[=AR1]{AR1()}} AR1 prior for main effects.
\item \link{priors} Overview of priors implemented in \code{bage}.
\item The values for \code{min} and \code{max} are based on the
defaults for function \code{forecast::ets()}.
}
}
