% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{ELin}
\alias{ELin}
\title{Exchangeable Linear Prior}
\usage{
ELin(s = 1, sd = 1, ms = 1, along = NULL)
}
\arguments{
\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\tau}).
Defaults to 1.}

\item{sd}{A postive number. Default is 1.}

\item{ms}{A positive number. Default is 1.}

\item{along}{Name of one of the dimensions
in the interaction. Optional, provided
the data contain a time or age dimension.}
}
\value{
An object of class \code{"bage_prior_elin"}.
}
\description{
Prior for an interaction where, within each combination
of the 'by' variables, the elements of the 'along'
variable follow a straight line, with idiosyncratic
errors around that line.
Each combination of the 'by' variables has a different
line, with a different slope. The slopes are drawn
from a common distribution.
}
\section{Mathematical details}{


If \eqn{\beta_{u,v}} is the \eqn{v}th element of the interaction
within the \eqn{u}th combination of the
'by' variables, then

\deqn{\beta_{u,v} \sim \text{N}(\eta_u q_v, \tau^2)}

where

\deqn{q_v = - (V+1)/(V-1) + 2v/(V-1).}

The slopes \eqn{\eta_u} are drawn from a common distribution,

\deqn{\eta_u \sim \text{N}(\eta, \omega^2).}

The absolute value of mean \eqn{\eta}
is governed by parameter \code{s}:

\deqn{\eta \sim \text{N}(0, \text{s}^2).}

Larger values for \eqn{\omega} imply more variability
in slopes across different combinations of the 'by'
variables. The size of \eqn{\omega} is governed by
parameter \code{ms},

\deqn{\omega \sim \text{N}^+(0, (\text{ms})^2)}

Larger values for \eqn{\tau} imply more variability
around each line. The size of \eqn{\tau} is
governed by parameter \code{s}:

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

(\eqn{\text{N}^+} denotes a half-normal distribution,
which has the same shape as a normal
distribution, but is defined only for non-negative
values.)
}

\section{'Along' and 'by' variables}{


Some priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the 'along' variable is likely to be time,
and the 'by' variables are likely to be region and sex.

If no \code{along} argument is supplied, a default value
is chosen as follows:
\itemize{
\item if the interaction contains a time variable, then
the 'along' variable is assumed to be time;
\item otherwise, if the interaction contains an
age variable, then the 'along' variable
is assumed to be age;
\item otherwise, an error is raised.
}
}

\examples{
ELin()
ELin(s = 0.5, sd = 2, ms = 0.1, along = "cohort")
}
\seealso{
\itemize{
\item \code{\link[=Lin]{Lin()}} Linear prior for main effects.
\item \link{priors} Overview of priors implemented in \code{bage}.
}
}
