% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{ERW}
\alias{ERW}
\title{Exchangeable Random Walk Prior}
\usage{
ERW(s = 1, along = NULL)
}
\arguments{
\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\sigma}).
Defaults to 1.}

\item{along}{Name of one of the dimensions
in the interaction. Optional, provided
the data contain a time or age dimension.}
}
\value{
An object of class \code{bage_prior_erw}.
}
\description{
Prior for an interaction,
where a (first order) random walk model
is applied to the 'along' variable, within each
combination of values of the 'by' variables.
Standard deviations are shared across
different combinations of the 'by' variables.
The series within each combination of the
'by' variables are treated as exchangeable.
}
\section{Mathematical details}{


If \eqn{\beta_{u,v}} is the \eqn{v}th element of the interaction
within the \eqn{u}th combination of the
'by' variables, then

\deqn{x_{u,v} = x_{u,v-1} + \epsilon_{u,v}}
\deqn{\epsilon_{u,v} \sim \text{N}(0, \tau^2)}

Standard deviation \eqn{\tau} is drawn from a
half-normal distribution,

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{s} lead to smoother series.
}

\section{'Along' and 'by' variables}{


Some priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the 'along' variable is likely to be time,
and the 'by' variables are likely to be region and sex.

If no \code{along} argument is supplied, a default value
is chosen as follows:
\itemize{
\item if the interaction contains a time variable, then
the 'along' variable is assumed to be time;
\item otherwise, if the interaction contains an
age variable, then the 'along' variable
is assumed to be age;
\item otherwise, an error is raised.
}
}

\examples{
ERW()
ERW(s = 0.5)
}
\seealso{
\itemize{
\item \code{\link[=RW]{RW()}} Random walk prior for main effects.
\item \link{priors} Overview of priors implemented in \code{bage}.
}
}
