% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{ESeas}
\alias{ESeas}
\title{Exchangeable Seasonal Prior}
\usage{
ESeas(n, s = 1, along = NULL)
}
\arguments{
\item{n}{Number of seasons.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\tau}).
Defaults to 1.}

\item{along}{Name of one of the dimensions
in the interaction. Optional, provided
the data contain a time or age dimension.}
}
\value{
An object of class \code{bage_prior_eseas}.
}
\description{
Prior for a seasonal effect, within an interaction.
Each combination of the 'by' variables,
has a separate seasonal effect.
The \code{along} variable is almost always
time.
}
\section{Mathematical details}{


If \eqn{\beta_{u,v}^{\text{seas}}} is the \eqn{v}th
element of the seasonal effect within
within the \eqn{u}th combination of the
'by' variables, then

\deqn{\beta_{u,v}^{\text{seas}} \sim \text{N}(\beta_{u,v-n}^{\text{seas}}, \tau^2),}

where \eqn{n} is the number of seasons.

Larger values for \eqn{\tau} imply more variability
in seasonal effects over time. The size of \eqn{\tau} is
governed by parameter \code{s}:

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

(\eqn{\text{N}^+} denotes a half-normal distribution,
which has the same shape as a normal
distribution, but is defined only for non-negative
values.)
}

\section{'Along' and 'by' variables}{


Some priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the 'along' variable is likely to be time,
and the 'by' variables are likely to be region and sex.

If no \code{along} argument is supplied, a default value
is chosen as follows:
\itemize{
\item if the interaction contains a time variable, then
the 'along' variable is assumed to be time;
\item otherwise, if the interaction contains an
age variable, then the 'along' variable
is assumed to be age;
\item otherwise, an error is raised.
}
}

\examples{
ESeas(n = 4)
ESeas(n = 12, s = 0.5, along = "cohort")
}
\seealso{
\itemize{
\item \code{\link[=Seas]{Seas()}} Seasonal effect for main effects.
\item Seasonal effects are always
used within 'composite' priors, created using
\code{\link[=compose_time]{compose_time()}}.
\item \link{priors} Overview of priors implemented in \strong{bage}
}
}
