% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{ESVD}
\alias{ESVD}
\title{Exchangeable SVD Prior for Interactions Involving Age}
\usage{
ESVD(ssvd, n = 5)
}
\arguments{
\item{ssvd}{Object of class \code{"bage_ssvd"}
holding results from a scaled SVD.
See below for current options.}

\item{n}{Number of vectors from scaled SVD
to use in modelling.
Default is half the number of components of \code{ssvd}.}
}
\value{
An object of class \code{"bage_prior_esvd"}.
}
\description{
Specify a Singular Value Decomposition (SVD) prior
for interactions involving age but not sex/gender.
Age profiles are "exchangeable" in that
profiles for all combinations of the non-age variables
are drawn from the same distribution.
The age profiles are assumed to look like
they were drawn at random
from an external demographic database. For instance,

\if{html}{\out{<div class="sourceCode">}}\preformatted{set_prior(mod, age ~ ESVD(HMD))
}\if{html}{\out{</div>}}

specifies that the age profiles should look like they
were drawn from the
\href{https://www.mortality.org}{Human Mortality Database}.
}
\section{Mathematical details}{


With an ESVD prior,

\deqn{\pmb{\beta}_u = \pmb{F} \pmb{\alpha}_u + \pmb{g}}

where
\itemize{
\item \eqn{\pmb{\beta}}_u is the age profile associated with
combination \eqn{u} of the non-age variables;
\item \eqn{\pmb{X}} is a known \eqn{V \times K} matrix, where
\eqn{V} is the number of age groups, and \eqn{K} is much
less than \eqn{V};
\item \eqn{\pmb{\alpha}}_u is a vector with \eqn{V} elements; and
\item \eqn{\pmb{g}} is a known vector with \eqn{V} elements.
}

The elements of \eqn{\pmb{\alpha}}_u have prior

\deqn{\alpha_{uk} \sim \text{N}(0, 1), \quad i = 1, \cdots, K}

\eqn{\pmb{F}} and \eqn{\pmb{g}} are constructed from
a large database of age-specific demographic estimates
by preforming an SVD and then extracting means and variances.
For details see TODO - REFERENCE TO VIGNETTE.
}

\section{Scaled SVDs currently implemented in bage}{

\itemize{
\item \code{\link{HMD}} Values from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item TODO - add others
}
}

\examples{
ESVD(HMD)
ESVD(HMD, n = 3)
}
\seealso{
\itemize{
\item \code{\link[=set_prior]{set_prior()}} Specify the prior for a model
\item \code{\link[=SVD]{SVD()}} SVD prior for age main effect.
\item \code{\link[=SVDS]{SVDS()}} SVD prior for interaction between
age and sex/gender.
\item \code{\link[=ESVDS]{ESVDS()}} Exchangeable SVD prior for interaction involving
age, sex/gender, and other dimension(s).
\item XXX Create a new scaled SVD.
}
}
