% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{ESVDS}
\alias{ESVDS}
\title{Exchangeable SVD Prior for Interactions Involving Age
and Sex/Gender}
\usage{
ESVDS(ssvd, n = 5, joint = FALSE)
}
\arguments{
\item{ssvd}{Object of class \code{"bage_ssvd"}
holding results from a scaled SVD.
See below for current options.}

\item{n}{Number of components from scaled SVD
to use in modelling. The default is half
the number of components of \code{ssvd}.}

\item{joint}{Whether to use combined or
separate SVDs. Default is \code{FALSE}.
See below for details.}
}
\value{
An object of class \code{"bage_prior_esvd"}.
}
\description{
Specify a Singular Value Decomposition (SVD) prior
for an interaction involving age, sex/gender,
and at least one other dimension.
Age-sex profiles are "exchangeable", in that
profiles for all combinations of the
non-age, non-sex variables are drawn
from the same distribution.
The age-sex profiles are assumed to look like they
were drawn at random
from an external demographic database. For instance,

\if{html}{\out{<div class="sourceCode">}}\preformatted{set_prior(mod, age:sex ~ ESVDS(HMD))
}\if{html}{\out{</div>}}

specifies that age-sex profiles should look
like they was drawn from the
\href{https://www.mortality.org}{Human Mortality Database}.
}
\section{Mathematical details}{


Let \eqn{V} denote the number of age groups, \eqn{S} the
number of sexes/genders, and \eqn{U} the number of combinations
of non-age, non-sex variables.
When \code{joint} is \code{FALSE},

\deqn{\pmb{\beta}_{us} = \pmb{F}_s \pmb{\alpha}_{us} + \pmb{g}_s,
  \quad, u = 1, \cdots, U, s = 1, \cdots S}

where \eqn{\pmb{\beta}_{us}}, \eqn{\pmb{\alpha}_{us}},
and \eqn{\pmb{g}_{us}} each have \eqn{V}
elements, and \eqn{\pmb{F}} has \eqn{V} rows.

When \code{joint} is \code{TRUE},

\deqn{\pmb{\beta}_u = \pmb{F} \pmb{\alpha}_u + \pmb{g}}

where

where \eqn{\pmb{\beta}_u}, \eqn{\pmb{\alpha}_u},
and \eqn{\pmb{g}} each have \eqn{V \times S}
elements, and \eqn{\pmb{F}} has \eqn{V \times S} rows.

\eqn{\pmb{F}_s}, \eqn{\pmb{g}_s}, \eqn{\pmb{F}}, and \eqn{\pmb{g}},
are all estimated externally, using estimates from a
demographic database, and are treated as fixed and known.

The elements of \eqn{\pmb{\alpha}_{us}} and \eqn{\pmb{\alpha}_u} have prior

\deqn{\alpha_{uk} \sim \text{N}(0, 1)}

For a description of the construction of \eqn{\pmb{F}_s},
\eqn{\pmb{F}}, \eqn{\pmb{g}_s}, and \eqn{\pmb{g}},
see TODO - REFERENCE TO VIGNETTE.
}

\section{Joint or independent SVDs}{


To possible ways of extracting patterns
from age-sex-specific data are
\enumerate{
\item carry out separate SVDs on separate datasets for
each sex/gender; or
\item carry out a single SVD on dataset that has separate
entries for each sex/gender.
}

Option 1 is more flexible. Option 2 is
more robust to sampling or measurement errors.
Option 1 is obtained by setting the \code{joint}
argument  to \code{FALSE}. Option 2
is obtained by setting the \code{joint} argument to \code{TRUE}.
The default is \code{FALSE}.
}

\section{Scaled SVDs currently implemented in bage}{

\itemize{
\item \code{\link{HMD}} Values from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item TODO - add others
}
}

\examples{
ESVDS(HMD)
ESVDS(HMD, joint = TRUE)
}
\seealso{
\itemize{
\item \code{\link[=set_prior]{set_prior()}} Specify the prior for a model
\item \code{\link[=SVD]{SVD()}} SVD prior for age main effect.
\item \code{\link[=SVDS]{SVDS()}} SVD prior for interaction between
age and sex/gender.
\item \code{\link[=ESVD]{ESVD()}} Exchangeable SVD prior for interaction involving
age and other dimensions, but not sex/gender.
}
}
