% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poputils.R
\name{find_age}
\alias{find_age}
\title{Identify an age variable}
\usage{
find_age(nms)
}
\arguments{
\item{nms}{A character vector}
}
\value{
An element of \code{nms} or \code{NULL}.
}
\description{
Find the element of \code{nms} that looks like an age variable.
If no elements look like an age variable, or if
two or more elements do,
then return \code{NULL}.
}
\examples{
find_age(c("Sex", "Year", "AgeGroup", NA)) ## one valid
find_age(c("Sex", "Year"))                 ## none valid
find_age(c("age", "age.years"))            ## two valid
}
\seealso{
\code{\link[=find_time]{find_time()}}, \code{\link[=find_sexgender]{find_sexgender()}}
}
