% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poputils.R
\name{find_female}
\alias{find_female}
\title{Identify female sex or gender labels}
\usage{
find_female(nms)
}
\arguments{
\item{nms}{A character vector}
}
\value{
An element of \code{nms} or \code{NULL}.
}
\description{
Given labels for sex or gender, try to infer
which (if any) refer to females.
If no elements look like a label for females,
or if two or more elements do,
then return \code{NULL}.
}
\examples{
find_female(c("Female", "Male")) ## one valid
find_female(c("0-4", "5-9"))     ## none valid
find_female(c("F", "Fem"))       ## two valid
}
\seealso{
\code{\link[=find_male]{find_male()}}, \code{\link[=find_sexgender]{find_sexgender()}}
}
