% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poputils.R
\name{find_male}
\alias{find_male}
\title{Identify male sex or gender labels}
\usage{
find_male(nms)
}
\arguments{
\item{nms}{A character vector}
}
\value{
An element of \code{nms} or \code{NULL}.
}
\description{
Given labels for sex or gender, try to infer
which (if any) refer to males.
If no elements look like a label for males,
or if two or more elements do,
then return \code{NULL}.
}
\examples{
find_male(c("Female", "Male")) ## one valid
find_male(c("0-4", "5-9"))     ## none valid
find_male(c("male", "m"))      ## two valid
}
\seealso{
\code{\link[=find_female]{find_female()}}, \code{\link[=find_sexgender]{find_sexgender()}}
}
