% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poputils.R
\name{find_sexgender}
\alias{find_sexgender}
\title{Identify a sex or gender variable}
\usage{
find_sexgender(nms)
}
\arguments{
\item{nms}{A character vector}
}
\value{
An element of \code{nms} or \code{NULL}.
}
\description{
Find the element of \code{nms} that looks like
a sex or gender variable.
If no elements look like a sex or gender variable,
or if two or more elements do,
then return \code{NULL}.
}
\examples{
find_sexgender(c("Sex", "Year", "AgeGroup", NA)) ## one valid
find_sexgender(c("Age", "Region"))               ## none valid
find_sexgender(c("sexgender", "sexes"))          ## two valid
}
\seealso{
\code{\link[=find_age]{find_age()}}, \code{\link[=find_time]{find_time()}}, \code{\link[=find_female]{find_female()}},
\code{\link[=find_male]{find_male()}}
}
