% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poputils.R
\name{find_time}
\alias{find_time}
\title{Identify a time variable}
\usage{
find_time(nms)
}
\arguments{
\item{nms}{A character vector}
}
\value{
An element of \code{nms} or \code{NULL}.
}
\description{
Find the element of \code{nms} that looks like an time variable.
If no elements look like a time variable, or if
two or more elements do,
then return \code{NULL}.
}
\examples{
find_time(c("Sex", "Year", "AgeGroup", NA)) ## one valid
find_time(c("Sex", "Region"))               ## none valid
find_time(c("time", "year"))                ## two valid
}
\seealso{
\code{\link[=find_age]{find_age()}}, \code{\link[=find_sexgender]{find_sexgender()}}
}
