% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-methods.R
\name{generate.bage_prior_norm}
\alias{generate.bage_prior_norm}
\alias{generate.bage_prior_rw}
\title{Generate Values from Priors}
\usage{
\method{generate}{bage_prior_norm}(x, n = 20, n_draw = 25, standardize = c("terms", "none"))

\method{generate}{bage_prior_rw}(x, n = 20, n_draw = 25, standardize = c("terms", "none"))
}
\arguments{
\item{x}{Object of class \code{"bage_prior"}}

\item{n}{Number of elements in term.
Default is \code{20}.}

\item{n_draw}{Number of draws. Default
is \code{25}.}

\item{standardize.}{Whether to standardize
the values for the effect. If \code{"terms"} (the default),
values are centered.}
}
\value{
A \link[tibble:tibble]{tibble}
}
\description{
Generate draws from priors for model terms.
The terms are assumed to be main effects,
except in the case of dynamic SVD priors
(eg \code{\link[=SVD_RW]{SVD_RW()}}) where they assumed to
be age-time interactions.
}
\examples{
rw <- RW()
generate(rw)
}
\seealso{
\itemize{
\item \link{priors} Overview of priors implemented in \strong{bage}
\item \code{\link[=augment]{augment()}}, \code{\link[=compose]{compose()}} Draw from posterior or prior
distribution for whole model
}
}
