% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_ssvd-methods.R
\name{generate.bage_ssvd}
\alias{generate.bage_ssvd}
\title{Generate Random Age or Age-Sex Profiles}
\usage{
\method{generate}{bage_ssvd}(x, n_draw = 20, n_comp = NULL, joint = NULL, age_labels = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"bage_ssvd"}.}

\item{n_draw}{Number of random draws to generate.}

\item{n_comp}{The number of components
to use when generating profiles. Default is
half the number of components of \code{x}.}

\item{joint}{Whether to use joint or
independent SVDs for each sex/gender. If
no value is supplied, an SVD for all sexes/genders
combined is used.}

\item{age_labels}{Age labels for the
desired profile. If no labels are supplied, the
most detailed profile available is used.}

\item{...}{Not currently used.}
}
\value{
A tibble
}
\description{
Generate random age or age-sex profiles from
an object of class \code{"bage_ssvd"}. An object
of class \code{"bage_ssvd"} holds results from
an \link[base:svd]{SVD} decomposition of demographic
data.
}
\section{Scaled SVDs currently implemented in bage}{

\itemize{
\item \code{\link{HMD}} Mortality rates from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item \code{\link{LFP}} Labour forcce participation
rates from the \href{https://data-explorer.oecd.org}{OECD}.
}
}

\examples{
## SVD for females and males combined
generate(HMD)

## separate SVDs for females and males
generate(HMD, joint = FALSE) 

## specify age groups
labels <- poputils::age_labels(type = "lt", max = 60)
generate(HMD, age_labels = labels)
}
\seealso{
\itemize{
\item \code{\link[=SVD]{SVD()}} SVD prior for age main effect.
\item \code{\link[=SVDS]{SVDS()}} SVD prior for interaction between
age and sex or gender.
\item \code{\link[poputils:age_labels]{poputils::age_labels()}} Generate age labels.
}
}
