% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HIMD_R}
\alias{HIMD_R}
\alias{HIMD_P1}
\alias{HIMD_P5}
\title{Scaled SVD Components from Human
Internal Migration Database}
\format{
An object of class bage_ssvd of length 1.

An object of class \code{bage_ssvd} of length 1.

An object of class \code{bage_ssvd} of length 1.
}
\source{
Dyrting, S. (2024, October 23).
Data from: Estimating Complete Migration Probabilities
from Grouped Data. Retrieved from osf.io/vmrfk
on 1 September 2025.
}
\usage{
HIMD_R

HIMD_P1

HIMD_P5
}
\description{
Objects of class \code{"bage_ssvd"}
holding scaled SVD components derived from
data from the
\href{https://osf.io/vmrfk/}{Human Internal Migration Database}.
Each object holds 5 components
}
\details{
\itemize{
\item \code{HIMD_P1} is derived from data on
1-year migration probabilities, ie the
probability that a person will migrate
during an interval of 1 year.
\item \code{HIMD_P5} is derived from data on
5-year migration probabilities, ie the
probability that a person will migrate
during an interval of 5 years.
\item \code{HIMD_R} is derived from data on 1-year
migration probabilities, using the
formula \eqn{r = \log(1 - p)}.
}

\strong{Versions:}
\itemize{
\item \code{"v2024"} Data from 23 October 2024 (the default)
}
}
\keyword{datasets}
