% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-methods.R
\name{is_fitted}
\alias{is_fitted}
\title{Test whether a model has been fitted}
\usage{
is_fitted(x)
}
\arguments{
\item{x}{A model object.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Test whether \link[=fit.bage_mod]{fit()} has been
called on a model object.
}
\examples{
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)
is_fitted(mod)
mod <- fit(mod)
is_fitted(mod)
}
