% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{ISeas}
\alias{ISeas}
\title{Independent Seasonal Prior}
\usage{
ISeas(n, s = 1, along = NULL)
}
\arguments{
\item{n}{Number of seasons.}

\item{s}{A positive number. Default is 1.}

\item{along}{Name of one of the dimensions
in the interaction. Optional, provided
the data contain a time or age dimension.}
}
\value{
An object of class \code{bage_prior_iseas}.
}
\description{
Prior for a seasonal effect, within an interaction.
Each combination of the 'by' variables,
has a separate seasonal effect.
The \code{along} variable is almost always
time.
}
\section{Statistical model}{


The model is

\deqn{x_{uv} \sim \text{N}(x_{u,v-n}, \sigma^2)}

where
\itemize{
\item \eqn{u} is the index for a combination of 'by' variables,
\item \eqn{v} is the index of the 'along' variable,
\item \eqn{n} is the number of seasons
}

Larger values for \eqn{\sigma} imply more variability
in seasonal effects over time. The size of \eqn{\sigma} is
governed by parameter \code{s}:

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(\eqn{\text{N}^} denotes a half-normal distribution,
which has the same shape as a normal
distribution, but is defined only for non-negative
values.)
}

\section{'Along' and 'by' variables}{


Multivariate priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the by variables are likely to be region and sex.

If no \code{along} argument is supplied, then:
\itemize{
\item if the interaction contains a time variable, then
it is assumed to be the 'along' variable;
\item otherwise, if the interaction contains an
age variable, then it is assumed to be
the 'along' variable;
\item otherwise, an error is raised.
}
}

\examples{
ISeas(n = 4)
ISeas(n = 12, s = 0.5, along = "cohort")
}
\seealso{
\itemize{
\item \code{\link[=Seas]{Seas()}} etc
\item \code{\link[=set_var_time]{set_var_time()}} to specify the time variable
\item \code{\link[=set_var_age]{set_var_age()}} to specify the age variable
}
}
