% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isl_deaths}
\alias{isl_deaths}
\title{Deaths in Iceland}
\format{
A \link[tibble:tibble-package]{tibble} with 5,300 rows and the following
columns:
\itemize{
\item \code{age} Single year of age, up to \code{"105+"}
\item \code{sex} \code{"Female"} and \code{"Male"}
\item \code{time} Calendar year, 1998-2022
\item \code{deaths} Counts of deaths
\item \code{popn} Mid-year population
}
}
\source{
Tables "Deaths by municipalities, sex and age 1981-2022",
and "Average annual population by municipality,
age and sex 1998-2022 - Current municipalities",
on the Statistics Iceland website. Data downloaded
on 12 July 2023.
}
\usage{
isl_deaths
}
\description{
Deaths and mid-year populations in Iceland,
by age, sex, and calendar year.
}
\seealso{
\itemize{
\item \link{datasets} Overview of datasets in \pkg{bage}
}
}
\keyword{datasets}
