% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{Known}
\alias{Known}
\title{Treat a model term as known}
\usage{
Known(values)
}
\arguments{
\item{values}{A numeric vector}
}
\value{
An object of class \code{bage_prior_known}.
}
\description{
Treat the intercept, a main effect, or an interaction
in an model as known.
}
\examples{
Known(-2.3)
Known(c(0.1, 2, -0.11))

}
\seealso{
\code{Known} is usually called within \code{\link[=set_prior]{set_prior()}}.
}
