% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{kor_births}
\alias{kor_births}
\title{Births in South Korea}
\format{
A \link[tibble:tibble-package]{tibble} with 1,872 rows and the following
columns:
\itemize{
\item \code{age} Five-year age groups from \verb{"10-14" to }"50-54"`
\item \code{region} Administrative region
\item \code{time} Calendar year, 2011-2023
\item \code{births} Counts of births
\item \code{popn} Mid-year population
}
}
\source{
Tables "Live Births by Age Group of Mother,
Sex and Birth Order for Provinces",
and "Resident Population in Five-Year Age Groups",
on the Korean Statistical Information Service website.
Data downloaded on 24 September 2024.
}
\usage{
kor_births
}
\description{
Births by age of mother, region, and calendar year,
2011-2023.
}
\keyword{datasets}
