% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LFP}
\alias{LFP}
\title{Scaled SVD Components from OECD
Labor Force Participation Data}
\format{
Object of class \code{"bage_ssvd"}.
}
\source{
Derived from data in the "Labor Force Indicators"
table of the OECD Data Explorer.
Code to create \code{LFS}
is in folder 'data-raw/ssvd_lfp' in
the source code for the \pkg{bage}
package.
}
\usage{
LFP
}
\description{
An object of class \code{"bage_ssvd"}
holding components extracted from labor force participation
data from the
\href{https://data-explorer.oecd.org}{OECD Data Explorer}.
\code{LFP} holds 5 components.
}
\details{
\strong{Versions:}
\itemize{
\item \code{"v2025"} data downloaded on 2025-10-17
}
}
\seealso{
\itemize{
\item \link[=svds]{Scaled SVDs} Overview of scaled SVDs
implemented in \pkg{bage}
}
}
\keyword{datasets}
