% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lfpr_oecd}
\alias{lfpr_oecd}
\title{Labour force participation rates from the OECD}
\format{
A matrix, in which each row holds a combination
of sex and age, and each column holds a combination
of country and year.
}
\source{
OECD (2022), "Labour Market Statistics:
Labour force statistics by sex and age: indicators",
OECD Employment and Labour Market Statistics (database),
https://doi.org/10.1787/data-00310-en
(accessed on 12 October 2022).
}
\usage{
lfpr_oecd
}
\description{
Estimated labour force participation rates,
is the proportion of the population that is in the
labour force, by age and sex.
}
\keyword{datasets}
