% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{LinAR}
\alias{LinAR}
\title{Linear Prior with Autoregressive Errors}
\usage{
LinAR(n = 2, s = 1, sd = 1)
}
\arguments{
\item{n}{The order of the model for the
errors, i.e. the number of lagged
terms that are included. Default is \code{2}.}

\item{s}{Scale for the prior for the errors.
Smaller values imply a smoother series. Default is 1.}

\item{sd}{Standard deviation in the prior for
the slope of the line. Larger values imply
steeper slopes. Default is 1.}
}
\value{
An object of class \code{"bage_prior_linar"}.
}
\description{
Prior for a main effect where the elements
follow a straight line, with autoregressive
errors around this line.
}
\section{Mathematical details}{


With a \code{LinAR()} prior,

\deqn{\beta_j = \eta q_j + \epsilon_j}

where
\itemize{
\item \eqn{\beta_j} is the \eqn{j}th element of the main effect
\item \eqn{q_j} is defined as
\deqn{q_j = - (J+1)/(J-1) + 2j/(J-1)}
\item \eqn{\epsilon_j} has an autoregressive structure
\deqn{\epsilon_j = \phi_1 \epsilon_{j-1} + \cdots + \phi_n \epsilon_{j-k} + \varepsilon_j}
\deqn{\varepsilon_j \sim \text{N}(0, \omega^2).}
}

The value of \eqn{\omega} is chosen so that each \eqn{\epsilon_j} has
marginal variance \eqn{\tau^2}.

Parameter \eqn{\tau} has a half-normal prior

\deqn{\tau \sim \text{N}^+(0, \text{s}^2),}

where a value for \code{s} is provided by the user.

The \eqn{\phi_1, \cdots, \phi_k} are restricted to values
between -1 and 1 that jointly lead to a stationary model. The quantity
\eqn{r = \sqrt{\phi_1^2 + \cdots + \phi_k^2}} is given a
boundary-avoiding prior

\deqn{r \sim \text{Beta}(2, 2).}

Slope \eqn{\eta} is drawn from a normal distribution,

\deqn{\eta \sim \text{N}(0, (\text{sd})^2).}
}

\examples{
LinAR()
LinAR(n = 3, s = 0.5, sd = 2)
}
\seealso{
\itemize{
\item \code{\link[=LinAR1]{LinAR1()}} Linear prior with AR1 errors
\item \code{\link[=Lin]{Lin()}} Linear prior with independent normal errors
\item \code{\link[=ELin]{ELin()}} Exchangeable version of \code{Lin()},
used with interactions
\item \link{priors} Overview of priors implemented in \strong{bage}
}
}
