% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{LinAR1}
\alias{LinAR1}
\title{Linear Prior with AR1 Errors}
\usage{
LinAR1(min = 0.8, max = 0.98, s = 1, sd = 1)
}
\arguments{
\item{min, max}{Minimum and maximum values
for autocorrelation coefficient (\eqn{\phi}).
Defaults are \code{0.8} and \code{0.98}.}

\item{s}{Scale for the prior for the errors.
Smaller values imply a smoother series. Default is 1.}

\item{sd}{Standard deviation in the prior for
the slope of the line. Larger values imply
steeper slopes. Default is 1.}
}
\value{
An object of class \code{"bage_prior_linar"}.
}
\description{
Prior for a main effect where the elements
follow a straight line, with errors that form
an autoregressive series of order 1.
}
\section{Mathematical details}{


With a \code{LinAR1()} prior,

\deqn{\beta_j = \eta q_j + \epsilon_j}

where
\itemize{
\item \eqn{\beta_j} is the \eqn{j}th element of the main effect
\item \eqn{q_j} is defined as
\deqn{q_j = - (J+1)/(J-1) + 2j/(J-1),}
\item \eqn{\epsilon_j} has an AR1 structure
\deqn{\epsilon_j = \phi \epsilon_{j-1} + \varepsilon_j}
\deqn{\varepsilon_j \sim \text{N}(0, \omega^2).}
}

The value of \eqn{\omega} is chosen so that each \eqn{\epsilon_j} has
marginal variance \eqn{\tau^2}.

Parameter \eqn{\tau} has a half-normal prior

\deqn{\tau \sim \text{N}^+(0, \text{s}^2),}

where a value for \code{s} is provided by the user.

Coefficient \eqn{\phi} is constrained
to lie between \code{min} and \code{max}.
Its prior distribution is

\deqn{\phi = (\text{max} - \text{min}) \phi' - \text{min}}

where

\deqn{\phi' \sim \text{Beta}(2, 2).}

Slope \eqn{\eta} is drawn from a normal distribution,

\deqn{\eta \sim \text{N}(0, (\text{sd})^2).}
}

\examples{
LinAR1()
LinAR1(min = 0, s = 0.5, sd = 2)
}
\seealso{
\itemize{
\item \code{\link[=LinAR]{LinAR()}} Generalisation of \code{LinAR1()}
\item \code{\link[=Lin]{Lin()}} Linear prior with independent normal errors
\item \code{\link[=ELin]{ELin()}} Exchangeable version of \code{Lin()},
used with interactions
\item \link{priors} Overview of priors implemented in \strong{bage}
}
}
