% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_age_effect}
\alias{make_age_effect}
\title{Assemble effects involving age for a single
draw from the posterior distribution}
\usage{
make_age_effect(terms, dim, mappings, b, X)
}
\arguments{
\item{terms}{List of numeric vectors.}

\item{dim}{The dimensions of the array
that could be formed by all the non-age
dimensions.}

\item{mappings}{List of integer vectors,
the same length as \code{terms}.}

\item{b, X}{Numeric vector and matrix used
to convert back to the original space.}
}
\value{
An array, the first dimension
of which is age or age-sex.
}
\description{
Assemble terms that represent an age or age-sex
effect within a subspace, then transform back
to the original space. The calculations are
carried out on a single draw from the posterior
distribution.
}
\details{
Each element of \code{terms} is composed
of one or more vectors describing a point
within the subspace. For instance an age-time
interaction would be a matrix, each column
of which represented a single time point,
and each row of which represented distance
along the c'th dimension of the subspace.
The matrix (or array) is implicit, in that
the terms themselves do not have \code{dim}
attributes.

\code{dim} is the dimension of an array
formed by combining all the non-age (or non-age-sex)
dimensions in \code{terms}. For instance, if the
elements of \code{terms} represent an age effect,
an age-time interaction, and an age-region
interaction, then the (implicit) array
has two dimensions: time and region.

Each element of \code{mappings} gives the
indices of the array described above
that the corresponding element of \code{terms}
maps on to. An element of length 0
is used to represent a pure age (or age-sex)
effect.
}
