% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_linear_pred}
\alias{make_linear_pred}
\title{Assemble linear predictor by assembling contributions
of main effect and interactions}
\usage{
make_linear_pred(terms, dim, mappings)
}
\arguments{
\item{terms}{List of numeric vectors.}

\item{dim}{The dimensions of the array
to be constructed.}

\item{mappings}{List of integer vectors,
the same length as \code{terms}.}
}
\value{
An array, with the dimension
specified by \code{dim}.
}
\description{
Assemble terms that representing main effects
or interactions (eg age:time interactions),
to create linear predictor with the same
dimensions as the array of outcomes being modelled.
}
\details{
Each element of \code{terms} is a single draw from the
posterior distribution for tha that term.

\code{dim} is the dimension of the array
of outcomes, and hence of the linear predictor.

Each element of \code{mappings} gives the
indices of the array that the corresponding
element of \code{terms} maps on to.
}
