% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-constructors.R
\name{mod_norm}
\alias{mod_norm}
\title{Specify a Normal Model}
\usage{
mod_norm(formula, data, weights)
}
\arguments{
\item{formula}{An R \link[stats:formula]{formula},
specifying the outcome and predictors.}

\item{data}{A data frame containing outcome,
predictor, and, optionally, weights variables.}

\item{weights}{Name of the weights variable,
a \code{1}, or a formula. See below for details.}
}
\value{
An object of class \code{bage_mod_norm}.
}
\description{
Specify a model where the outcome is drawn from
a normal distribution.
}
\details{
The model is hierarchical. The means in the normal distribution
are described by a prior model formed from dimensions such
as age, sex, and time. The terms for these dimension themselves
have models, as described in \link{priors}. These priors all have defaults,
which depend on the type of term (eg an intercept, an age main effect,
or an age-time interaction.)
}
\section{Scaling of outcome and weights}{


Internally, \code{mod_norm()} scales the outcome variable
to have mean 0 and standard deviation 1, and
scales the weights to have mean 1.
This scaling allows \code{mod_norm()} to use the
same menu of priors as \code{\link[=mod_pois]{mod_pois()}} and \code{\link[=mod_binom]{mod_binom()}}.
(\code{\link[=mod_pois]{mod_pois()}} works on the log scale,
and \code{\link[=mod_binom]{mod_binom()}} works on the logit scale.)

\code{\link[=augment]{augment()}} always returns values on the
original scale, rather than the transformed scale.

\code{\link[=components]{components()}} by default returns values on
the transformed scale. But if \code{original_scale} is
\code{TRUE}, it returns some types of values on the
original scale. See the documentation for
\code{\link[=components]{components()}} for details.
}

\section{Specifying weights}{


The \code{weights} argument can take three forms:
\itemize{
\item the name of a variable in \code{data}, with or without
quote marks, eg \code{"wt"} or \code{wt};
\item the number \code{1}, in which no weights are used; or
\item a formula, which is evaluated with \code{data} as its
environment (see below for example).
}
}

\section{Mathematical details}{


The likelihood is

\deqn{\tilde{y}_i \sim \text{N}(\mu_i, \xi^2 / \tilde{w}_i)}

where

\deqn{\tilde{y}_i = (y_i - \bar{y}) / s}

and

\deqn{\tilde{w}_i = w_i / \bar{w},}

and where
\itemize{
\item subscript \eqn{i} identifies some combination of the
classifying variables, such as age, sex, and time,
\item \eqn{y_i} is the value of the outcome variable,
\item \eqn{w_i} is a weight,
\item \eqn{\bar{y}} is the mean across \eqn{i} of \eqn{y_i},
\item \eqn{s} is the sample standard deviation across \eqn{i} of \eqn{y_i},
\item \eqn{\bar{w}} is the mean across \eqn{i} of \eqn{w_i},
\item \eqn{\mu_i} is a mean parameter, and
\item \eqn{\xi} is a standard deviation parameter.
}

In some applications, \eqn{w_i} is set to 1
for all \eqn{i}.

The mean parameter \eqn{\mu_i} is modelled as
the sum of terms formed
from classifying variables and covariates,

\deqn{\mu_i = \sum_{m=0}^{M} \beta_{j_i^m}^{(m)}}

where
\itemize{
\item \eqn{\beta^{0}} is an intercept;
\item \eqn{\beta^{(m)}}, \eqn{m = 1, \dots, M}, is a main effect
or interaction; and
\item \eqn{j_i^m} is the element of \eqn{\beta^{(m)}} associated with
cell \eqn{i},
}

The \eqn{\beta^{(m)}} are given priors, as described in \link{priors}.

The prior for \eqn{\xi} is described in \code{\link[=set_disp]{set_disp()}}.

The model for \eqn{\mu_i}
can also include covariates,
as described in \code{\link[=set_covariates]{set_covariates()}}.
}

\examples{
mod <- mod_norm(value ~ diag:age + year,
                data = nld_expenditure,
                weights = 1)

## use formula to specify weights
mod <- mod_norm(value ~ diag:age + year,
                data = nld_expenditure,
                weights = ~sqrt(value))
}
\seealso{
\itemize{
\item \code{\link[=mod_pois]{mod_pois()}} Specify Poisson model
\item \code{\link[=mod_binom]{mod_binom()}} Specify binomial model
\item \code{\link[=set_prior]{set_prior()}} Specify non-default prior for term
\item \code{\link[=set_disp]{set_disp()}} Specify non-default prior for standard deviation
\item \code{\link[=fit]{fit()}} Fit a model
\item \code{\link[=augment]{augment()}} Extract values for rates,
means, or probabilities,
together with original data
\item \code{\link[=components]{components()}} Extract values for hyper-parameters
\item \code{\link[=forecast]{forecast()}} Forecast a model
\item \code{\link[=report_sim]{report_sim()}} Do a simulation study
\item \code{\link[=replicate_data]{replicate_data()}} Generate replicate
data for a model
\item \href{https://bayesiandemography.github.io/bage/articles/vig2_math.html}{Mathematical Details}
vignette
}
}
