% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mx_hmd}
\alias{mx_hmd}
\title{Mortality rates from the Human Mortality Database}
\format{
A matrix, in which each row holds a combination
of sex and age, and each column holds a combination
of country and year.
}
\source{
HMD. Human Mortality Database. Max Planck
Institute for Demographic Research (Germany),
University of California, Berkeley (USA), and
French Institute for Demographic Studies (France).
Available at www.mortality.org
(data downloaded on 24 December 2021).
}
\usage{
mx_hmd
}
\description{
Estimated mortality rates, modified from estimates
from the Human Mortality Database.
}
\keyword{datasets}
