% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmod.R
\name{new_bage_sysmod}
\alias{new_bage_sysmod}
\title{par, index_par, matrices_par are all fixed by 'data',
and do not change. These are therefore all
part of the 'mod' object.}
\usage{
new_bage_sysmod(
  formula,
  data,
  nm_distn,
  is_mod_with_offset,
  vname_offset,
  nm_offset
)
}
\description{
priors, and hence index_priors, hyper, and index_hyper
can change via 'set_prior'. Only 'priors' is held
in the 'mod' object - all the other quantities
that depend on them are derived on the fly
within function 'fit'.
}
