% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{NFix}
\alias{NFix}
\title{Normal Prior with Fixed Standard Deviation}
\usage{
NFix(sd = 1)
}
\arguments{
\item{sd}{Standard deviation.
A positive, finite number.
Default is \code{1}.}
}
\value{
An object of class \code{"bage_prior_normfixed"}.
}
\description{
Normal prior where, in contrast to \code{\link[=N]{N()}}, the
standard deviation is treated as fixed and known.
}
\details{
\code{NFix()} is the default prior for the intercept.
}
\section{Mathematical details}{


If \eqn{\beta_j} is the \eqn{j}th element of a main effect
or interaction, then

\deqn{\beta_j \sim \text{N}(0, \text{sd}^2).}
}

\examples{
NFix()
NFix(sd = 10) ## prior used for intercept
}
\seealso{
\itemize{
\item \code{\link[=N]{N()}} Version of \code{NFix()} where the standard deviation
term is estimated from the data.
\item \link{priors} Overview of priors implemented in \strong{bage}
}
}
