% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{NFixed}
\alias{NFixed}
\title{Normal prior with fixed standard deviation}
\usage{
NFixed(sd = 1)
}
\arguments{
\item{sd}{Standard deviation.
A positive, finite number.
Default is \code{1}.}
}
\value{
An object of class \code{bage_prior_normfixed}.
\code{NFixed()} is usually called within \code{\link[=set_prior]{set_prior()}}.
Other priors are \code{\link[=N]{N()}}, \code{\link[=RW]{RW()}}, \code{\link[=RW2]{RW2()}}.
}
\description{
Normal prior where, in contrast to \code{\link[=N]{N()}}, the
standard deviation is treated as fixed and known.
}
\details{
The distribution as mean equal to \code{0} and
standard deviation \code{sd}, where \code{sd} is supplied
by the user.

\deqn{x \sim \text{N}(0, s^2)}

\code{NFixed()} is the default prior for the intercept.
}
\examples{
NFixed()
NFixed(sd = 10) ## prior used for intercept
}
