% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nld_expenditure}
\alias{nld_expenditure}
\title{Per Capita Health Expenditure in the Netherlands, 2003-2011}
\format{
A \link[tibble:tibble-package]{tibble} with 1,296 rows and the
following columns:
\itemize{
\item \code{diag} Diagnostic group
\item \code{age} 5-year age groups, with open age
group of 85+
\item \code{year} 2003, 2005, 2007, and 2011
}
}
\source{
Calculated from data in table
"Expenditure by disease, age and gender under the System
of Health Accounts (SHA) Framework :
Current health spending by age"
from OECD database 'OECD.Stat' (downloaded on 25 May 2016)
and in table "Historical population data and projections
(1950-2050)" from OECD database 'OECD.Stat' (downloaded
5 June 2016).
}
\usage{
nld_expenditure
}
\description{
Per capita health expenditure, in Euros,
by diagnostic group, age group, and year, in the Netherlands.
}
\keyword{datasets}
