% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nzl_divorces}
\alias{nzl_divorces}
\title{Divorces in New Zealand}
\format{
A \link[tibble:tibble-package]{tibble} with 242 rows
and the following columns:
\itemize{
\item \code{age}: 5-year age groups, \code{"15-19"} to \code{"65+"}
\item \code{sex}: \code{"Female"} or \code{"Male"}
\item \code{time}: Calendar year
\item \code{divorces}: Numbers of divorces during year
\item \code{population}: Person-years lived during year
}
}
\source{
Divorce counts from data in table "Age at divorces by
sex (marriages and civil unions) (Annual-Dec)"
in the online database Infoshare
on the Statistics New Zealand website.
Data downloaded on 22 March 2023.
Population estimates derived from data in table
"Estimated Resident Population by Age and Sex (1991+)
(Annual-Dec)" in the online
database Infoshare
on the Statistics New Zealand website.
Data downloaded on 26 March 2023.
}
\usage{
nzl_divorces
}
\description{
Counts of divorces and population, by age, sex,
and calendar year, in New Zealand, 2011-2021.
}
\keyword{datasets}
