% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nzl_households}
\alias{nzl_households}
\title{People in One-Person Households in New Zealand}
\format{
A \link[tibble:tibble-package]{tibble} with 528 rows and the
following columns:
\itemize{
\item \code{age}: 5-year age groups, with open age
group of 65+
\item \code{region}: Region within New Zealand
\item \code{year}: Calendar year
\item \code{oneperson}: Count of people living in one-person
households
\item \code{total}: Count of people living in all types of
household
}
}
\source{
Derived from data in table
"Household composition by age group, for people in
households in occupied private dwellings,
2006, 2013, and 2018 Censuses (RC, TA, DHB, SA2)"
in the online database NZ.Stat,
on the Statistics New Zealand website.
Data downloaded on 3 January 2023.
}
\usage{
nzl_households
}
\description{
Counts of people in one-person households, and
counts of people living in any household, by
age, region, and year.
}
\keyword{datasets}
