% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.R
\name{prepare_svd_probs}
\alias{prepare_svd_probs}
\title{Prepare a matrix of probabilities for 
a singular value decomposition}
\usage{
prepare_svd_probs(x)
}
\arguments{
\item{x}{A matrix of probabilities
or proportions.}
}
\value{
A modified version
of \code{x}.
}
\description{
Prepare a matrix of probabilities or
proportions, \code{x},
for a singular value decomposition,
typically via function \code{\link{scaled_svd}}.
The preparation consists of
valdity checking, replacing any
zeros and ones with modelled values,
and then transforming to a logit scale.
}
\examples{
lfpr <- bage::lfpr_oecd
logit_lfpr <- prepare_svd_probs(lfpr)
}
\seealso{
\code{\link{scaled_svd}},
\code{\link{prepare_svd_probs}}
}
