% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.R
\name{prepare_svd_rates}
\alias{prepare_svd_rates}
\title{Prepare a matrix of rates for 
a singular value decomposition}
\usage{
prepare_svd_rates(x)
}
\arguments{
\item{x}{A matrix of rates.}
}
\value{
A modified version
of \code{x}.
}
\description{
Prepare a matrix of rates, \code{x},
for a singular value decomposition,
typically via function \code{\link{scaled_svd}}.
The preparation consists of
valdity checking, replacing any
zeros with small modelled values,
and then taking logs.
}
\examples{
mx <- bage::mx_hmd
log_mx <- prepare_svd_rates(mx)
}
\seealso{
\code{\link{scaled_svd}},
\code{\link{prepare_svd_probs}}
}
