% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prt_deaths}
\alias{prt_deaths}
\title{Deaths in Portugal}
\format{
A \link[tibble:tibble-package]{tibble}
with 3,168 rows and the following columns:
\itemize{
\item \code{age}: Age groups \code{"0"}, \code{"1-4"}, "5-9", ..., \code{"95-99"}, \code{"100+"}
\item \code{sex}: \code{"Female"} or \code{"Male"}
\item \code{time}: Calendar year
\item \code{deaths}: Count of deaths
\item \code{exposure}: Person-years lived by population
}
}
\source{
Human Mortality Database. University of California, Berkeley (USA),
and Max Planck Institute for Demographic Research (Germany).
Available at \url{https://www.mortality.org}.
(data downloaded on 17 July 2018).
}
\usage{
prt_deaths
}
\description{
Deaths and exposure in Portugal, by age, sex, and year.
}
\details{
The data are from the Human Mortality Database. Deaths are
rounded to the nearest integer.  More recent versions,
and a comprehensive description of the data,
are available at the HMD website.
}
\seealso{
\itemize{
\item \link{datasets} Overview of datasets in \pkg{bage}
}
}
\keyword{datasets}
