% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{reset_seeds}
\alias{reset_seeds}
\title{Reset Random Seeds in Model Object}
\usage{
reset_seeds(mod, new_seeds = NULL)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod"},
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}

\item{new_seeds}{\code{NULL} (the default) or a list
of integers. This list must have names
\code{"seed_stored_draws"}, \code{"seed_components"},
\code{"seed_augment"}, \code{"seed_forecast_components"},
and \code{"seed_forecast_augment"}.}
}
\value{
A modified version of \code{mod}.
}
\description{
Reset random seeds stored in a model object.
When \code{new_seeds} is \code{NULL} (the default),
the new seeds are generated randomly; otherwise
they are taken from \code{new_seeds}.
}
\details{
An object of class \code{"bage_mod"}
includes 5 random seeds:
\itemize{
\item \code{seed_stored_draws}
\item \code{seed_components}
\item \code{seed_augment}
\item \code{seed_forecast_components}
\item \code{seed_forecast_augment}
}

When functions \code{\link[=fit]{fit()}}, \code{\link[=component]{component()}}, \code{\link[=augment]{augment()}},
and \code{\link[=forecast]{forecast()}} are called, these seeds are used internally
to make sure that that the same inputs always give the
same outputs, even though the outputs are generated
using random draws.

\code{reset_seeds()} is not often called by end-users,
though it can occasionally be useful, such as when
creating simulations.
}
\examples{
## fit model
mod <- mod_pois(injuries ~ age,
                data = nzl_injuries,
                exposure = popn) |>
  fit()

## call 'components()'
components(mod)

## call 'components()' again - same results
components(mod)

## reset seeds
mod <- reset_seeds(mod)

## call 'components()' again - different results
components(mod)
}
\seealso{
\itemize{
\item \code{\link[=report_sim]{report_sim()}} Do a simulation study. (\code{report_sim()}
calls \code{reset_seeds()} internally.)
\item \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}}, \code{\link[=mod_norm]{mod_norm()}} Specify a model
\item \code{\link[=fit]{fit()}} Fit a model
\item \code{\link[=unfit]{unfit()}} Reset model, deleting estimates
}
}
