% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW}
\alias{RW}
\title{Random walk prior}
\usage{
RW(s = 1)
}
\arguments{
\item{s}{A positive, finite number. Default is 1.}
}
\value{
An object of class \code{"bage_prior_rw"}.
}
\description{
Prior in which units follow a (first order)
random walk. Increments between neighbouring
elements are assumed to be normally distibuted.
}
\section{Mathematical details}{


If \eqn{\beta_j} is the \eqn{j}th element of a main effect, then

\deqn{\beta_1 \sim \text{N}(0, 1)}
\deqn{\beta_j \sim \text{N}(\beta_{j - 1}, \tau^2), \quad j = 2, \cdots, J.}

Standard deviation \eqn{\tau} is drawn from a
half-normal distribution,

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{scale} lead to smoother series of \code{x}s, and
higher values lead to rougher series.
}

\examples{
RW()
RW(s = 0.5)
}
\seealso{
\itemize{
\item \code{\link[=RW2]{RW2()}} Random walk with drift.
\item \code{\link[=ERW]{ERW()}} Exchangeable version of \code{RW()},
used with interactions.
\item \link{priors} Overview of priors implemented in \code{bage}.
}
}
