% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW2}
\alias{RW2}
\title{Random walk with drift prior}
\usage{
RW2(s = 1, flat = FALSE)
}
\arguments{
\item{s}{A positive, finite number. Default is 1.}

\item{flat}{Whether the overall slope of the
terms is constrained to be zero. Default
is \code{FALSE}.}
}
\value{
An object of class \code{bage_prior_rw}
or \code{bage_prior_rw2}.
}
\description{
Prior in which units follow a random walk with drift.
Second-order differences are normally distributed,
}
\details{
\deqn{(x_i - x_{i-1}) - (x_{i-1} - x_{i-2}) \sim \text{N}(0, \sigma^2),
      i = 3, \cdots, n}

For identifiability,
the mean value for the $x_i$ is assumed to be
(almost) exactly equal to zero.

When \code{flat} is \code{FALSE} (the default),
no constraint is placed on the overall slope of the $x_i$.
When \code{flat} is \code{TRUE}, a line through the middle of the
$x_i$ must be be horizontal.

Standard deviation \eqn{\sigma} is drawn from a
half-normal distribution,

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{scale} lead to smoother series of \code{x}s, and
higher values lead to rougher series.
}
\examples{
RW2()
RW2(flat = TRUE)
}
\seealso{
\itemize{
\item \code{\link[=RW]{RW()}} for a first-order random walk.
\item \code{RW2()} is usually called as part of
a call to \code{\link[=set_prior]{set_prior()}}
}
}
