% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW2_Infant}
\alias{RW2_Infant}
\title{Second-Order Random Walk Prior with 'Infant' Indicator}
\usage{
RW2_Infant(s = 1, sd_slope = 1, zero_sum = FALSE)
}
\arguments{
\item{s}{Scale for the prior for the innovations.
Default is \code{1}.}

\item{sd_slope}{Standard deviation
for initial slope of random walk. Default is \code{1}.}

\item{zero_sum}{If \code{TRUE}, values must
sum to 0 within each combination
of the 'by' variables. Default is \code{FALSE}.}
}
\value{
Object of class \code{"bage_prior_rw2infant"}.
}
\description{
Use a second-order random walk to model variation
over age, with an indicator variable for the first age group.
Designed for use in models of mortality rates.
}
\details{
A second-order random walk prior \code{\link[=RW2]{RW2()}}
does a good job of smoothing
mortality rates over age, except at age 0, where there
is typically a sudden jump in rates, reflecting the
special risks of infancy. The \code{RW2_Infant()}
is a \code{\link[=RW2]{RW2()}} prior with a special treatment of
the first age group.

If \code{RW2_Infant()} is used in an interaction,
the 'along' dimension is always age, implying that
there is a separate random walk along age within each
combination of the 'by' variables.

Argument \code{s} controls the size of innovations in the random walk.
Smaller values for \code{s} tend to give smoother series.

Argument \code{sd} controls the sl size of innovations in the random walk.
Smaller values for \code{s} tend to give smoother series.
}
\section{Mathematical details}{


When \code{RW2_Infant()} is used with a main effect,

\deqn{\beta_1 \sim \text{N}(0, 1)}
\deqn{\beta_2 \sim \text{N}(0, \omega^2)}
\deqn{\beta_3 \sim \text{N}(2 \beta_2, \tau^2)}
\deqn{\beta_j \sim \text{N}(2 \beta_{j-1} - \beta_{j-2}, \tau^2)}

and when it is used with an interaction,

\deqn{\beta_{u,1} \sim \text{N}(0, 1)}
\deqn{\beta_{u,2} \sim \text{N}(0, \omega^2)}
\deqn{\beta_{u,3} \sim \text{N}(2 \beta_{u,2}, \tau^2)}
\deqn{\beta_{u,j} \sim \text{N}(2 \beta_{u,v-1} - \beta_{u,v-2}, \tau^2)}

where
\itemize{
\item \eqn{\pmb{\beta}} is a main effect or interaction;
\item \eqn{j} denotes position within the main effect;
\item \eqn{v} denotes position within the 'along' variable of the interaction; and
\item \eqn{u} denotes position within the 'by' variable(s) of the interaction.
}

Parameter \eqn{\omega} has a half-normal prior
\deqn{\omega \sim \text{N}^+(0, \text{sd}^2),}
where \code{sd} can be specified by the user.

Parameter \eqn{\tau} has a half-normal prior
\deqn{\tau \sim \text{N}^+(0, \text{s}^2),}
where \code{s} can be specified by the user.
}

\examples{
RW2_Infant()
RW2_Infant(s = 0.1)
}
\seealso{
\itemize{
\item \code{\link[=RW2]{RW2()}} Second-order random walk, without infant indicator
\item \code{\link[=Sp]{Sp()}} Smoothing via splines
\item \code{\link[=SVD]{SVD()}} Smoothing over age via singular value decomposition
\item \link{priors} Overview of priors implemented in \strong{bage}
\item \code{\link[=set_prior]{set_prior()}} Specify prior for intercept,
main effect, or interaction
}
}
