% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW2_Seas}
\alias{RW2_Seas}
\title{Second-Order Random Walk Prior with Seasonal Effect}
\usage{
RW2_Seas(
  n_seas,
  s = 1,
  sd = 1,
  sd_slope = 1,
  s_seas = 0,
  sd_seas = 1,
  along = NULL,
  zero_sum = FALSE
)
}
\arguments{
\item{n_seas}{Number of seasons}

\item{s}{Scale for prior for innovations in
random walk. Default is \code{1}.}

\item{sd_slope}{Standard deviation
for initial slope of random walk. Default is \code{1}.}

\item{s_seas}{Scale for innovations
in seasonal effects. Default is \code{0}.}

\item{sd_seas}{Standard deviation for
initial values of seasonal effects.
Default is \code{1}.}

\item{along}{Name of the variable to be used
as the 'along' variable. Only used with
interactions.}

\item{zero_sum}{If \code{TRUE}, values must
sum to 0 within each combination
of the 'by' variables. Default is \code{FALSE}.}
}
\value{
Object of class
\code{"bage_prior_rw2randomseasvary"},
\code{"bage_prior_rw2randomseasfix"},
\code{"bage_prior_rw2zeroseasvary"}, or
\code{"bage_prior_rw2zeroseasfix"}.
}
\description{
Use a second-oder random walk with
seasonal effects as a model for
a main effect, or use multiple second-order random walks,
each with their own seasonal effects,
as a model for an interaction.
Typically used with temrs that involve time.
}
\details{
If \code{RW2_Seas()} is used with an interaction,
a separate series is constructed within each
combination of the 'by' variables.

Argument \code{s} controls the size of innovations in the random walk.
Smaller values for \code{s} tend to produce smoother series.

Argument \code{n_seas} controls the number of seasons.
When using quarterly data, for instance,
\code{n_seas} should be \code{4}.

By default, the magnitude of seasonal effects
is fixed. However, setting \code{s_seas} to a value
greater than zero produces seasonal effects
that evolve over time.
}
\section{Mathematical details}{


When \code{RW2_Seas()} is used with a main effect,

\deqn{\beta_j = \alpha_j + \lambda_j}
\deqn{\alpha_1 \sim \text{N}(0, \text{sd}^2)}
\deqn{\alpha_2 \sim \text{N}(0, \text{sd\_slope}^2)}
\deqn{\alpha_j \sim \text{N}(2 \alpha_{j-1} - \alpha_{j-2}, \tau^2), \quad j > 2}
\deqn{\lambda_j \sim \text{N}(\lambda_{j-n}, \omega^2),}

and when it is used with an interaction,

\deqn{\beta_{u,v} = \alpha_{u,v} + \lambda_{u,v}}
\deqn{\alpha_{u,1} \sim \text{N}(0, \text{sd}^2)}
\deqn{\alpha_{u,2} \sim \text{N}(0, \text{sd\_slope}^2)}
\deqn{\alpha_{u,v} \sim \text{N}(2 \alpha_{u,v-1} - \alpha_{u,v-2}, \tau^2), \quad j > 2}
\deqn{\lambda_{u,v} \sim \text{N}(\lambda_{u,v-n}, \omega^2)}

where
\itemize{
\item \eqn{\pmb{\beta}} is the main effect or interaction;
\item \eqn{\alpha_j} or \eqn{\alpha_{u,v}} is an element of the random walk;
\item \eqn{\lambda_j} or \eqn{\lambda_{u,v}} is an element of the seasonal effect;
\item \eqn{j} denotes position within the main effect;
\item \eqn{v} denotes position within the 'along' variable of the interaction;
\item \eqn{u} denotes position within the 'by' variable(s) of the interaction; and
\item \eqn{n} is \code{n_seas}.
}

Parameter \eqn{\omega} has a half-normal prior
\deqn{\omega \sim \text{N}^+(0, \text{s\_seas}^2),}
where \code{s_seas} is provided by the user. If
\code{s_seas} is set to 0, then \eqn{\omega} is 0,
and the seasonal effects are fixed over time.

Parameter \eqn{\tau} has a half-normal prior
\deqn{\tau \sim \text{N}^+(0, \text{s}^2),}
where \code{s} is provided by the user.
}

\examples{
RW2_Seas(n_seas = 4)               ## seasonal effects fixed
RW2_Seas(n_seas = 4, s_seas = 0.5) ## seasonal effects evolve
RW2_Seas(n_seas = 4, sd = 0)       ## first term in random walk fixed at 0
}
\seealso{
\itemize{
\item \code{\link[=RW2]{RW2()}} Second-order random walk without seasonal effect
\item \code{\link[=RW_Seas]{RW_Seas()}} Random walk with seasonal effect
\item \link{priors} Overview of priors implemented in \strong{bage}
\item \code{\link[=set_prior]{set_prior()}} Specify prior for intercept,
main effect, or interaction
}
}
