% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_scaled_svd-constructors.R
\name{scaled_svd}
\alias{scaled_svd}
\title{Create object to hold data from scaled SVDs}
\usage{
scaled_svd(data)
}
\arguments{
\item{data}{A data frame. See Details for description.}
}
\value{
An object of class \code{bage_scaled_svd}.
}
\description{
Create an object of class \code{bage_scaled_svd} to
hold scaled SVDs (singular value decompositions.)
}
\details{
\code{data} has the following columns:
\itemize{
\item \code{sexgender} Labels for sexes/genders. The column
can also be called \code{sex} or \code{gender}.
\item \code{matrix} List column of matrices created by
\code{\link[=scaled_svd_comp]{scaled_svd_comp()}}. Must have rownames.
Must not have NAs. Each matrix has 10 columns.
\item \code{offset} List column of vectors created by
\code{\link[=scaled_svd_comp]{scaled_svd_comp()}}. Must have names, which
are identical to the rownames of the corresponding
element of \code{matrix}.
}
}
