% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{Seas}
\alias{Seas}
\title{Prior for Seasonal Effect}
\usage{
Seas(n, s = 1)
}
\arguments{
\item{n}{The number of seasons.}

\item{s}{A positive, finite number. Default is 1.}
}
\value{
An object of class \code{"bage_prior_seas"}.
}
\description{
Prior for seasonal effect, where the size of
the effects evolve over time. Each season follows its
own random walk.
}
\section{Mathematical description}{


\deqn{x_j \sim \text{N}(0, 1), j = 1, \cdots, \text{n}}
\deqn{x_j - x_{j-n} \sim \text{N}(x_{j - n}, \sigma^2), j = n+1, \cdots, J}

Standard deviation \eqn{\sigma} is drawn from a
half-normal distribution,

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{scale} lead to smoother series.
}

\examples{
Seas(n = 4)
Seas(n = 12, s = 0.5)
}
\seealso{
\itemize{
\item \code{\link[=RW]{RW()}} for an ordinary random walk.
}
}
