% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_cyclical}
\alias{set_cyclical}
\title{Add a cyclical effect}
\usage{
set_cyclical(mod, n = 2, s = 1)
}
\arguments{
\item{mod}{A \code{bage_mod} object, typically
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}

\item{n}{Number of terms in model.
Default is 2. Minimum is 1.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\sigma}).
Defaults to 1.}
}
\value{
A modified \code{bage_mod} object.
}
\description{
Add a cyclical effect to a model.
}
\details{
TODO - description

If the \code{mod} argument to \code{set_cyclical} is
a fitted model, then \code{set_cyclical} 'unfits'
the model, by deleting existing estimates.
}
\examples{
mod <- mod_pois(deaths ~ month,
                data = us_acc_deaths,
                exposure = 1)
mod
mod |> set_cyclical()
}
\seealso{
\itemize{
\item \code{\link[=is_fitted]{is_fitted()}}
}
}
