% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_datamod_outcome_rr3}
\alias{set_datamod_outcome_rr3}
\title{Specify RR3 Data Model}
\usage{
set_datamod_outcome_rr3(mod)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod"},
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}
}
\value{
A revised version of \code{mod}.
}
\description{
#' `r lifecycle::badge('deprecated')
}
\details{
This function has been deprecated, and will
be removed from future versions of \code{bage}.
Please used function \code{\link[=set_confidential_rr3]{set_confidential_rr3()}}
instead.
}
\examples{
## 'injuries' variable in 'nzl_injuries' dataset
## has been randomly rounded to base 3
mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
                data = nzl_injuries,
                exposure = popn) |>
  set_confidential_rr3() |> ## rather than set_datamod_outcome_rr3
  fit()
}
