% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_datamod_overcount}
\alias{set_datamod_overcount}
\title{Specify Overcount Data Model}
\usage{
set_datamod_overcount(mod, rate)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod_pois"},
created with \code{\link[=mod_pois]{mod_pois()}}.}

\item{rate}{The prior for the overcoverage rate.
A data frame with a variable
called \code{"mean"}, a variable called \code{"disp"},
and, optionally, one or more 'by' variables.}
}
\value{
A revised version of \code{mod}.
}
\description{
Specify Overcount Data Model
}
