% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_datamod_undercount}
\alias{set_datamod_undercount}
\title{Specify Undercount Data Model}
\usage{
set_datamod_undercount(mod, prob)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod_pois"},
created with \code{\link[=mod_pois]{mod_pois()}}.}

\item{prob}{The prior for the probability
that a person or event in the target
population will correctly enumerated.
A data frame with a variable
called \code{"mean"}, a variable called \code{"disp"},
and, optionally, one or more 'by' variables.}
}
\value{
A revised version of \code{mod}.
}
\description{
Specify Undercount Data Model
}
