% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_disp}
\alias{set_disp}
\title{Set scale parameter for dispersion}
\usage{
set_disp(mod, s)
}
\arguments{
\item{mod}{A \code{bage_mod} object, typically
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}

\item{s}{Scale term. In Poisson and
binomial models, \code{s} must be non-negative.
In normal models, \code{s} must be positive.}
}
\value{
A \code{bage_mod} object
}
\description{
Specify the scale parameter \code{s} in the prior
for dispersion.
}
\details{
In Poisson and binomial models,
\code{s} can be set to \code{0}, implying
that the dispersion term is also \code{0}.
In normal models, \code{s} must be non-negative.

If the \code{mod} argument to \code{set_disp} is
a fitted model, then \code{set_disp} 'unfits'
the model, by deleting existing estimates.
}
\examples{
mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
                data = injuries,
                exposure = popn)
mod
mod |> set_disp(s = 0.1)
mod |> set_disp(s = 0)
}
\seealso{
\itemize{
\item \code{\link[=is_fitted]{is_fitted()}}
}
}
