% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_prior}
\alias{set_prior}
\title{Change the prior for a model term}
\usage{
set_prior(mod, formula)
}
\arguments{
\item{mod}{A \code{bage_mod} object, created with
\code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}

\item{formula}{A formula giving the term
and a function for creating a prior.}
}
\value{
A \code{bage_mod} object.
}
\description{
Specify a non-default prior distribution
for a main effect or interaction.
}
\details{
\code{formula} gives the name of a main
effect or interaction, and a function
specifying a prior, eg
\code{age ~ RW2()}.
}
\examples{
mod <- mod_pois(injuries ~ age + year,
                data = injuries,
                exposure = popn)
mod
mod |> set_prior(age ~ RW2())
}
\seealso{
\code{\link[=N]{N()}}, \code{\link[=RW]{RW()}}, \code{\link[=RW2]{RW2()}}
}
