% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_var_sexgender}
\alias{set_var_sexgender}
\title{Set a sex or gender variable}
\usage{
set_var_sexgender(mod, name)
}
\arguments{
\item{mod}{A \code{"bage_mod"} object, typically
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}

\item{name}{The name of the sex or gender variable.}
}
\value{
A \code{"bage_mod"} object
}
\description{
Specify which variable (if any) represents sex or gender.
Functions \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}},
and \code{\link[=mod_norm]{mod_norm()}} try to find a sex or gender variable
using variable names, but do not always get it right.
}
\details{
In an R \code{\link{formula}}, a 'variable' is different
from a 'term'. For instance,

\code{~ gender + region + gender:region}

contains variables \code{gender} and \code{region},
and terms \code{gender}, \code{region}, and \code{gender:region}.
}
\examples{
## rename 'sex' variable to something unexpected
injuries2 <- injuries
injuries2$biological_sex <- injuries2$sex

## mod_pois does not recognize sex variable
mod <- mod_pois(injuries ~ age * biological_sex + year,
                data = injuries2,
                exposure = popn)
mod

## so we set the sex variable explicitly
mod |>
  set_var_sexgender(name = "biological_sex")
}
\seealso{
\itemize{
\item \code{\link[=set_var_age]{set_var_age()}} sets the age variable
\item \code{\link[=set_var_time]{set_var_time()}} sets the time variable
\item internally, \code{bage} uses \code{\link[=find_sexgender]{find_sexgender()}} to locate sex or gender variables
\item internally, \code{bage} uses \code{\link[=find_female]{find_female()}} to locate female categories within a sex or gender variable
\item internally, \code{bage} uses \code{\link[=find_male]{find_male()}} to locate male categories within a sex or gender variable
}
}
