% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate from a model}
\usage{
sim(x_est, x_gen = NULL, n_sim = 1, known = "(Intercept)")
}
\arguments{
\item{x_est}{A \code{bage_mod} object.}

\item{x_gen}{\code{NULL} or a \code{bage_mod} object.}

\item{n_sim}{Number of draws from simulated posterior.
Defaults to 1.}
}
\value{
A \code{bage_mod_sim} object
}
\description{
Simulate from a model
}
