% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{Sp}
\alias{Sp}
\title{P-spline prior}
\usage{
Sp(n = NULL, s = 1, sd = 1)
}
\arguments{
\item{n}{Number of spline vectors.
By default is \code{NULL}, in which case the number of
vectors is set to \code{max(ceiling(0.7 * k), 4)}
where \code{k} is the number
of elements in the term being modelled.}

\item{s}{A positive, finite number. Default is 1.}

\item{sd}{A positive, finite number. Default is 1.}
}
\value{
An object of class \code{"bage_prior_spline"}.
}
\description{
Specify a P-spline (penalised spline) prior.
A P-spline is flexible, but
favours profiles that are relatively smooth.
}
\section{Mathematical details}{


The model for the effect, on the log scale, is

\deqn{\beta = X \gamma}

where
\itemize{
\item \eqn{\beta} is a main effect,
\item \eqn{X} is a matrix holding the basis functions
for the spline, with \code{n} columns, and
\item \eqn{\alpha} is a vector of coefficients,
with \code{n} elements.
}

The elements of \eqn{\gamma} are assumed to follow
a second order random walk (see \code{\link[=RW2]{RW2()}}).
}

\examples{
Sp()
Sp(n = 10)
}
\seealso{
\itemize{
\item \link{priors} Overview of priors implemented in \code{bage}.
}
}
