% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{Spline}
\alias{Spline}
\title{P-spline prior}
\usage{
Spline(n = NULL, s = 1)
}
\arguments{
\item{n}{Number of spline vectors.
By default is \code{NULL}, in which case the number of
vectors is set to \code{max(ceiling(0.7 * k), 4)}
where \code{k} is the number
of elements in the term being modelled.}

\item{s}{Scale for error term.
Default is 1.}
}
\value{
An object of class \code{"bage_prior_spline"}.
}
\description{
Specify a P-spline (penalised spline) prior.
A P-spline is flexible, but
favours profiles that are relatively smooth.
}
\section{Mathematical details}{


The model for the effect, on the log scale, is

\deqn{\beta = X \gamma}

where
\itemize{
\item \eqn{\beta} is a main effect,
\item \eqn{X} is a matrix holding the basis functions
for the spline, with \code{n} columns, and
\item \eqn{\alpha} is a vector of coefficients,
with \code{n} elements.
}

The elements of \eqn{\gamma} are assumed to follow
a second order random walk,

\deqn{(\gamma_j - \gamma_{j-1}) - (\gamma_{j-1} - \gamma_{j-2}) \sim \text{N}(0, \sigma^2)}

Parameter \eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, s^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for \eqn{s} is supplied by the user.
}

\examples{
Spline()
Spline(n = 10)
}
\seealso{
\itemize{
\item \code{\link[=RW2]{RW2()}} etc
}
}
