% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvd_comp.R
\name{ssvd_hmd}
\alias{ssvd_hmd}
\title{Create a Scaled SVD Object from a Human Mortality Database File}
\usage{
ssvd_hmd(zipfile)
}
\arguments{
\item{zipfile}{The name of a zipped file downloaded
from the Human Mortality Database.
Any path name that can be handled by \code{\link[utils:unzip]{utils::unzip()}}.
Defaults to \code{"log"}.}
}
\value{
An object of class \code{\link[=ssvd]{"bage_ssvd"}}
}
\description{
Create an object of class \code{\link[=ssvd]{"bage_ssvd"}}
from a zipped file downloaded from the
\href{https://www.mortality.org/Data/ZippedDataFiles}{Human Mortality Database}.
}
\section{Warning}{


This function, an associated functions, will probably
be moved into a separate package in future.
}

\section{Usage}{

\strong{Step 1: Download data}

Register or log in at the Human Mortality Database, and go to page
\href{https://www.mortality.org/Data/ZippedDataFiles}{Downloading the HMD in zipped data files}.
Go to the "Previous Versions" table at the bottom of the page, and
download a file from the "Statistics" column, eg file

https://www.mortality.org/File/Download/hmd.v6/zip/all_hmd/hmd_statistics_20240226.zip

To download a file, you will need to register with the Human Mortality Databse
and log in.

\strong{Step 2: Call function 'svd_hmd'}

Put the file into the working directory, and supply the name of the file
to function \code{svd_hmd()}, eg

\if{html}{\out{<div class="sourceCode">}}\preformatted{MyHMD <- svd_hmd("hmd_statistcs_20240226")
}\if{html}{\out{</div>}}

\strong{Step 3: Use object to create a prior}

\if{html}{\out{<div class="sourceCode">}}\preformatted{mod |>
  set_prior(age ~ SVD(MyHMD))
}\if{html}{\out{</div>}}
}

\seealso{
\itemize{
\item \code{\link[=SVD]{SVD()}}, \code{\link[=SVDS]{SVDS()}}
to use an object of class
\code{\link[=ssvd]{"bage_ssvd"}} in a prior.
}
}
