% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{standardize_vals_effect}
\alias{standardize_vals_effect}
\title{Standardize Effects}
\usage{
standardize_vals_effect(mod, vals_effect)
}
\arguments{
\item{mod}{}
}
\description{
Apply same standardization to simulated effects
that is applied to estimated effects.
}
