% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{SVD2}
\alias{SVD2}
\title{Two-Sex SVD prior}
\usage{
SVD2(scaled_svd, n = 5, joint = FALSE)
}
\arguments{
\item{scaled_svd}{An object created by \code{\link[=scaled_svd]{scaled_svd()}},
holding scaled values from an SVD of age-specific values.}

\item{n}{Number of components from scaled SVD
to use in modelling. Must be between 1 and 10.
Default is 5.}

\item{joint}{If \code{TRUE}, use a SVD calculated for
females and males jointly. If \code{FALSE}, use one SVD
for females, and a separate one for males.
See below for details. Default is \code{FALSE}.}
}
\value{
An object of class \code{"bage_prior_svd"}.
}
\description{
Specify a Singular Value Decomposition (SVD) prior
for an age-sex interaction involving two sexes
}
\details{
A sex-specific SVD prior assumes that age-sex-profile
for the quantity being modelled looks like
an age-sex profile drawn at random
from an external demographic database. For instance,

\if{html}{\out{<div class="sourceCode">}}\preformatted{set_prior(mod, age:sex ~ SVD2(HMD))
}\if{html}{\out{</div>}}

specifies that the quantity being modelled in \code{mod}
has an age-sex profile that looks like it was drawn from the
\href{https://www.mortality.org}{Human Mortality Database}.
A Singular Value Decomposition is closely related
to a Principal Components analysis, and is a way of
extracting patterns from large volumes of data.
}
\section{Joint or independent SVDs}{


To possible ways of extracting patterns
from age-sex-specific data are
\enumerate{
\item carry out a SVD on dataset for females, and a separate
SVD on a dataset for males; or
\item carry out a single SVD on dataset that combines data for
females and data for males.
}

Option 1 is more flexible. Option 2 is
more robust to sampling or measurement errors.
Option 1 is obtained by setting the \code{joint}
argument in \code{SVD2()} to \code{FALSE}. Option 2
is obtained by setting the
\code{joint} argument to \code{TRUE}.
The default is \code{FALSE}.
}

\section{Mathematical details}{


When \code{joint} is \code{FALSE},

\deqn{\pmb{\beta}_{\text{F}} = \pmb{F}_{\text{F}} \pmb{\alpha}_{\text{F}} + \pmb{g}_{\text{F}}}
\deqn{\pmb{\beta}_{\text{M}} = \pmb{F}_{\text{M}} \pmb{\alpha}_{\text{M}} + \pmb{g}_{\text{M}}}

and when \code{joint} is \code{TRUE},

\deqn{\pmb{\beta}_{\text{FM}} = \pmb{F}_{\text{FM}} \pmb{\alpha}_{\text{FM}} + \pmb{g}_{\text{FM}}}
where
\itemize{
\item \eqn{\pmb{\beta}_{\text{F}}} is a vector with \eqn{J/2} elements, specifying the age profile for females;
\item \eqn{\pmb{\beta}_{\text{M}}} is a vector with \eqn{J/2} elements, specifying the age profile for males;
\item \eqn{\pmb{\beta}_{\text{FM}}} is a vector with \eqn{J} elements, specifying the age profile for females, followed by the age profile for males;
\item \eqn{\pmb{X}_{\text{F}}} is a known \eqn{J/2 \times K} matrix, estimated from data for females;
\item \eqn{\pmb{X}_{\text{M}}} is a known \eqn{J/2 \times K} matrix, estimated from data for males;
\item \eqn{\pmb{X}_{\text{FM}}} is a known \eqn{J \times K} matrix, estimated from data for females and males;
\item \eqn{\pmb{\alpha}_{\text{F}}}, \eqn{\pmb{\alpha}_{\text{M}}} and \eqn{\pmb{\alpha}_{\text{FM}}} are vectors with \eqn{K} elements;
\item \eqn{\pmb{g}_{\text{F}}} is a known vector with \eqn{J/2} elements, estimated from data for females;
\item \eqn{\pmb{g}_{\text{M}}} is a known vector with \eqn{J/2} elements, estimated from data for females; and
\item \eqn{\pmb{g}_{\text{FM}}} is a known vector with \eqn{J} elements, estimated from data for females and males.
}

The elements of \eqn{\pmb{\alpha}_{\text{F}}}, \eqn{\pmb{\alpha}_{\text{M}}}, and \eqn{\pmb{\alpha}_{\text{FM}}} have prior

\deqn{\alpha_i \sim \text{N}(0, 1), \quad i = 1, \cdots, K}

For a description of the construction of \eqn{\pmb{F}_{\text{F}}},
\eqn{\pmb{F}_{\text{M}}}, \eqn{\pmb{F}_{\text{FM}}},
\eqn{\pmb{g}_{\text{F}}}, \eqn{\pmb{g}_{\text{M}}},
and \eqn{\pmb{g}_{\text{FM}}}, see TODO - REFERENCE TO VIGNETTE.
}

\section{Scaled SVDs in bage}{

\itemize{
\item \code{\link{HMD}} Values from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item TODO - add others
}
}

\examples{
SVD2(HMD)
SVD2(HMD, joint = TRUE)
}
\seealso{
\itemize{
\item \code{\link[=set_prior]{set_prior()}} to specify the prior for a model
\item \code{\link[=SVD]{SVD()}} Combined-sexes version of \code{SVD2()}
\item \code{\link[=ESVD]{ESVD()}} Exchangeable SVD prior for interactions
\item \code{\link[=ESVD2]{ESVD2()}} Exchangeable two-sex SVD prior
for interactions
\item XXX to create a new scaled SVD.
}
}
