% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.R
\name{svd_transform}
\alias{svd_transform}
\title{Construct transformations based on SVD
of age-sex-specific rates}
\usage{
svd_transform(x, n = 5, scale = c("log", "logit", "none"), concat = NULL)
}
\arguments{
\item{x}{A data frame with rates, probabilities,
or means classified by age, sex/gender, and other
variables.}

\item{n}{Number of components of SVD.}

\item{scale}{\code{"log"}, \code{"logit"}, or \code{"none"}.
Defaults to \code{"log"}.}

\item{concat}{Levels of the sex/gender
variable to use for a \code{"concat"}
transformation.}
}
\value{
A named list, with one element
for each level of the sex/gender variable,
plus one more element if a value for
\code{concat} is supplied.
Each element in the return value
is itself a named list,
holding a matrix called \code{"transform"}
and a vector called \code{"translate"}.
}
\description{
Apply the \link[base:svd]{Singular Value Decomposition}
(SVD) to a dataset of ages-sex-specific rates,
probabilities, or means, and use the results
to construct transformations that allow
age-sex profiles to be represented more
parsimoniously.
}
\details{
Data frame \code{x} must have
\itemize{
\item a column named \verb{"age}"
\item a column named \code{"sex"} or \code{"gender"}
\item one or more columns containing additional
classifying variables
\item a column called \code{"value"} holding the rates,
probabilities, or means.
}

The number of components used by the SVD
is governed by argument \code{n}, which defaults
to 5.

Rates are typically transformed to the
log scale, and probabilities to the logit scale,
before the SVD is applied. The type of
transformation is specified through the
\code{scale} argument (which defaults to \code{"log"}.)

When a log or logit scale is used, \code{svd_transform}
converts zeros in the \code{value} column to
values just above zero before transforming.
When a logit scale is used, \code{svd_transform}
also converts ones to values just
below one. The substitute
values are calculated by fitting a main-effects
model to the data, and then shifting the predicted
value down slightly (in the case of zeros),
or up slightly (in the case of ones.)

By default, \code{svd_transform} creates a
transformation for each level of the sex/gender
variable. If a value is supplied for the
\code{concat} argument, then \code{svd_transform}
also creates a transformation that concatenates
age levels for each sex/gender included
in the value supplied for \code{concat}. See below for
an example. Concatenated transforms are useful,
for instance, when jointly modelling female
and male rates.
}
\examples{
set.seed(0)
x <- expand.grid(age = 0:10,
                 gender = c("Female", "Male", "Total"),
                 year = 2020:2021,
                 country = LETTERS)
x$value <- rgamma(n = nrow(x), shape = 0.2)

## "Female", "Male", and "Total"
tr <- svd_transform(x)
str(tr)

## "Female", "Male", "Total", and "concat"
tr_concat <- svd_transform(x, concat = c("Female", "Male"))
str(tr_concat)
}
