% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{SVDS_AR}
\alias{SVDS_AR}
\alias{SVDS_AR1}
\alias{SVDS_RW}
\alias{SVDS_RW2}
\title{SVD Priors for Age-Sex or Age-Gender Interactions,
With Time-Varying Coefficients}
\usage{
SVDS_AR(ssvd, n_comp = NULL, joint = FALSE, n_coef = 2, s = 1)

SVDS_AR1(ssvd, n_comp = NULL, joint = FALSE, min = 0.8, max = 0.98, s = 1)

SVDS_RW(ssvd, n_comp = NULL, joint = FALSE, s = 1)

SVDS_RW2(ssvd, n_comp = NULL, joint = FALSE, s = 1)
}
\arguments{
\item{ssvd}{Object of class \code{"bage_ssvd"}
holding a scaled SVD. See below for scaled SVDs
of databases currently available in \strong{bage}.}

\item{n_comp}{Number of components from scaled SVD
to use in modelling. The default is half
the number of components of \code{ssvd}.}

\item{joint}{Whether to use combined or
separate SVDs. Default is \code{FALSE}.
See below for details.}

\item{n_coef}{Number of AR coefficients in \code{SVD_RW()}.}

\item{s}{Scale for standard deviations terms.}

\item{min, max}{Minimum and maximum values
for autocorrelation coefficient in \code{SVD_AR()}.
Defaults are \code{0.8} and \code{0.98}.}
}
\value{
An object of class \code{"bage_prior_svd_ar"},
\code{"bage_prior_svd_rw"}, or \code{"bage_prior_svd_rw2"}.
}
\description{
Singular Value Decomposition (SVD) priors
for interactions involving age, sex or gender,
time, and possibly other variables.
}
\details{
\code{SVD_AR()}, \code{SVD_AR1()}, \code{SVD_RW()}, and \code{SVD_RW2()}
priors assume that, in any given period,
the age-sex profile or age-gender
profiles for the quantity being modelled look
they were drawn at random
from an external demographic database. For instance,
the prior obtained via

\if{html}{\out{<div class="sourceCode">}}\preformatted{SVDS_AR1(HMD)
}\if{html}{\out{</div>}}

favours age-sex profiles that look like
they were obtained from the
\href{https://www.mortality.org}{Human Mortality Database}.

Time-varying age profiles with no sex or gender
dimension should be modelled using \code{\link[=SVD_AR]{SVD_AR()}},
\code{\link[=SVD_AR1]{SVD_AR1()}}, \code{\link[=SVD_RW]{SVD_RW()}}, or \code{\link[=SVD_RW2]{SVD_RW2()}} priors,
rather than \code{SVDS_AR()},
\code{SVDS_AR1()}, \code{SVDS_RW()}, or \code{SVDS_RW2()}.
}
\section{Mathematical details}{


\strong{Case 1: Interaction involving age, sex/gender, time, and no other variables}

When \code{joint} is \code{FALSE},

\deqn{\pmb{\beta}_{s,t} = \pmb{F}_s \pmb{\alpha}_{s,t} + \pmb{g}_s,}

and when \code{joint} is \code{TRUE},

\deqn{\pmb{\beta}_t = \pmb{F} \pmb{\alpha}_t + \pmb{g},}

where
\itemize{
\item \eqn{\pmb{\beta}} is an interaction involving age, sex/gender, and time;
\item \eqn{\pmb{\beta}_{s,t}} is a subvector of \eqn{\pmb{\beta}},
holding values for sex/gender \eqn{s} in period \eqn{t};
\item \eqn{\pmb{\beta}_t} is a subvector of \eqn{\pmb{\beta}},
holding values for all combinations of age and sex/gender
in period \eqn{t};
\item \eqn{J} is the number of elements in \eqn{\pmb{\beta}_t};
\item \eqn{S} is the number of sexes/genders;
\item \eqn{n} is \code{n_comp};
\item \eqn{\pmb{F}_s} is a known \eqn{(J/S) \times n} matrix, specific
to sex/gender \eqn{s};
\item \eqn{\pmb{g}_s} is a known vector with \eqn{J/S} elements,
specific to sex/gender \eqn{s};
\item \eqn{\pmb{F}} is a known \eqn{J \times n} matrix, with values
for all combinations of age and sex/gender; and
\item \eqn{\pmb{g}} is a known vector with \eqn{J} elements, with values
for all combinations of age and sex/gender.
}

\strong{Case 2: Interaction involving age, sex/gender, time, and one or more other variables}

When \code{joint} is \code{FALSE},

\deqn{\pmb{\beta}_{u,s,t} = \pmb{F}_s \pmb{\alpha}_{u,s,t} + \pmb{g}_s,}

and when \code{joint} is \code{TRUE},

\deqn{\pmb{\beta}_{u,t} = \pmb{F} \pmb{\alpha}_{u,t} + \pmb{g},}

where
\itemize{
\item \eqn{\pmb{\beta}} is an interaction involving sex/gender;
\item \eqn{\pmb{\beta}_{u,s,t}} is a subvector of \eqn{\pmb{\beta}},
holding values for sex/gender \eqn{s}, for the \eqn{u}th
combination of the other variables, in period \eqn{t};
\item \eqn{\pmb{\beta}_{u,t}} is a subvector of \eqn{\pmb{\beta}},
holding values for all combinatinos of age and sex/gender,
for the \eqn{u}th combination of the other variables,
in period \eqn{t};
\item \eqn{V} is the number of elements in \eqn{\pmb{\beta}_{u,t}};
\item \eqn{S} is the number of sexes/genders;
\item \eqn{n} is \code{n_comp};
\item \eqn{\pmb{F}_s} is a known \eqn{(V/S) \times n} matrix, specific
to sex/gender \eqn{s};
\item \eqn{\pmb{g}_s} is a known vector with \eqn{V/S} elements,
specific to sex/gender \eqn{s};
\item \eqn{\pmb{F}} is a known \eqn{V \times n} matrix, with values
for all combinations of age and sex/gender; and
\item \eqn{\pmb{g}} is a known vector with \eqn{V} elements, with values
for all commbinations of age and sex/gender.
}

\strong{Priors for coefficients}

With \code{SVDS_AR()}, the prior for the \eqn{k}th elements of
\eqn{\pmb{\alpha}_{s,t}} or \eqn{\pmb{\alpha}_{u,s,t}} is

\deqn{\alpha_{k,s,t} = \phi_1 \alpha_{k,s,t-1} + \cdots + \phi_n \beta_{k,s,t-n} + \epsilon_{k,s,t}}

or

\deqn{\alpha_{k,u,s,t} = \phi_1 \alpha_{k,u,s,t-1} + \cdots + \phi_n \beta_{k,u,s,t-n} + \epsilon_{k,u,t}};

with \code{SVDS_AR1()}, it is

\deqn{\alpha_{k,s,t} = \phi \alpha_{k,s,t-1} + \epsilon_{k,s,t}}

or

\deqn{\alpha_{k,u,s,t} = \phi \alpha_{k,u,s,t-1} + \epsilon_{k,u,s,t}};

with \code{SVDS_RW()}, it is

\deqn{\alpha_{k,s,t} = \alpha_{k,s,t-1} + \epsilon_{k,s,t}}

or

\deqn{\alpha_{k,u,s,t} = \alpha_{k,u,s,t-1} + \epsilon_{k,u,s,t}};

and with \code{SVDS_RW2()}, it is

\deqn{\alpha_{k,s,t} = 2 \alpha_{k,s,t-1} - \alpha_{k,s,t-2} + \epsilon_{k,s,t}}

or

\deqn{\alpha_{k,u,s,t} = 2 \alpha_{k,u,s,t-1} - \alpha_{k,u,s,t-2} + \epsilon_{k,u,s,t}}.

For more on the \eqn{\phi} and \eqn{\epsilon}, see \code{\link[=AR]{AR()}}, \code{\link[=AR1]{AR1()}},
\code{\link[=RW]{RW()}}, and \code{\link[=RW2]{RW2()}}.
}

\section{Joint or independent SVDs}{


Two possible ways of extracting patterns
from age-sex-specific data are
\enumerate{
\item carry out separate SVDs on separate datasets for
each sex/gender; or
\item carry out a single SVD on dataset that has separate
entries for each sex/gender.
}

Option 1 is more flexible. Option 2 is
more robust to sampling or measurement errors.
Option 1 is obtained by setting the \code{joint}
argument  to \code{FALSE}. Option 2
is obtained by setting the \code{joint} argument to \code{TRUE}.
The default is \code{FALSE}.
}

\section{Scaled SVDs of demographic databases in bage}{

\itemize{
\item \code{\link{HMD}} Mortality rates from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item \code{\link{LFP}} Labour forcce participation
rates from the \href{https://data-explorer.oecd.org}{OECD}.
}
}

\examples{
SVDS_AR1(HMD)
SVDS_RW(HMD, n_comp = 3)
}
\references{
\itemize{
\item For details of the construction of
scaled SVDS see the vignette
\href{https://bayesiandemography.github.io/bage/articles/vig4_svd.html}{here}.
}
}
\seealso{
\itemize{
\item \code{\link[=SVDS]{SVDS()}} SVD prior for non-time-varying age-sex or age-gender profile
\item \code{\link[=SVD_AR]{SVD_AR()}}, \code{\link[=SVD_AR1]{SVD_AR1()}}, \code{\link[=SVD_RW]{SVD_RW()}}, \code{\link[=SVD_RW2]{SVD_RW2()}} SVD priors for
for time-varying age profiles with no sex/gender dimension;
\item \code{\link[=RW]{RW()}} Smoothing via random walk
\item \code{\link[=RW2]{RW2()}} Smoothing via random walk with drift
\item \code{\link[=Sp]{Sp()}} Smoothing via splines
\item \link{priors} Overview of priors implemented in \strong{bage}
\item \code{\link[=set_prior]{set_prior()}} Specify prior for intercept,
main effect, or interaction
}
}
