% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swe_infant}
\alias{swe_infant}
\title{Infant Mortality in Sweden}
\format{
A tibble with 441 rows and the following columns:
\itemize{
\item \code{county}: A factor with 21 levels, where the levels are
ordered by number of births, from \code{"Stockholm"} down to \code{"Gotland"}
\item `time: Calendar year
\item \code{births}: Count of births
\item \code{deaths}: Count of infant deaths
}
}
\source{
Database "Live births by region, mother's age and child's sex.
Year 1968 - 2017" and database "Deaths by region, age (during the year) and sex.
Year 1968 - 2017" on the Statistics Sweden website. Downloaded on 13 July 2018.
}
\usage{
swe_infant
}
\description{
Counts of births and infant deaths in Sweden by county and year, 1995-2015
}
\details{
Dataset used in Chapter 11 of the book
\emph{Bayesian Demographic Estimation and Forecasting}.
}
\references{
Bryant J and Zhang J. 2018.
\emph{Bayesian Demographic Estimation and Forecasting}. CRC Press.
}
\seealso{
\itemize{
\item \link{datasets} Overview of datasets in \pkg{bage}
}
}
\keyword{datasets}
