% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poputils.R
\name{to_matrix}
\alias{to_matrix}
\title{Build a matrix from measure and classification variables}
\usage{
to_matrix(x, rows, cols, measure)
}
\arguments{
\item{x}{A data frame.}

\item{rows}{The classification variable(s)
used to distinguish rows in the matrix.}

\item{cols}{The classification variable(s)
used to distinguish columns in the matrix.}

\item{measure}{The measure variable, eg
some rates or counts.}
}
\value{
A matrix
}
\description{
Build a matrix where the elements are values of
a measure variable, and the rows and columns
are formed by interactions between classification
variables. The classification variables picked
out by \code{rows} and \code{cols} must uniquely identify
cells. \code{to_matrix()}, unlike \code{stats::xtabs()},
does not sum across multiple combinations of
classification variables.
}
\examples{
m1 <- to_matrix(injuries,
                rows = c(age, sex),
                cols = c(ethnicity, year),
                measure = injuries)
m1[1:5, 1:5]
dimnames(m1)

m2 <- to_matrix(injuries,
                rows = c(age, sex, ethnicity),
                cols = year,
                measure = injuries)
m2[1:5, 1:5]
dimnames(m2)

## cells not uniquely identified
try(
to_matrix(injuries,
          rows = age,
          cols = sex,
          measure = injuries)
)
}
