% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_acc_deaths}
\alias{us_acc_deaths}
\title{Accidental deaths in the USA}
\format{
A data frame with 72 rows and the
following variables:
\itemize{
\item \code{month}: Year and month.
\item \code{deaths}: Count of deaths.
}
}
\source{
Reformatted version of
\code{datasets::USAccDeaths}.
}
\usage{
us_acc_deaths
}
\description{
Counts of accidental deaths in the USA,
by month, for 1973-1978.
}
\keyword{datasets}
