% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WMD}
\alias{WMD}
\title{Scaled SVD Components from
World Marriage Database}
\format{
Object of class \code{"bage_ssvd"}.
}
\source{
Derived from data from the
World Marriage Data 2019 database,
which is assembled by the UN Population
Division from national census and survey data.
Code to create \code{WMD}
is in folder 'data-raw/ssvd_wmd'
in the source code for thet \pkg{bage}
package.
}
\usage{
WMD
}
\description{
An object of class \code{"bage_ssvd"}
holding scaled SVD components derived from
data from the
\href{https://www.un.org/development/desa/pd/data/world-marriage-data}{World Marriage Data database}.
The object holds 5 components.
}
\details{
\strong{Versions:}
\itemize{
\item \code{"v2019"} (default) data published in 2019
}
}
\keyword{datasets}
