% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WMD_C}
\alias{WMD_C}
\alias{WMD_E}
\title{Scaled SVD Components from
World Marriage Database}
\format{
Object of class \code{"bage_ssvd"}.

An object of class \code{bage_ssvd} of length 1.
}
\source{
Derived from data from the
World Marriage Data 2019 database,
which is assembled by the UN Population
Division from national census and survey data.
Code to create \code{WMD}
is in folder 'data-raw/ssvd_wmd'
in the source code for thet \pkg{bage}
package.
}
\usage{
WMD_C

WMD_E
}
\description{
Object of class \code{"bage_ssvd"}
holding scaled SVD components derived from
data from the
\href{https://www.un.org/development/desa/pd/data/world-marriage-data}{World Marriage Data database}.
Each object holds 5 components.
}
\details{
\itemize{
\item \code{WMD_C} is based on data on the proportion
of the population that is currently married.
It should be used for modelling the proportion
of people whose marital status is
"Currently Married"
\item \code{WMD_E} is based on data on the proportion
of the population that has ever been married.
It should be used for modelling the proportion
of people whose marital status is
"Ever Married".
}

In both cases "marriage" includes de facto
marriages and consensual unions.

\strong{Versions:}
\itemize{
\item \code{"v2019"} (default) data published in 2019
}
}
\seealso{
\itemize{
\item \link[=svds]{Scaled SVDs} Overview of scaled SVDs
implemented in \pkg{bage}
}
}
\keyword{datasets}
