% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{AR}
\alias{AR}
\title{Autoregressive Prior}
\usage{
AR(n = 2, s = 1)
}
\arguments{
\item{n}{The order of the model.
Default is \code{2}.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\sigma}).
Default is \code{1}.}
}
\value{
An object of class \code{bage_prior_ar}.
}
\description{
Autoregressive prior with order \code{k}.
}
\section{Mathematical description}{


The model is

\deqn{x_i = \phi_1 x_{i-1} + \cdots + \phi_k x_{i-k} + \epsilon_i}
\deqn{\epsilon_i \sim \text{N}(0, \omega^2)}

where \eqn{\omega} is chosen so that each \eqn{x_i} has
marginal variance \eqn{\sigma^2}. The value of
\eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

The \eqn{\phi_j} are contrained values between -1 and 1.
}

\examples{
AR(n = 3)
AR(n = 3, s = 2.4)
}
\references{
TMB documentation for
\href{http://kaskr.github.io/adcomp/classdensity_1_1ARk__t.html#details}{ARk}
}
\seealso{
\code{\link[=N]{N()}}, \code{\link[=RW]{RW()}}, \code{\link[=RW2]{RW2()}}, \code{\link[=Known]{Known()}}.
The values for \code{min} and \code{max} are based on the
defaults for function \code{forecast::ets()}.
\code{\link[=AR1]{AR1()}}
}
