% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{AR}
\alias{AR}
\title{Autoregressive Prior}
\usage{
AR(n = 2, s = 1)
}
\arguments{
\item{n}{The order of the model.
Default is \code{2}.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\tau}).
Default is \code{1}.}
}
\value{
An object of class \code{"bage_prior_ar"}.
}
\description{
Autoregressive prior with order \code{k}. Used to model main effects:
typically time main effects.
}
\section{Mathematical details}{


If \eqn{\beta_j} is the \eqn{j}th element of the main effect, then

\deqn{\beta_j = \phi_1 \beta_{j-1} + \cdots + \phi_k \beta_{j-k} + \epsilon_j}
\deqn{\epsilon_j \sim \text{N}(0, \omega^2)}

where \eqn{\omega} is chosen so that each \eqn{\beta_j} has
marginal variance \eqn{\sigma^2}. The value of
\eqn{\tau} has prior

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

The \eqn{\phi_1, \cdots, \phi_k} are restricted to values
in the interval \eqn{(-1,1)} that yield stationary series.
}

\examples{
AR(n = 3)
AR(n = 3, s = 2.4)
}
\references{
TMB documentation for
\href{http://kaskr.github.io/adcomp/classdensity_1_1ARk__t.html#details}{ARk}
}
\seealso{
\itemize{
\item \code{\link[=AR1]{AR1()}} Special case of \code{AR()}, though with
more options for damping.
\item \code{\link[=EAR]{EAR()}} Exchangeable version of \code{AR()},
used with interactions.
\item \link{priors} Overview of priors implemented in \code{bage}.
}
}
