% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{AR1}
\alias{AR1}
\title{AR1 Prior}
\usage{
AR1(min = 0.8, max = 0.98, s = 1)
}
\arguments{
\item{min, max}{Minimum and maximum values
for autocorrelation parameter (\eqn{\phi}).
Defaults are \code{0.8} and \code{0.98}.}

\item{s}{Scale for the prior for
the errors. Default is \code{1}.}
}
\value{
An object of class \code{"bage_prior_ar"}.
}
\description{
Autoregressive prior of order 1. Typically used with time effects.
}
\details{
Parameter \code{s} controls the size of errors. Smaller values
for \code{s} may lead to smoother series.
}
\section{Mathematical details}{


With an \code{AR1()} prior,

\deqn{\beta_j = \phi \beta_{j-1} + \epsilon_j}
\deqn{\epsilon \sim \text{N}(0, \omega^2)}

where \eqn{\beta_j} is the \eqn{j}th element of the main effect.

The value of \eqn{\omega} is chosen so that each \eqn{\beta_j} has
marginal variance \eqn{\tau^2}.

Parameter \eqn{\tau} has half-normal prior

\deqn{\tau \sim \text{N}^+(0, \text{s}^2),}

where a value for \code{s} is provided by the user.

Parameter \eqn{\phi} is constrained
to lie between \code{min} and \code{max}.
Its prior distribution is

\deqn{\phi = (\text{max} - \text{min}) \phi' - \text{min}}

where

\deqn{\phi' \sim \text{Beta}(2, 2).}
}

\examples{
AR1()
AR1(min = 0, max = 1, s = 2.4)
}
\references{
\itemize{
\item \code{AR1()} is based on the TMB function
\href{http://kaskr.github.io/adcomp/classdensity_1_1AR1__t.html#details}{AR1}
\item The values for \code{min} and \code{max} are based on the
defaults for function \code{forecast::ets()}.
}
}
\seealso{
\itemize{
\item \code{\link[=AR]{AR()}} Generalisation of \code{AR1()}
\item \code{\link[=EAR1]{EAR1()}} Exchangeable version of \code{AR1()},
used with interactions
\item \link{priors} Overview of priors implemented in \strong{bage}
}
}
