% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{AR1}
\alias{AR1}
\title{AR1 Prior}
\usage{
AR1(min = 0.8, max = 0.98, s = 1)
}
\arguments{
\item{min, max}{Minimum and maximum values
for autocorrelation parameter (\eqn{\phi}).
Defaults are \code{0.8} and \code{0.98}.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\tau}).
Default is \code{1}.}
}
\value{
An object of class \code{"bage_prior_ar"}.
}
\description{
Autoregressive prior of order 1
}
\section{Mathematical details}{


If \eqn{\beta_j} is the \eqn{j}th element of the main effect, then
\deqn{\beta_0 \sim \text{N}(0, \sigma^2)}
\deqn{\beta_j = \phi \beta_{j-1} + \sqrt{1 - \phi^2}\epsilon_j}
\deqn{\epsilon \sim \text{N}(0, \tau^2)}

\eqn{\tau} is drawn from a half-normal distribition
with scale set by the \code{s} argument.

Correlation parameter \eqn{\phi} is constrained
to lie in the interval \verb{(\\text\{min\}, \\text\{max\})}.
The prior distribution for \eqn{\phi} is

\deqn{\phi = (\text{max} - \text{min}) \phi' - \text{min}}

where

\deqn{\phi' \sim \text{beta}(2, 2)}.
}

\examples{
AR1()
AR1(min = 0, max = 1, s = 2.4)
}
\references{
TMB documentation of
\href{http://kaskr.github.io/adcomp/classdensity_1_1AR1__t.html#details}{AR1}
}
\seealso{
\itemize{
\item \code{\link[=AR]{AR()}} General case of \code{AR()}.
\item \code{\link[=EAR1]{EAR1()}} Exchangeable version of \code{AR1()},
used with interactions.
\item \link{priors} Overview of priors implemented in \code{bage}.
\item The values for \code{min} and \code{max} are based on the
defaults for function \code{forecast::ets()}.
}
}
