% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-methods.R
\name{augment.bage_mod}
\alias{augment.bage_mod}
\title{Data and values from a fitted model}
\usage{
\method{augment}{bage_mod}(x, ...)
}
\arguments{
\item{x}{A fitted \code{bage_mod} object.}

\item{...}{Unused. Included for generic consistency only.}
}
\value{
A \link[tibble:tibble-package]{tibble},
consisting of the original \code{data} argument,
to the original model function (eg \code{\link[=mod_pois]{mod_pois()}})
plus two new columns:
\itemize{
\item \code{.fitted} An \link[rvec:rvec]{rvec} holding
draws from the posterior distribution.
\item \code{.observed} Direct estimates of the rate,
probability, or mean.
}
}
\description{
Data and values from a fitted model
}
\section{Warning}{

The tidymodels
\href{https://www.tidymodels.org/learn/develop/broom/}{website}
states that the augment function may change soon.
The method here will be updated to match the new interface.
}

\examples{
mod_pois(injuries ~ age + sex + year,
         data = injuries,
         exposure = popn) |>
  fit() |>
  augment()
}
