% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-methods.R
\name{augment.bage_mod}
\alias{augment.bage_mod}
\title{Extract Data and Modelled Values}
\usage{
\method{augment}{bage_mod}(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"bage_mod"}.}

\item{quiet}{Whether to suppress messages.
Default is \code{FALSE}.}

\item{...}{Unused. Included for generic consistency only.}

\item{center}{Whether to center effects
when drawing from prior distribution.
Default is \code{TRUE}. See below for details.}
}
\value{
A \link[tibble:tibble-package]{tibble}, with the original
data plus one or more of the following columns:
\itemize{
\item \verb{.<outcome>} Corrected or extended version of
the outcome variable, in applications where the
outcome variable has missing values, or a data model
is being used.
\item \code{.observed} 'Direct' estimates of rates or
probabilities, ie counts divided by exposure or size
(in Poisson and binomial models.)
\item \code{.fitted} Draws of rates, probabilities,
or means.
\item \code{.expected} Draws of expected values for
rates or probabilities (in Poisson
that include exposure, or in binomial models.)
}

Uncertain quantities are represented using
\link[rvec:rvec]{rvecs}.
}
\description{
Extract data and rates, probabilities, or means
from a model object.
The return value consists of the original
data and one or more columns of modelled values.
}
\section{Fitted vs unfitted models}{


\code{augment()} is typically called on a \link[=fit]{fitted}
model. In this case, the modelled values are
draws from the joint posterior distribution for rates,
probabilities, or means.

\code{augment()} can, however, be called on an
unfitted model. In this case, the modelled values
are draws from the joint prior distribution.
In other words, the modelled values are informed by
model priors, and by values for \code{exposure}, \code{size}, or \code{weights},
but not by observed outcomes.
}

\section{Centering in unfitted models}{


Most sets of priors used in \strong{bage} place weak constraints
on the overall level for rates, probabilities, or means,
even when the priors place strong constraints
on variation within each main
effect or interaction. Drawing straight from these
most sets of priors
can produces rates, probabilities, or means that are
unrealistically high or low, by several orders of magnitude.

By default, when called on an unfitted
model, \code{augment()} and \code{components()} draw from
from a modified set of priors. The modification preserves
differentials across dimensions such as age and
sex, but gives more demographically-plausible
levels. The interaction is set to 0,
and all main effects and interactions are scaled to
have mean zero, with one exception. The exception is that, if the
a model has one or more terms with SVD-based priors
such as \code{\link[=SVD]{SVD()}} or \code{\link[=SVD_AR]{SVD_AR()}}, then the first such term
is not scaled. SVD-based priors are designed to
produce values demographically-plausible values without
the need for scaling.

Centering can be overriden by setting \code{center} to \code{FALSE}.
}

\section{Imputed values for outcome variable}{


\code{augment()} automatically imputes any missing
values for the outcome variable. If outcome variable
\code{var} has one or more \code{NA}s, then \code{augment}
creates a variable \code{.var}
holding original and imputed values.
}

\section{Data model for outcome variable}{


If the overall model includes a data model
for the outcome variable \code{var},
then \code{augment()} creates a new variable \code{.var} containing
estimates of the true value for the outcome.
}

\examples{
## specify model
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)

## look at prior distribution
mod |> augment()

## fit model
mod <- mod |>
  fit()

## look at posterior distribution
mod |> augment()

## insert a missing value into outcome variable
injuries_missing <- injuries
injuries_missing$injuries[1] <- NA

## fitting model and calling 'augument'
## creates a new variable called '.injuries'
## holding observed and imputed values
mod_pois(injuries ~ age + sex + year,
         data = injuries_missing,
         exposure = popn) |>
  fit() |>
  augment()

## specifying a data model for the
## original data also leads to a new
## variable called '.injuries'
mod_pois(injuries ~ age + sex + year,
         data = injuries,
         exposure = popn) |>
  set_datamod_outcome_rr3() |>
  fit() |>
  augment()
}
\seealso{
\itemize{
\item \code{\link[=components]{components()}} Extract values for hyper-parameters from a model
\item \code{\link[=tidy]{tidy()}} Short summary of a model
\item \code{\link[=mod_pois]{mod_pois()}} Specify a Poisson model
\item \code{\link[=mod_binom]{mod_binom()}} Specify a binomial model
\item \code{\link[=mod_norm]{mod_norm()}} Specify a normal model
\item \code{\link[=fit]{fit()}} Fit a model
\item \code{\link[=is_fitted]{is_fitted()}} See if a model has been fitted
\item \link{datamods} Overview of data models implemented in \strong{bage}
}
}
