% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage-package.R
\docType{package}
\name{bage-package}
\alias{bage}
\alias{bage-package}
\title{Bayesian estimation and forecasting of age-specific rates}
\description{
Bayesian estimation and forecasting of age-specific rates, probabilities, and means. Uses singular value decompositions and the package TMB (Template Model Builder), to make the estimation and forecasting scalable and fast.
}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

Authors:
\itemize{
  \item Junni Zhang \email{junni@bayesiandemography.com}
}

}
\keyword{internal}
