% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-methods.R
\name{components.bage_mod}
\alias{components.bage_mod}
\title{Extract Values for Hyper-Parameters}
\usage{
\method{components}{bage_mod}(object, center = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{"bage_mod"}.}

\item{center}{Whether to center simulation draws.
Used only with unfitted models. See below for details.}

\item{quiet}{Whether to suppress messages.
Default is \code{FALSE}.}

\item{...}{Unused. Included for generic consistency only.}
}
\value{
A \link[tibble:tibble-package]{tibble}
with four columns columns:

The return value contains the following columns:
\itemize{
\item \code{term} Model term that the hyper-parameter belongs to.
\item \code{component} Component within term.
\item \code{level} Element within component .
\item \code{.fitted} An \link[rvec:rvec]{rvec} containing
draws from the posterior distribution.
}
}
\description{
Extract values for hyper-parameters
from a model object. Includes values for
main effects and interactions, and values
for any disperion or variance terms.
}
\section{Fitted vs unfitted models}{


\code{components()} is typically called on a \link[=fit]{fitted}
model. In this case, the modelled values are
draws from the joint posterior distribution for the
hyper-parameters in the model.

\code{components()} can, however, be called on an
unfitted model. In this case, the modelled values
are draws from the joint prior distribution.
In other words, the modelled values are informed by
model priors, and by any \code{exposure}, \code{size}, or \code{weights}
argument in the model, but not by the observed outcomes.
}

\section{The 'center' argument}{


TODO - WRITE THIS
}

\examples{
## specify model
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)

## look at prior distribution
mod |> components()

## fit model
mod <- mod |>
  fit()

## look at posterior distribution
mod |> components() ## posterior distribution
}
\seealso{
\itemize{
\item \code{\link[=augment]{augment()}} Extract data and values for rates,
means, or probabilities
\item \code{\link[=tidy]{tidy()}} Extract a one-line summary of a model
\item \code{\link[=mod_pois]{mod_pois()}} Specify a Poisson model
\item \code{\link[=mod_binom]{mod_binom()}} Specify a binomial model
\item \code{\link[=mod_norm]{mod_norm()}} Specify a normal model
\item \code{\link[=fit]{fit()}} Fit a model
\item \code{\link[=is_fitted]{is_fitted()}} See if a model has been fitted
}
}
