% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-methods.R
\name{components.bage_mod}
\alias{components.bage_mod}
\title{Extract Components from a Model}
\usage{
\method{components}{bage_mod}(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{A \code{bage_mod} object.}

\item{quiet}{Whether to suppress messages.
Default is \code{FALSE}.}

\item{...}{Unused. Included for generic consistency only.}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Extract batches of hyper-parameters
from a model object
with class \code{"bage_mod"}.
\link[tibble:tibble-package]{tibble}
draws from the posterior distribution if the model has been
fitted, or the prior distribution if it has not.
}
\details{
The return value contains the following columns:
\itemize{
\item \code{term} Model term that the hyper-parameter belongs to.
\item \code{component} Component within term.
\item \code{level} Element within component .
\item \code{.fitted} An \link[rvec:rvec]{rvec} containing
draws from the posterior distribution.
}
}
\examples{
## specify model
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)

## look at prior distribution
mod |> components()

## fit model
mod <- mod |>
  fit()

## look at posterior distribution
mod |> components() ## posterior distribution
}
\seealso{
\itemize{
\item \code{\link[=augment]{augment()}} A tibble with data and parameters.
\item \code{\link[=tidy]{tidy()}} A one-line summary of a model.
\item \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}}, \code{\link[=mod_norm]{mod_norm()}} Functions
to specify a model
\item \code{\link[=fit]{fit()}} Function to fit a model
}
}
