% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{compose_time}
\alias{compose_time}
\title{Compose a Prior for a Time Main Effect or Interaction}
\usage{
compose_time(trend, cyclical = NULL, seasonal = NULL, error = NULL)
}
\arguments{
\item{trend}{Prior describing the long-run behavior
of the series. See below for choices. Required.}

\item{cyclical}{Prior describing departures from the
long-run trend that do not have a fixed
period. See below for choices. Optional.}

\item{seasonal}{Prior describing departures from the
long-run trend that have a fixed period.
See below for choices. Optional.}

\item{error}{Prior for additional idiosyncratic variation.
See below for choices. Optional}
}
\value{
An object of class \code{"bage_prior_compose"}
}
\description{
Create a composite prior for a time main effect,
or for an interaction involving time.
The composite prior always contains a trend,
and may contain a cyclical effect,
a seasonal effect, and an error.
}
\details{
\tabular{lll}{
   Term \tab Main effects \tab Interactions \cr
   \code{trend} \tab \code{\link[=Lin]{Lin()}}, \code{\link[=RW]{RW()}}, \code{\link[=RW2]{RW2()}}, \code{\link[=Sp]{Sp()}} \tab \code{\link[=ILin]{ILin()}} \cr
   \code{seasonal} \tab \code{\link[=Seas]{Seas()}} \tab \code{\link[=ISeas]{ISeas()}} \cr
   \code{cyclical} \tab \code{\link[=AR]{AR()}}, \code{\link[=AR1]{AR1()}} \tab \code{\link[=IAR]{IAR()}}, \code{\link[=IAR1]{IAR1()}} \cr
   \code{error} \tab \code{\link[=N]{N()}} \tab \code{\link[=N]{N()}} \cr
}
}
\examples{
compose_time(
  trend = Lin(),
  cyclical = AR()
)

compose_time(
  trend = RW2(),
  error = N()
)

compose_time(
  trend = ILin(),
  cyclical = IAR(),
  season = ISeas(n = 4)
)
}
