% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamods.R
\name{datamods}
\alias{datamods}
\title{Data Models}
\description{
The models for rates, probabilities, or means
created with functions \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}},
and \code{\link[=mod_norm]{mod_norm()}} can be extended by adding data models,
also referred to as measurement error models.
Data models can be applied to the outcome variable,
the exposure variable (in Poisson models), or the
size variable (in binomial models).
}
\details{
\strong{Data models for outcome variable}\tabular{lllll}{
   \strong{Function} \tab \strong{Assumptions about data} \tab \strong{pois} \tab \strong{binom} \tab \strong{norm} \cr
   \code{\link[=set_datamod_outcome_rr3]{set_datamod_outcome_rr3()}} \tab Outcome randomly rounded to base 3 \tab Yes \tab Yes \tab No \cr
}


\strong{Data models for exposure or size variable}

\emph{None implemented yet}
}
