% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamods.R
\name{datamods}
\alias{datamods}
\title{Data Models}
\description{
The models for rates, probabilities, or means
created with functions \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}},
and \code{\link[=mod_norm]{mod_norm()}} can be extended by adding data models,
also referred to as measurement error models.
}
\details{
\tabular{lllll}{
   \strong{Function} \tab \strong{Assumptions about measurement error} \tab \strong{Poisson} \tab \strong{Binomial} \tab \strong{Normal} \cr
   \code{\link[=set_datamod_miscount]{set_datamod_miscount()}} \tab Reported outcome has undercount and overcount \tab Yes \tab No \tab No \cr
   \code{\link[=set_datamod_undercount]{set_datamod_undercount()}} \tab Reported outcome has undercount \tab Yes \tab Yes \tab No \cr
   \code{\link[=set_datamod_overcount]{set_datamod_overcount()}} \tab Reported outcome has overcount \tab Yes \tab No \tab No \cr
   \code{\link[=set_datamod_noise]{set_datamod_noise()}} \tab Reported outcome unbiased, but with positive and negative measurement errors \tab Yes* \tab No \tab Yes \cr
   \code{\link[=set_datamod_exposure]{set_datamod_exposure()}} \tab Reported exposure unbiased, but with positive and negative measurement errors \tab Yes* \tab No \tab No \cr
}


*Models with no dispersion term for rates.
}
