% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{EAR}
\alias{EAR}
\title{Exchangeable Autoregressive Prior}
\usage{
EAR(n = 2, s = 1, along = NULL)
}
\arguments{
\item{n}{The order of the model, i.e. the number
of lagged terms included in the model. The default
is \code{2}.}

\item{s}{Scale for the prior for the errors.
Default is \code{1}.}

\item{along}{Name of one of the variables
in the interaction. Optional, provided
the interaction includes a time or age variable.}
}
\value{
An object of class \code{"bage_prior_ear"}.
}
\description{
An \code{EAR()} prior is designed for use with interactions,
typically involving time or age.

With this prior, an autogressive time series model is
applied to the 'along'
variable, within each combination
of the 'by' variables. For instance,
an autoregressive time series
model might be applied to the time variable,
within each combination of age and sex variables.
The same set of coefficients is shared
across all combinations of the 'by' variables.
The time series are "exchangeable" in the sense
that they are drawn from a common distribution.
}
\section{'Along' and 'by' variables}{


Some priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the 'along' variable is likely to be time,
and the 'by' variables are likely to be region and sex.

If no \code{along} argument is supplied, a default value
is chosen as follows:
\itemize{
\item if the interaction contains a time variable, then
the 'along' variable is assumed to be time;
\item otherwise, if the interaction contains an
age variable, then the 'along' variable
is assumed to be age;
\item otherwise, an error is raised.
}
}

\section{Mathematical details}{


With an \code{EAR()} prior,

\deqn{\beta_{u,v} = \phi_1 x_{u,v-1} + \cdots + \phi_n x_{u,v-n} + \epsilon_{u,v}}
\deqn{\epsilon_{u,v} \sim \text{N}(0, \omega^2)}

where
\itemize{
\item \eqn{u} denotes a position within the the 'along' variable;
\item \eqn{v} denotes a position within the classification
formed by the 'by' variable(s);
\item \eqn{\beta_{u,v}} is an element of an interaction;
\item \eqn{\phi_1, \dots, \phi_n} are autoregression coefficients; and
\item \eqn{\epsilon_{u,v}} is an error term.
}

The value for \eqn{\omega} is chosen so that each
\eqn{\beta_{u,v}} has marginal variance \eqn{\tau^2}.

Parameter \eqn{\tau} has prior

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

The \eqn{\phi_1, \cdots, \phi_k} are jointly restricted to values
between -1 and 1 that lead to stationary models. The quantity
\eqn{r = \sqrt{\phi_1^2 + \cdots + \phi_k^2}} is given a
boundary-avoiding prior

\deqn{r \sim \text{Beta}(2, 2)}.
}

\examples{
EAR(n = 3)
EAR(n = 3, s = 2.4)
EAR(along = "cohort")
}
\references{
\itemize{
\item \code{EAR()} is based partly on the TMB function
\href{http://kaskr.github.io/adcomp/classdensity_1_1ARk__t.html#details}{ARk}
}
}
\seealso{
\itemize{
\item \code{\link[=EAR1]{EAR1()}} Special case of \code{EAR()}
\item \code{\link[=AR]{AR()}} Autoregressive prior for main effects
\item \link{priors} Overview of priors implemented in \strong{bage}
}
}
