% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{EAR1}
\alias{EAR1}
\title{Exchangeable AR1 Prior}
\usage{
EAR1(min = 0.8, max = 0.98, s = 1)
}
\arguments{
\item{min, max}{Minimum and maximum values
for autocorrelation parameter (\eqn{\phi}).
Defaults are \code{0.8} and \code{0.98}.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\sigma}).
Default is \code{1}.}
}
\value{
An object of class \code{bage_prior_ear}.
}
\description{
Autogressive prior for an interaction,
where an AR model order 1 is applied
to the "along" variable, within each
combination of values of the "by" variable.
The damping coefficient is shared across
different combinations of the "by" variables.
The series within each combination of the
'by' variables are, however, treated as
exchangeable.
}
\section{Mathematical description}{


The model is

\deqn{x_{u,v} = \phi x_{u,i-1} + \epsilon_{u,v}}
\deqn{\epsilon_{u,v} \sim \text{N}(0, \omega^2)}

where \eqn{\omega} is chosen so that each \eqn{x_{u,v}} has
marginal variance \eqn{\sigma^2}. The value of
\eqn{\sigma} has prior

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

Correlation parameter \eqn{\phi} is constrained
to lie in the interval \verb{(a, b)},
where \eqn{a} = \code{min} and \eqn{b} = \code{max}.
The prior distribution is for \eqn{\phi} is
\deqn{\phi = (b - a) \phi' - a}
where
\deqn{\phi' \sim \text{beta}(2, 2)}.
}

\section{'Along' and 'by' variables}{


Multivariate priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the by variables are likely to be region and sex.

If no \code{along} argument is supplied, then:
\itemize{
\item if the interaction contains a time variable, then
it is assumed to be the 'along' variable;
\item otherwise, if the interaction contains an
age variable, then it is assumed to be
the 'along' variable;
\item otherwise, an error is raised.
}
}

\examples{
EAR1()
EAR1( min = 0, max = 1, s = 2.4)
}
\references{
TMB documentation for
\href{http://kaskr.github.io/adcomp/classdensity_1_1ARk__t.html#details}{ARk}
}
\seealso{
\code{\link[=N]{N()}}, \code{\link[=RW]{RW()}}, \code{\link[=RW2]{RW2()}}, \code{\link[=Known]{Known()}}.
The values for \code{min} and \code{max} are based on the
defaults for function \code{forecast::ets()}.
\code{\link[=EAR]{EAR()}}
}
