% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{EAR1}
\alias{EAR1}
\title{Exchangeable AR1 Prior}
\usage{
EAR1(min = 0.8, max = 0.98, s = 1, along = NULL)
}
\arguments{
\item{min, max}{Minimum and maximum values
for autocorrelation parameter.
Defaults are \code{0.8} and \code{0.98}.}

\item{s}{Scale for the prior for the errors.
Default is \code{1}.}

\item{along}{Name of one of the variables
in the interaction. Optional, provided
the interaction includes a time or age variable.}
}
\value{
An object of class \code{"bage_prior_ear"}.
}
\description{
An \code{EAR1()} prior is designed for use with interactions,
typicallly involving time or age.
}
\details{
With this prior, an AR1 time series model
is applied to the 'along' variable, within
each combination of the 'by' variables.
For instance, an AR1 model might be applied
to the time veriable, which each combination
of age and sex variables.
The same coefficient is shared across
combinations of the 'by' variables.
The time series are "exchangeable" in the sense
that they are drawn from a common distribution.
}
\section{Mathematical details}{


With an \code{EAR1()} prior,

\deqn{\beta_{u,v} = \phi \beta_{u,v-1} + \epsilon_{u,v}}
\deqn{\epsilon_{u,v} \sim \text{N}(0, \omega^2)}

where
\itemize{
\item \eqn{u} denotes a position within the the 'along' variable;
\item \eqn{v} denotes a position within the classification
formed by the 'by' variable(s);
\item \eqn{\beta_{u,v}} is an element of an interaction;
\item \eqn{\phi} is the autoregression coefficient; and
\item \epsilon_{u,v} is an error term.
}

The value for \eqn{\omega} is chosen so that each
\eqn{\beta_{u,v}} has marginal variance \eqn{\tau^2}.

Parameter \eqn{\tau} has prior

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

Correlation parameter \eqn{\phi} is constrained
to lie in the interval \verb{(\\text\{min\}, \\text\{max\})}.
The prior distribution for \eqn{\phi} is

\deqn{\phi = (\text{max} - \text{min}) \phi' - \text{min}}

where

\deqn{\phi' \sim \text{Beta}(2, 2)}.
}

\section{'Along' and 'by' variables}{


Some priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the 'along' variable is likely to be time,
and the 'by' variables are likely to be region and sex.

If no \code{along} argument is supplied, a default value
is chosen as follows:
\itemize{
\item if the interaction contains a time variable, then
the 'along' variable is assumed to be time;
\item otherwise, if the interaction contains an
age variable, then the 'along' variable
is assumed to be age;
\item otherwise, an error is raised.
}
}

\examples{
EAR1()
EAR1(min = 0, max = 1, s = 2.4)
EAR1(along = "cohort")
}
\references{
\itemize{
\item `EAR1()' is partly based on the TMB function
\href{http://kaskr.github.io/adcomp/classdensity_1_1ARk__t.html#details}{ARk}
\item The values for \code{min} and \code{max} are based on the
defaults for function \code{forecast::ets()}.
}
}
\seealso{
\itemize{
\item \code{\link[=EAR]{EAR()}} More general exchangeable AR model
\item \code{\link[=AR1]{AR1()}} AR1 prior for main effects
\item \link{priors} Overview of priors implemented in \strong{bage}
}
}
