% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{ELin}
\alias{ELin}
\title{Exchangeable Linear Prior}
\usage{
ELin(s = 1, sd = 1, ms = 1, along = NULL)
}
\arguments{
\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\tau}).
Defaults to 1.}

\item{sd}{A postive number. Default is 1.}

\item{ms}{A positive number. Default is 1.}

\item{along}{Name of one of the dimensions
in the interaction. Optional, provided
the data contain a time or age dimension.}
}
\value{
An object of class \code{bage_prior_elin}.
}
\description{
Prior for an interaction where, within each combination
of the 'by' variables, the elements of the 'along'
variable follow (approximately) a straight line.
The slopes of the line vary across
different combinations of 'by' variables'.
}
\section{Statistical model}{


The model is

\deqn{x_{uv} \sim \text{N}(\text{slope}_u q_v, \sigma^2)}

where
\itemize{
\item \eqn{u} is the index for a combination of 'by' variables,
\item \eqn{v} is the index of the 'along' variable,
\item \eqn{q_j} is a rescaled version of \eqn{j}, with mean 0,
minimum -1, and maximum 1: \eqn{q_j = - (J+1)/(J-1) + 2j/(J-1)}.
}

The slopes are drawn from a common distribution with mean
\eqn{\text{slope}}.

\deqn{\text{slope}_u \sim \text{N}(\text{slope}, \omega^2)}

Larger absolute values for \eqn{\text{slope}} imply
steeper lines. The absolute value of \eqn{slope}
is governed by parameter \code{s}:

\deqn{\text{slope} ~ \text{N}(0, \text{s}^2)}.

Larger values for \eqn{\omega} imply more variability
in slopes across different combinations of the 'by'
variables. The size of \eqn{\omega} is governed by
parameter \code{ms}

\deqn{\omega \sim \text{N}^+(0, \text{ms}^2)}

Larger values for \eqn{\sigma} imply more variability
around each line. The size of \eqn{\sigma} is
governed by parameter \code{s}:

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(\eqn{\text{N}^} denotes a half-normal distribution,
which has the same shape as a normal
distribution, but is defined only for non-negative
values.)
}

\section{'Along' and 'by' variables}{


Some priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the 'along' variable is likely to be time,
and the 'by' variables are likely to be region and sex.

If no \code{along} argument is supplied, a default value
is chosen as follows:
\itemize{
\item if the interaction contains a time variable, then
the 'along' variable is assumed to be time;
\item otherwise, if the interaction contains an
age variable, then the 'along' variable
is assumed to be age;
\item otherwise, an error is raised.
}
}

\examples{
ELin()
ELin(s = 0.5, sd = 2, ms = 0.1, along = "cohort")
}
\seealso{
\itemize{
\item \code{\link[=N]{N()}} etc
\item \code{\link[=set_var_time]{set_var_time()}} to specify the time variable
\item \code{\link[=set_var_age]{set_var_age()}} to specify the age variable
}
}
