% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{ERW}
\alias{ERW}
\title{Exchangeable Random Walk Prior}
\usage{
ERW(s = 1, along = NULL)
}
\arguments{
\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\sigma}).
Defaults to 1.}

\item{along}{Name of one of the dimensions
in the interaction. Optional, provided
the data contain a time or age dimension.}
}
\value{
An object of class \code{bage_prior_erw}.
}
\description{
Prior for an interaction,
where a (first order) random walk model
is applied to the "along" variable, within each
combination of values of the "by" variable.
Standard deviations are shared across
different combinations of the "by" variables.
The series within each combination of the
'by' variables are, however, treated as exchangeable.
}
\section{Mathematical description}{


The model is

\deqn{x_{u,v} = x_{u,i-1} + \epsilon_{u,v}}
\deqn{\epsilon_{u,v} \sim \text{N}(0, \sigma^2)}

Standard deviation \eqn{\sigma} is drawn from a
half-normal distribution,

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{scale} lead to smoother series of \code{x}s, and
higher values lead to rougher series.
}

\section{'Along' and 'by' variables}{


Multivariate priors for interactions distinguish
between 'along' and 'by' variables. The 'along'
variable is typically time, or, in interactions
not involving time, is typically age. The 'by'
variables are everything else. In an interaction
between region, sex, and time, for instance,
the by variables are likely to be region and sex.

If no \code{along} argument is supplied, then:
\itemize{
\item if the interaction contains a time variable, then
it is assumed to be the 'along' variable;
\item otherwise, if the interaction contains an
age variable, then it is assumed to be
the 'along' variable;
\item otherwise, an error is raised.
}
}

\examples{
ERW()
ERW(s = 0.5)
}
\seealso{
\code{\link[=N]{N()}}, \code{\link[=RW]{RW()}}, \code{\link[=RW2]{RW2()}}, \code{\link[=Known]{Known()}}.
}
