% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a model}
\usage{
fit(mod)
}
\arguments{
\item{mod}{An object of class \code{bage_mod},
typically created by a call to \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}
}
\value{
An object of class \code{bage_mod}
}
\description{
Fit a model
}
