% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-methods.R
\name{forecast.bage_mod}
\alias{forecast.bage_mod}
\title{Use a Model to Construct Forecasts}
\usage{
\method{forecast}{bage_mod}(
  object,
  output = c("augment", "components"),
  include_estimates = TRUE,
  labels,
  ...
)
}
\arguments{
\item{object}{An fitted model.}

\item{output}{Type of output returned}

\item{include_estimates}{Whether to
include historical estimates along
with the forecasts. Default is \code{TRUE}.}

\item{labels}{Labels for future values. WARNING
this argument is only temporary.}

\item{...}{Not currently used.}
}
\value{
A \code{bage_mod} object
}
\description{
DESCRIBE HOW FORECASTS CONSTRUCTED.
}
\section{Type of output}{


\code{forecast()} always returns a \link[tibble:tibble]{tibble}.
However, the contents of the tibble depend on the
\code{output} argument:
\itemize{
\item \code{"augment"} (The default.) Like the
output from \code{\link[=augment]{augment()}}, but without the
outcome variable or any exposure,
size, or weights variables.
\item \code{"components"} Like the output from
\code{\link[=components]{components()}}.
}
}

