% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HFD}
\alias{HFD}
\title{Scaled SVD Components from Human Fertility Database}
\format{
Object of class \code{"bage_ssvd"}.
}
\source{
Derived from data from the
Human Fertility Database.Max Planck Institute
for Demographic Research (Germany) and Vienna
Institute of Demography (Austria). .
Code to create \code{HFD}
is in folder 'data-raw/ssvd_hfd'
in the source code for the \pkg{bage}
package.
}
\usage{
HFD
}
\description{
An object of class \code{"bage_ssvd"}
holding scaled SVD components derived from
data from the
\href{https://www.humanfertility.org}{Human Fertility Database}.
\code{HFD} holds 5 components.
}
\details{
\strong{Versions:}
\itemize{
\item \code{"v2024"} (default) version of data from
23 October 2024
}
}
\seealso{
\itemize{
\item \link[=svds]{Scaled SVDs} Overview of scaled SVDs
implemented in \pkg{bage}
}
}
\keyword{datasets}
