% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HFD}
\alias{HFD}
\title{Scaled SVD Components from Human Fertility Database}
\format{
Object of class \code{"bage_ssvd"}.

Versions:
\itemize{
\item \code{"v2025"} (default) Data published on 2025-07-24
\item \code{"v2024"} Data published on October 2024-10-23
}
}
\source{
Derived from data from the
\href{https://www.humanfertility.org}{Human Fertility Database}.
Max Planck Institute
for Demographic Research (Germany) and Vienna
Institute of Demography (Austria).
Code to create \code{HFD}
is in folder \file{data-raw/ssvd_hfd}
in the source code for the \pkg{bage}
package.
}
\usage{
HFD
}
\description{
An object of class \code{"bage_ssvd"}
holding scaled SVD components derived from
data from the Human Fertility Database.
\code{HFD} holds 5 components.
}
\seealso{
\itemize{
\item \link[=svds]{Scaled SVDs} Overview of scaled SVDs
implemented in \pkg{bage}
\item \code{\link[=SVD]{SVD()}} A prior based on a scaled SVD
}
}
\keyword{datasets}
