% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HMD}
\alias{HMD}
\title{Components from Human Mortality Database}
\format{
An object of class \code{bage_scaled_svd} of length 1.
}
\source{
Derived from data at
Human Mortality Database. Max Planck Institute for
Demographic Research (Germany), University of California,
Berkeley (USA), and French Institute for Demographic Studies
(France). Available at www.mortality.org.
Code for processing the data is
\href{https://github.com/bayesiandemography/svd_hmd}{here}.
}
\usage{
HMD
}
\description{
An object of class \link[=scaled_svd]{bage_scaled_svd}
holding components extracted from mortality
data from the \href{https://www.mortality.org}{Human Mortality Database}.
}
\keyword{datasets}
