% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HMD}
\alias{HMD}
\title{Components from Human Mortality Database}
\format{
Object of class \code{"bage_ssvd"}.Versions
\itemize{
\item \code{"v2025"} (default) data published on
2025-09-25, all countries and years
\item \code{"v2025-50"} data published on
2025-09-25, all countries, 1950 and later
\item \code{"v2024"} data published on
2024-02-26, all countries and years
}
}
\source{
Derived from data from the
Human Mortality Database. Max Planck Institute for
Demographic Research (Germany), University of California,
Berkeley (USA), and French Institute for Demographic Studies
(France). Code to create \code{HMD}
is in folder 'data-raw/ssvd_hmd'
in the source code for \strong{bage}
package.
}
\usage{
HMD
}
\description{
An object of class \code{"bage_ssvd"}
holding components extracted from mortality
data from the
\href{https://www.mortality.org}{Human Mortality Database}.
The object holds 5 components.
}
\keyword{datasets}
