% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMD}
\alias{HMD}
\title{Scaled SVD Components from
Human Mortality Database}
\format{
Object of class \code{"bage_ssvd"}.

Versions:
\itemize{
\item \code{"v2025"} (default) Data published on
2025-09-25, all years
\item \code{"v2025-50"} Data published on
2025-09-25, 1950 and later
\item \code{"v2024"} Data published on
2024-02-26, all years
}
}
\source{
Derived from data from the
\href{https://www.mortality.org}{Human Mortality Database}.
Max Planck Institute for
Demographic Research (Germany), University of California,
Berkeley (USA), and French Institute for Demographic Studies
(France). Code to create \code{HMD}
is in folder \file{data-raw/ssvd_hmd}
in the source code for the \pkg{bage}
package.
}
\usage{
HMD
}
\description{
An object of class \code{"bage_ssvd"}
holding scaled SVD components derived from
data from the
Human Mortality Database.
\code{HMD} holds 5 components.
}
\seealso{
\itemize{
\item \link[=svds]{Scaled SVDs} Overview of scaled SVDs
implemented in \pkg{bage}
\item \code{\link[=SVD]{SVD()}} A prior based on a scaled SVD
}
}
\keyword{datasets}
