% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{injuries}
\alias{injuries}
\title{Fatal Injuries in New Zealand}
\format{
A \link[tibble:tibble-package]{tibble}
with 912 rows and the following columns:
\itemize{
\item \code{age}: 5-year age groups, up to age 55-59
\item \code{sex}: \code{"Female"} or \code{"Male"}
\item \code{ethnicity}: \code{"Maori"} or \code{"Non Maori"}
\item \code{year}: Calendar year
\item \code{injuries}: Count of injuries, randomly rounded to base 3
\item \code{popn}: Population on 30 June
}
}
\source{
Derived from data in tables "Estimated Resident
Population by Age and Sex (1991+) (Annual-Jun)" and
"Maori Ethnic Group Estimated Resident Population
by Age and Sex (1991+) (Annual-Jun)" in the online
database Infoshare, and table
"Count of fatal and serious non-fatal
injuries by sex, age group, ethnicity, cause, and severity
of injury, 2000-2021" in the online database NZ.Stat,
on the Statistics New Zealand website.
Data downloaded on 1 January 2023.
}
\usage{
injuries
}
\description{
Counts of fatal injuries in New Zealand, by age, sex,
ethnicity, and year, plus estimates of the population
at risk.
}
\keyword{datasets}
