% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{Known}
\alias{Known}
\title{Known Prior}
\usage{
Known(values)
}
\arguments{
\item{values}{A numeric vector}
}
\value{
An object of class \code{"bage_prior_known"}.
}
\description{
Treat an intercept, a main effect, or an interaction
as fixed and known.
}
\examples{
Known(-2.3)
Known(c(0.1, 2, -0.11))
}
\seealso{
\itemize{
\item \code{\link[=NFix]{NFix()}} Prior where level unknown, but variability known.
\item \link{priors} Overview of priors implemented in \strong{bage}
\item \code{\link[=set_prior]{set_prior()}} Specify prior for intercept,
main effect, or interaction
}
}
