% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{kor_births}
\alias{kor_births}
\title{Births in South Korea}
\format{
A \link[tibble:tibble-package]{tibble} with 1,872 rows and the following
columns:
\itemize{
\item \code{age} Five-year age groups from \verb{"10-14" to }"50-54"`
\item \code{region} Administrative region
\item \code{time} Calendar year, 2011-2023
\item \code{births} Counts of births
\item \code{popn} Mid-year population
\item \code{gdp_pc_2023} Regional GDP per capita in 2023
\item \code{dens_2020} Regional population density
(people per km-squared) in 2020
}
}
\source{
Tables "Live Births by Age Group of Mother,
Sex and Birth Order for Provinces",
and "Resident Population in Five-Year Age Groups",
on the Korean Statistical Information Service website.
Data downloaded on 24 September 2024. Data on GDP per capita
and population density from Wikipedia
https://w.wiki/DMFA, data downloaded on 8 March 2025, and
https://w.wiki/DMF9, data downloaded on 8 March 2025.
}
\usage{
kor_births
}
\description{
Births and mid-year population by age of mother,
region, and calendar year,
2011-2023, plus regional data on GDP per capita and
population density.
}
\keyword{datasets}
