% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{Lin}
\alias{Lin}
\title{Linear Prior}
\usage{
Lin(s = 1, sd = 1, along = NULL)
}
\arguments{
\item{s}{Scale for the prior for the errors.
Smaller values can lead to smoother series.
Default is \code{1}.}

\item{sd}{Standard deviation in prior for slope
of line. Default is 1.}

\item{along}{Name of the variable to be used
as the "along" variable. Only used with
interactions.}
}
\value{
An object of class \code{"bage_prior_lin"}.
}
\description{
Prior for main effect or interaction.
A straight line or lines, combined with independent
normal errors. Typically used with time.
}
\details{
If \code{Lin()} is used with an interaction,
then separate lines are constructed along
the "along" variable, within each combination
of the "by" variables.

Argument \code{s} controls the size of the errors.
}
\section{Mathematical details}{


When \code{Lin()} is used with a main effect,

\deqn{\beta_j = \eta q_j + \epsilon_j}
\deqn{\epsilon_j \sim \text{N}(0, \tau^2),}

and when it is used with an interaction,

\deqn{\beta_{u,v} \sim \eta_u q_v + \epsilon_{u,v}}
\deqn{\epsilon_{u,v} \sim \text{N}(0, \tau^2),}

where
\itemize{
\item \eqn{\pmb{\beta}} is the main effect or interaction;
\item \eqn{j} denotes position within the main effect;
\item \eqn{u} denotes position within the "along" variable of the interaction;
\item \eqn{u} denotes position within the "by" variable(s) of the interaction;
\item \deqn{q = - (J+1)/(J-1) + 2j/(J-1);} and
\item \deqn{q_v = - (V+1)/(V-1) + 2v/(V-1)}.
}

The slopes have priors
\deqn{\eta \sim \text{N}(0, \text{sd}^2)}
and
\deqn{\eta_u \sim \text{N}(0, \text{sd}^2).}
Larger values for \code{sd} permit steeper slopes.

Parameter \eqn{\tau} has a half-normal prior
\deqn{\tau \sim \text{N}^+(0, \text{s}^2),}
where \code{s} is provided by the user.
}

\examples{
Lin()
Lin(s = 0.5, sd = 2)
Lin(along = "cohort")
}
\seealso{
\itemize{
\item \code{\link[=LinAR]{LinAR()}} Linear with AR errors
\item \code{\link[=LinAR1]{LinAR1()}} Linear with AR1 errors
\item \code{\link[=RW2]{RW2()}} Random walk with drift
\item \link{priors} Overview of priors implemented in \strong{bage}
\item \code{\link[=set_prior]{set_prior()}} Specify prior for intercept,
main effect, or interaction
}
}
