% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{Lin}
\alias{Lin}
\title{Linear Prior}
\usage{
Lin(s = 1, sd = 1)
}
\arguments{
\item{s}{A positive number. Default is 1.}

\item{sd}{A postive number. Default is 1.}
}
\value{
An object of class \code{bage_prior_known}.
}
\description{
Prior in which units are assumed to follow a
straight line (with idiosyncratic errors around
the line).
}
\details{
\deqn{x_j \sim \text{N}(\text{slope} q_j, \sigma^2)}

where \eqn{q_j} is a rescaled version of \eqn{j}, with mean 0,
minimum -1, and maximum 1: \eqn{q_j = - (J+1)/(J-1) + 2j/(J-1)}.

The slope term is drawn from a normal distribution,

\deqn{\text{slope} ~ \text{N}(0, \text{sd}^2)}.

The standard deviation for the normal distribution,
\code{sd}, defaults to 1, but can be set to other values.
Lower values for \code{sd} lead to smaller values for
the slope term.

Standard deviation \eqn{\sigma} is drawn from a
half-normal distribution,

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{scale} lead to a smoother line.
}
\examples{
Lin()
Lin(s = 0.5, sd = 2)
}
\seealso{
\itemize{
\item \code{\link[=N]{N()}} etc
}
}
