% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{LinAR}
\alias{LinAR}
\title{Linear Prior with Autoregressive Errors}
\usage{
LinAR(n = 2, s = 1, sd = 1, along = NULL)
}
\arguments{
\item{n}{The order of the model for the
errors, i.e. the number of lagged
terms that are included. Default is \code{2}.}

\item{s}{Scale for the prior for the errors.
Smaller values imply a smoother series. Default is 1.}

\item{sd}{Standard deviation in the prior for
the slope of the line. Larger values imply
steeper slopes. Default is 1.}
}
\value{
An object of class \code{"bage_prior_linar"}.
}
\description{
Prior for main effect or interaction.
A straight line or lines, combined with
autoregressive errors. Typically used with time.
}
\details{
If \code{LinAR()} is used with an interaction,
then, within each combination of the "by" variables,
the "along" variable is modelled as a straight line
with autoregressive errors.

Argument \code{s} controls the size of the errors.
}
\section{Mathematical details}{


When \code{LinAR()} is being used with a main effect,

\deqn{\beta_j = \eta q_j + \epsilon_j}
\deqn{\epsilon_j = \phi_1 \epsilon_{j-1} + \cdots + \phi_n \epsilon_{j-n} + \varepsilon_j,}
\deqn{\varepsilon_j \sim \text{N}(0, \omega^2).}

and when it is being used with an interaction,

\deqn{\beta_{u,v} = \eta_j q_{u,v} + \epsilon_{u,v}}
\deqn{\epsilon_{u,v} = \phi_1 \epsilon_{u,v-1} + \cdots + \phi_n \epsilon_{u,v-n} + \varepsilon_{u,v},}
\deqn{\varepsilon_{u,v} \sim \text{N}(0, \omega^2).}

where
\itemize{
\item \eqn{\pmb{\beta}} is the main effect or interaction;
\item \eqn{j} denotes position within the main effect;
\item \eqn{u} denotes position within the "along" variable of the interaction;
\item \eqn{u} denotes position within the "by" variable(s) of the interaction;
\item \eqn{q = - (J+1)/(J-1) + 2j/(J-1);} and
\item \eqn{q_v = - (V+1)/(V-1) + 2v/(V-1)}.
}

The slopes have priors
\deqn{\eta \sim \text{N}(0, \text{sd}^2)}
and
\deqn{\eta_u \sim \text{N}(0, \text{sd}^2).}
Larger values for \code{sd} permit steeper slopes.

Internally, \code{LinAR()} derives a value for \eqn{\omega} that
gives \eqn{\epsilon_j} or \eqn{\epsilon_{u,v}} a marginal
variance of \eqn{\tau^2}. Parameter \eqn{\tau}
has a half-normal prior
\deqn{\tau \sim \text{N}^+(0, \text{s}^2),}
where a value for \code{s} is provided by the user.

The \eqn{\phi_1, \cdots, \phi_k} are restricted to values
between -1 and 1 that jointly lead to a stationary model. The quantity
\eqn{r = \sqrt{\phi_1^2 + \cdots + \phi_k^2}} is given a
boundary-avoiding prior

\deqn{r \sim \text{Beta}(2, 2).}

Slope \eqn{\eta} is drawn from a normal distribution,
\deqn{\eta \sim \text{N}(0, (\text{sd})^2).}
}

\examples{
LinAR()
LinAR(n = 3, s = 0.5, sd = 2)
}
\seealso{
\itemize{
\item \code{\link[=LinAR1]{LinAR1()}} Special case of \code{LinAR()}
\item \code{\link[=AR]{AR()}} AR process with no line
\item \link{priors} Overview of priors implemented in \strong{bage}
\item \code{\link[=set_prior]{set_prior()}} Specify prior for intercept,
main effect, or interaction
}
}
