% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-constructors.R
\name{mod_norm}
\alias{mod_norm}
\title{Specify a normal model}
\usage{
mod_norm(formula, data, weights)
}
\arguments{
\item{formula}{An R \link[stats:formula]{formula},
specifying the outcome and predictors.}

\item{data}{A data frame containing the outcome,
predictors, and, optionally, weights.}

\item{weights}{Name of the weights variable,
or a \code{1}.

Internally, outcome scaled to have mean 0 and sd 1;
weights scaled to have mean 1.}
}
\value{
An object of class \code{bage_mod}.
}
\description{
Specify a model where the outcome is drawn from
a normal distribution.
}
\details{
\itemize{
\item \code{formula} specifies the outcome and predictors,
including interactions between predictors.
It follows standard R \link[stats:formula]{formula}
conventions, except that it cannot include
transformations (e.g. \code{sqrt(deaths)}).
\item \code{data} holds the outcome, the predictors, and,
optionally, weights.
\item \code{weights} is the name (bare or quoted) of the variable
in \code{data} used as weights, or, if the
model does include weights, a \code{1}.
}

If the model includes weights, then the
the first level of the model is

\deqn{y \sim \text{norm}(\mu, \sigma^2 / w)}

where \eqn{\mu} is the underlying rate, and
\eqn{w} is weights. If the model does not
include weights, then the first level is

\deqn{y \sim \text{norm}(\mu, \sigma^2)}

TODO - Include error term once specification finalised.
}
\examples{
mod <- mod_norm(value ~ diag:age + year,
                data = expenditure,
                weights = 1)
}
\seealso{
\itemize{
\item \code{\link[=mod_pois]{mod_pois()}} and \code{\link[=mod_binom]{mod_binom()}} for specification
of Poisson and binomial models
\item \code{\link[=fit]{fit()}} to fit a model
\item \code{\link[=simulate]{simulate()}} to create simulated data
\item \code{\link[=set_prior]{set_prior()}} to specify non-default priors
}
}
