% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{mod_pois}
\alias{mod_pois}
\title{Specify a Poisson model}
\usage{
mod_pois(formula, data, exposure)
}
\arguments{
\item{formula}{An R \link[stats:formula]{formula},
specifying the outcome and predictors.}

\item{data}{A data frame containing the outcome,
predictors, and, optionally, exposure.}

\item{exposure}{Name of the exposure variable,
or a \code{1}.}
}
\value{
An object of class \code{bage_mod}.
}
\description{
Specify a model where the outcome is drawn from
a Poisson distribution.
}
\details{
\itemize{
\item \code{formula} is a standard R \link[stats:formula]{formula},
specifying the outcome and predictors, including interactions
between predictors.
\item \code{data} holds the outcome, the predictors, and,
optionally, exposure.
\item \code{exposure} is the name (bare or quoted) of the variable
in \code{data} used to measure exposure, or, if the
model does include exposure, a \code{1}.
}

If the model includes exposure, then the
the first level of the model is

\deqn{y \sim \text{Poisson}(\mu w)}

where \eqn{\mu} is the underlying rate, and
\eqn{w} is exposure. If the model does not
include exposure, then the first level is

\deqn{y \sim \text{Poisson}(\mu)}

TODO - Include error term once specification finalised.
}
\examples{
## model with exposure
mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
                data = injuries,
                exposure = popn)

## model without exposure
mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
                data = injuries,
                exposure = 1)
}
\seealso{
\itemize{
\item \code{\link[=mod_binom]{mod_binom()}} and \code{\link[=mod_norm]{mod_norm()}} for specification
of binomial and normal models
\item \code{\link[=fit]{fit()}} to fit a model
\item \code{\link[=simulate]{simulate()}} to create simulated data
\item \code{\link[=set_prior]{set_prior()}} to specify non-default priors
}
}
