% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{N}
\alias{N}
\title{Normal Prior}
\usage{
N(s = 1)
}
\arguments{
\item{s}{A positive, finite number.}
}
\value{
An object of class \code{bage_prior_norm}.
}
\description{
Prior in which units are drawn independently from a normal
distribution. The default prior for most terms.
}
\section{Mathematical details}{


If \eqn{\beta_j} is the \eqn{j}th element of a main effect
or interaction, then

\deqn{\beta_j \sim \text{N}(0, \tau^2).}

Standard deviation \eqn{\tau} is drawn from a half-normal
distribution,

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution defaults
to 1, but can be set to other values. Lower values
lead to more tightly concentrated
estimates for \verb{\\beta_j}, and higher values lead to less tightly
concentrated estimates.
}

\examples{
N()
N(s = 0.5)
}
\seealso{
\itemize{
\item \code{\link[=NFix]{NFix()}} Version of \code{N()} where the standard deviation
term is supplied rather than estimated from the data.
\item \link{priors} Overview of priors implemented in \code{bage}.
}
}
