% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{N}
\alias{N}
\title{Normal prior}
\usage{
N(s = 1)
}
\arguments{
\item{s}{A positive, finite number.}
}
\value{
An object of class \code{bage_prior_norm}.
}
\description{
Prior in which units are drawn independently from a normal
distribution. The default prior for most terms.
}
\details{
The normal distribution has mean \eqn{0} and standard
deviation \eqn{\sigma}.

\deqn{x \sim \text{N}(0, \sigma^2)}

Standard deviation \eqn{\sigma} is drawn from a half-normal
distribution,

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution defaults
to 1, but can be set to other values. Lower values
lead to more tightly concentrated
estimates for \code{x}, and higher values lead to less tightly
concentrated estimates.
}
\examples{
N()
N(s = 0.5)
}
\seealso{
\code{N()} is usually called within \code{\link[=set_prior]{set_prior()}}.
Other priors are \code{\link[=RW]{RW()}}, \code{\link[=RW2]{RW2()}}.
}
