% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priors}
\title{Priors for Intercept, Main Effects, Interactions}
\description{
The models for rates, probabilities, or means
created with functions \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}},
and \code{\link[=mod_norm]{mod_norm()}} always include an intercept,
and typically include main effects,
and interactions. Most models, for instance
include an age effect, and many include
and interaction between age and sex/gender,
or age and time.
}
\details{
The intercept, main effects, and interactions
all have prior models that capture the expected
behavior of the term. Current choices of prior
models are summarised here.

Priors with links to help have been implemented.
Priors without links are next in the queue for implementation.\tabular{llll}{
   \strong{Prior} \tab \strong{Description} \tab \strong{Uses} \tab \strong{Can be forecasted} \cr
   \code{\link[=N]{N()}} \tab Elements drawn from a common normal distribution with mean 0 and standard devation estimated from the data. \tab Main effects or interactions with no natural order or neighbourbood structure. \tab Yes \cr
   \code{\link[=NFix]{NFix()}} \tab As for \code{N()}, but standard deviation supplied (fixed) by user. \tab Like \code{N()}, but where there are too few elements to estimate the standard deviation, or where there is strong prior information about its value. Intercept and sex/gender effects. \tab Yes \cr
   \code{\link[=RW]{RW()}} \tab Random walk. Elements have a natural ordering, and neighbouring elements more strongly correlated than distant elements. \tab Modelling age, time, cohort, or other smoothly-varying dimensions. \tab Yes \cr
   \code{\link[=RW2]{RW2()}} \tab Second-order random walk. Similar to \code{RW()}, but with persistent trends. \tab As for \code{RW()}, but where greater smoothness expected. \tab Yes \cr
   \code{\link[=AR]{AR()}} \tab Auto-regressive prior of order \emph{k}. \tab Modelling variables that revert towards the mean, which often includes time. \tab Yes \cr
   \code{\link[=AR1]{AR1()}} \tab Auto-regressive prior of order 1. Special case of \code{AR()}, though has more options for controlling damping. \tab As for \code{AR()}. \tab Yes \cr
   \code{\link[=Known]{Known()}} \tab Values supplied by treated and treated as known. \tab Simulations and experiments. \tab No \cr
   \code{\link[=Lin]{Lin()}} \tab Linear trend, with independent normal. \tab Parsimonious model for time. \tab Yes \cr
   \code{\link[=LinAR]{LinAR()}} \tab Linear trend, with autoregressive errors. \tab Model for time, where departures from trend persist over time. \tab Yes \cr
   \code{\link[=LinAR1]{LinAR1()}} \tab Linear trend, with autoregressive errors of order 1. \tab Simpler version of \code{LinAR()}. \tab Yes \cr
   \code{\link[=Sp]{Sp()}} \tab P-Spline (penalised spline). \tab Modelling varies that smoothly, eg age. \tab No \cr
   \code{\link[=SVD]{SVD()}} \tab Age profile that looks like it was drawn at random from a database of rates, which are summarised using a singular value decomposition (SVD). \tab Main effects or interactions involving age, but not sex/gender. \tab No \cr
   \code{\link[=SVDS]{SVDS()}} \tab Age-sex or age-gender profile, similar to \code{SVD()} \tab Interactions involving age, sex/gender, and possibly other dimensions. \tab No \cr
}
}
