% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-methods.R
\name{replicate_data}
\alias{replicate_data}
\title{Create Replicate Data}
\usage{
replicate_data(x, condition_on = c("fitted", "expected"), n = 19)
}
\arguments{
\item{x}{A fitted model, typically created by
calling \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}},
and then \code{\link[=fit]{fit()}}.}

\item{n}{Number of replicate datasets to create.
Default is 19.}

\item{conditition_on}{Parameters to condition
on. Either \code{"fitted"} or \code{"expected"}. See
details.}
}
\value{
A tibble with the following structure:\tabular{ll}{
   \code{.replicate} \tab data \cr
   `"Original" \tab Original data supplied to \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_nbinom]{mod_nbinom()}}, \code{\link[=mod_norm]{mod_norm()}} \cr
   \code{"Replicate 1"} \tab Original data, except that actual outcome replaced by simulated values. \cr
   \code{"Replicate 2"} \tab Original data, except that actual outcome replaced by simulated values. \cr
   \dots \tab \dots \cr
   \code{"Replicate <n>"} \tab Original data, except that actual outcome replaced by simulated values. \cr
}
}
\description{
Use a fitted model to create replicate datasets,
typically as a way of checking a model.
}
\details{
Use \code{n} draws from the posterior distribution
for model parameters to generate \code{n} simulated datasets.
If the model is working well, these simulated
datasets should look similar to the actual dataset.
}
\section{}{
 The \code{condition_on} argument

With Poisson and binomial models that include
dispersion terms (which is the default), there are
two options for constructing replicate data.
\itemize{
\item When \code{condition_on} is \code{"fitted"},
the replicate data is created by (i) drawing values
from the posterior distribution for rates or probabilities
(the \eqn{\gamma_i} defined in \code{\link[=mod_pois]{mod_pois()}}
and \code{\link[=mod_binom]{mod_binom()}}), and (ii)  conditional on these
rates or probabilities, drawing values for the
outcome variable.
\item When \code{condition_on} is \code{"expected"},
the replicate data is created by (i) drawing
values from hyper-parameters governing
the rates or probabilities
(the \eqn{\mu_i} and \eqn{\xi} defined
in \code{\link[=mod_pois]{mod_pois()}} and \code{\link[=mod_binom]{mod_binom()}}),
then (ii) conditional on these hyper-parameters,
drawing values for the rates or probabilties,
and finally (iii) conditional on these
rates or probabilities, drawing values for the
outcome variable.
}

The \code{"expected"} option is a more severe test for
a model, it gives the original data less opportunity
to influence the replicate data.

As described in \code{\link[=mod_norm]{mod_norm()}}, normal models
have a slightly different structure from Poisson
and binomial models, and the distinction between
fitted and expected values does not apply.
}

\examples{
mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
                data = injuries,
                exposure = 1) |>
  fit()

rep_data <- mod |>
  replicate_data()

library(dplyr)
replicate_data |>
  group_by(.replicate) |>
  count(wt = injuries)
}
\seealso{
\itemize{
\item \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_nbinom]{mod_nbinom()}}, \code{\link[=mod_norm]{mod_norm()}} to create models
\item \code{\link[=fit]{fit()}} to fit models
}
}
