% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW}
\alias{RW}
\title{Random walk prior}
\usage{
RW(s = 1)
}
\arguments{
\item{s}{A positive, finite number. Default is 1.}
}
\value{
An object of class \code{bage_prior_rw}.
}
\description{
Prior in which units follow a (first order)
random waok. Increments between neighbouring
\code{x}s are assumed to be normally distibuted,
}
\details{
\deqn{x_1 \sim \text{N}(0, 1)}
\deqn{x_i - x_{i-1} \sim \text{N}(0, \sigma^2), i = 2, \cdots, n}

Standard deviation \eqn{\sigma} is drawn from a
half-normal distribution,

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{scale} lead to smoother series of \code{x}s, and
higher values lead to rougher series.
}
\examples{
RW()
RW(s = 0.5)
}
\seealso{
\itemize{
\item \code{\link[=RW2]{RW2()}} for a random walk with drift
(a second-order random walk).
\item \code{RW()} is usually called as part of
a call to \code{\link[=set_prior]{set_prior()}}
}
}
