% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW}
\alias{RW}
\alias{RW2}
\title{Random walk priors}
\usage{
RW(scale = 1)

RW2(scale = 1)
}
\arguments{
\item{scale}{A positive, finite number.}
}
\value{
An object of class \code{bage_prior_rw}
or \code{bage_prior_rw2}.
}
\description{
Priors in which units follow a one-dimensional
random walk or random walk with drift.
}
\details{
With \code{RW()}, increments between neighbouring
\code{x}s are normally distibuted,

\deqn{x_i - x_{i-1} \sim \text{N}(0, s^2)}

With \code{RW2()}, increments in increments
are normally distributed,

\deqn{(x_i - x_{i-1}) - (x_{i-1} - x_{i-2}) \sim \text{N}(0, s^2)}

In addition, there is a linear term TODO - EXPLAIN.

In both cases, standard deviation \code{s} is drawn from a
half-normal distribution,

\deqn{s \sim \text{N}^+(0, \text{scale}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution defaults
to 1, but can be set to other values. Lower values
for \code{scale} lead to smoother series of \code{x}s, and
higher values lead to rougher series.
}
\examples{
RW()
RW(scale = 0.5)
RW2()
}
\seealso{
\code{RW()} or \code{RW2()} are usually called within
\code{\link[=set_prior]{set_prior()}}. Other priors are \code{\link[=N]{N()}}.
}
