% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW2}
\alias{RW2}
\title{Random Walk with Drift Prior}
\usage{
RW2(s = 1)
}
\arguments{
\item{s}{A positive, finite number. Default is 1.}
}
\value{
An object of class \code{"bage_prior_rw2"}.
}
\description{
Prior in which units follow a random walk with drift,
also known as a second-order random walk.
}
\section{Mathematical details}{


If \eqn{\beta_j} is the \eqn{j}th element of a main effect, then

\deqn{\beta_j \sim \text{N}(0, 1), \quad j = 1,2}
\deqn{\beta_j \sim \text{N}(\beta_{j-1} + \delta_{j-1}, \tau^2), \quad j = 3, \cdots, J,}
where
\deqn{\delta_j = \beta_j - \beta_{j-1}.}

Standard deviation \eqn{\tau} is drawn from a
half-normal distribution,

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{s} lead to smoother series .
}

\examples{
RW2()
RW2(s = 0.2)
}
\seealso{
\itemize{
\item \code{\link[=RW]{RW()}} Ordinary random wark.
\item \code{\link[=ERW2]{ERW2()}} Exchangeable version of \code{RW2()},
used with interactions.
\item \link{priors} Overview of priors implemented in \strong{bage}
}
}
