% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW2}
\alias{RW2}
\title{Random Walk with Drift Prior}
\usage{
RW2(s = 1, sd = 1)
}
\arguments{
\item{s}{A positive, finite number. Default is 1.}

\item{sd}{A positive, finite number. Default is 1.}
}
\value{
An object of class \code{"bage_prior_rw2"}.
}
\description{
Prior in which units follow a random walk with drift,
also known as a second-order random walk.
Second-order differences are normally distributed.
}
\section{Mathematical details}{


If \eqn{\beta_j} is the \eqn{j}th element of a main effect, then

\deqn{(\beta_j - \beta_{j-1}) - (\beta_{j-1} - x_{j-2}) \sim \text{N}(0, \tau^2)},

Standard deviation \eqn{\tau} is drawn from a
half-normal distribution,

\deqn{\tau \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{s} lead to smoother series .

Parameter \code{sd} governs the expected size of
increments between neighbouring units.
It defaults to 1, but can be set to other values.
}

\examples{
RW2()
RW2(s = 0.2)
}
\seealso{
\itemize{
\item \code{\link[=RW]{RW()}} for a first-order random walk.
\item \code{RW2()} is usually called as part of
a call to \code{\link[=set_prior]{set_prior()}}
}
}
