% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW2}
\alias{RW2}
\title{Random walk with drift prior}
\usage{
RW2(s = 1, sd = 1)
}
\arguments{
\item{s}{A positive, finite number. Default is 1.}

\item{sd}{A positive, finite number. Default is 1.}
}
\value{
An object of class \code{bage_prior_rw}
or \code{bage_prior_rw2}.
}
\description{
Prior in which units follow a random walk with drift.
Second-order differences are normally distributed,
}
\details{
\deqn{x_1 \sim \text{N}(0, 1)}
\deqn{x_2 \sim \text{N}(x_1, sd^2)}
\deqn{(x_i - x_{i-1}) - (x_{i-1} - x_{i-2}) \sim \text{N}(0, \sigma^2),
      i = 3, \cdots, n}

Standard deviation \eqn{\sigma} is drawn from a
half-normal distribution,

\deqn{\sigma \sim \text{N}^+(0, \text{s}^2)}

(A half-normal distribution has the same shape as a normal
distribution, but is defined only for non-negative
values.)

The scale for the half-normal distribution, \code{s}, defaults
to 1, but can be set to other values. Lower values
for \code{scale} lead to smoother series of \code{x}s, and
higher values lead to rougher series.

Parameter \code{sd} governs the expected size of
increments between neighbouring units.
It defaults to 1, but can be set to other values.
}
\examples{
RW2()
RW2(s = 0.2)
}
\seealso{
\itemize{
\item \code{\link[=RW]{RW()}} for a first-order random walk.
\item \code{RW2()} is usually called as part of
a call to \code{\link[=set_prior]{set_prior()}}
}
}
