% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RW2_AR1}
\alias{RW2_AR1}
\title{Second-Order Random Walk Prior with
First Order Autoregressive Errors}
\usage{
RW2_AR1(
  s_rw = 1,
  sd = 1,
  sd_slope = 1,
  s_ar = 1,
  shape1 = 5,
  shape2 = 5,
  min = 0.8,
  max = 0.98,
  along = NULL,
  con = c("none", "by")
)
}
\arguments{
\item{s_rw}{Scale for the innovations in the
RW2 process. Default is \code{1}.}

\item{sd}{Standard deviation for initial term
in RW2 process. Default is \code{1}. Can be \code{0}.}

\item{sd_slope}{Standard deviation in the prior for
the initial slope of RW2 process. Larger values imply
steeper slopes. Default is 1.}

\item{s_ar}{Scale for the innovations in the
AR1 process. Default is \code{1}.}

\item{shape1, shape2}{Parameters for beta-distribution prior
for coefficients. Defaults are \code{5} and \code{5}.}

\item{min, max}{Minimum and maximum values
for autocorrelation coefficient in AR1 process.
Defaults are \code{0.8} and \code{0.98}.}

\item{along}{Name of the variable to be used
as the 'along' variable. Only used with
interactions.}

\item{con}{Constraints on parameters.
Current choices are \code{"none"} and \code{"by"}.
Default is \code{"none"}. See below for details.}
}
\value{
An object of class \code{"bage_prior_rw2randomar"}
or \code{"bage_prior_rw2zeroar"}.
}
\description{
Use one or more second-order random walks,
combined with an AR1
error term, to model a main effect
or an interaction. Typically used with time.
}
\details{
If \code{RW2_AR1()} is used with an interaction,
separate random walks are constructed along
the 'along' variable, within each combination
of the 'by' variables.

Parameters controlling the RW2 process:
\itemize{
\item \code{s_rw}
\item \code{sd}
\item \code{sd_slope}
}

Parameters controlling the AR1 process:
\itemize{
\item \code{s_ar}
\item \code{shape1}
\item \code{shape2}
\item \code{min}
\item \code{max}
}
}
\section{Mathematical details}{


When \code{RW2_AR1()} is used with a main effect,

\deqn{\beta_j = \alpha_j + \epsilon_j}
\deqn{\alpha_1 \sim \text{N}(0, \mathtt{sd}^2)}
\deqn{\alpha_2 \sim \text{N}(\alpha_1, \mathtt{sd\_slope}^2)}
\deqn{\alpha_j \sim \text{N}(2\alpha_{j-1} - \alpha_{j-2}, \tau^2), \quad j = 3, \cdots, J}
\deqn{\epsilon_j = \phi \epsilon_{j-1} + \varepsilon_j}
\deqn{\varepsilon_j \sim \text{N}(0, \omega^2),}

and when it is used with an interaction,

\deqn{\beta_{u,v} = \alpha_{u,v} + \epsilon_{u,v}}
\deqn{\alpha_{u,1} \sim \text{N}(0, \mathtt{sd}^2)}
\deqn{\alpha_{u,2} \sim \text{N}(\alpha_{u,1}, \mathtt{sd\_slope}^2)}
\deqn{\alpha_{u,v} \sim \text{N}(2\alpha_{u,v-1} - \alpha_{u,v-2}, \tau^2), \quad v = 3, \cdots, V}
\deqn{\epsilon_{u,v} = \phi \epsilon_{u,v-1} + \varepsilon_{u,v}}
\deqn{\varepsilon_{u,v} \sim \text{N}(0, \omega^2),}

where
\itemize{
\item \eqn{\pmb{\beta}} is the main effect or interaction;
\item \eqn{j} denotes position within the main effect;
\item \eqn{u} denotes position within the 'by' variable(s) of the interaction; and
\item \eqn{v} denotes position within the 'along' variable of the interaction.
}

The \eqn{\tau} parameter in the random walk has prior
\deqn{\tau \sim \text{N}^+(0, \mathtt{s\_rw}^2)}

Internally, \code{RW2_AR()} derives a value for \eqn{\omega} that
gives \eqn{\epsilon_j} or \eqn{\epsilon_{u,v}} a marginal
variance of \eqn{\nu^2}. Parameter \eqn{\nu}
has a half-normal prior
\deqn{\nu \sim \text{N}^+(0, \mathtt{s\_ar}^2).}

Coefficient \eqn{\phi} is constrained
to lie between \code{min} and \code{max}.
Its prior distribution is

\deqn{\phi = (\mathtt{max} - \mathtt{min}) \phi' - \mathtt{min}}

where

\deqn{\phi' \sim \text{Beta}(\mathtt{shape1}, \mathtt{shape2}).}
}

\section{Constraints}{


With some combinations of terms and priors, the values of
the intercept, main effects, and interactions
are only weakly identified.
This weak identifiability is
typically harmless. However, in some applications, such as
when trying to obtain interpretable values
for main effects and interactions, it can be helpful to increase
identifiability through the use of constraints, specified through the
\code{con} argument.

Current options for \code{con} are:
\itemize{
\item \code{"none"} No constraints. The default.
\item \code{"by"} Only used in interaction terms that include 'along' and
'by' dimensions. Within each value of the 'along'
dimension, terms across each 'by' dimension are constrained
to sum to 0.
}
}

\examples{
RW2_AR1()
RW2_AR1(sd_slope = 2, s_ar = 0.5)
}
\seealso{
\itemize{
\item \code{\link[=RW2_AR]{RW2_AR()}} Generalization of \code{RW2_AR1()}
\item \code{\link[=Lin_AR1]{Lin_AR1()}} Sepcial case of \code{RW2_AR1()}
\item \code{\link[=RW2]{RW2()}} Second-order random walk
\item \code{\link[=AR1]{AR1()}} AR1 process
\item \link{priors} Overview of priors implemented in \pkg{bage}
\item \code{\link[=set_prior]{set_prior()}} Specify prior for intercept,
main effect, or interaction
\item \href{https://bayesiandemography.github.io/bage/articles/vig02_math.html}{Mathematical Details}
vignette
}
}
