% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RWSeas}
\alias{RWSeas}
\title{Random Walk with Seasonal Effect}
\usage{
RWSeas(n, s = 1, seas = 1, along = NULL)
}
\arguments{
\item{n}{Number of seasons}

\item{s}{Scale for prior for innovations in
the random walk. Default is \code{1}.}

\item{seas}{Scale for prior for innovations
in the seasonal effect. Default is \code{1}.
Can be \code{0}.}
}
\value{
Object of class \code{"bage_prior_rwseasvary"}
or \code{"bage_prior_rwseasfix"}.
}
\description{
Random Walk with Seasonal Effect
}
\examples{
RWSeas(n = 4)           ## seasonal effects evolve
RWSeas(n = 4, a = 0) ## seasonal effects fixed
}
