% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{RWSeas}
\alias{RWSeas}
\title{Random Walk with Seasonal Effect}
\usage{
RWSeas(n, s = 1, s_seas = 1, along = NULL)
}
\arguments{
\item{n}{Number of seasons}

\item{s}{Scale for prior for innovations in
the random walk. Default is \code{1}.}

\item{s_seas}{Scale for prior for innovations
in the seasonal effect. Default is \code{1}.
Can be \code{0}.}

\item{along}{Name of the variable to be used
as the "along" variable. Only used with
interactions.}
}
\value{
Object of class \code{"bage_prior_rwseasvary"}
or \code{"bage_prior_rwseasfix"}.
}
\description{
TODO - NEED TO FINISH THIS
}
\examples{
RWSeas(n = 4)             ## seasonal effects evolve
RWSeas(n = 4, s_seas = 0) ## seasonal effects fixed
}
