% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_scaled_svd-constructors.R
\name{scaled_svd}
\alias{scaled_svd}
\title{Create object to hold data from scaled SVDs}
\usage{
scaled_svd(data)
}
\arguments{
\item{data}{A data frame. See Details for description.}
}
\value{
An object of class \code{bage_scaled_svd}.
}
\description{
Create an object of class \code{bage_scaled_svd} to
hold scaled SVDs (singular value decompositionsk.)
}
\details{
\code{data} has the following columns:
\itemize{
\item \code{type} Type of decomposition. Choices are "total",
"joint", and "indep". See details.
\item \code{labels_age} Age labels for individual rows of
matrices within \code{matrix} and individual elements
of vectors within \code{offset}.
\item \code{labels_sexgender} Sex/gender labels for individual rows of
matrices within \code{matrix} and individual elements
of vectors within \code{offset}, or \code{NULL}.
\code{NULL} when \code{sexgender} is \code{"total"}, since in
this case results average across sexes/genders.
\item \code{matrix} List column of sparse matrices created by
\code{\link[=scaled_svd_comp]{scaled_svd_comp()}}. Must have rownames.
Must not have NAs. When \code{type} is \code{"total"} or \code{"joint"},
each matrix has 10 columns. When \code{"type"} is \code{"indep"},
each matrix has 20 columns.
\item \code{offset} List column of vectors created by
\code{\link[=scaled_svd_comp]{scaled_svd_comp()}}. Must have names, which
are identical to the rownames of the corresponding
element of \code{matrix}.
}
}
