% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_svd_comp.R
\name{scaled_svd_comp}
\alias{scaled_svd_comp}
\title{Create components needed by a scaled_svd object}
\usage{
scaled_svd_comp(x, n = 5, transform = c("log", "logit", "none"))
}
\arguments{
\item{x}{A matrix with value such as rates,
probabilities, or means.}

\item{n}{Number of components of the SVD to use.
Defaults to 5.}

\item{transform}{\code{"log"}, \code{"logit"}, or \code{"none"}.
Defaults to \code{"log"}.}
}
\value{
A named list with two elements:
\itemize{
\item \code{transform}, a numeric matrix
\item \code{translate}, a numeric vector
}
}
\description{
Use the \link[base:svd]{Singular Value Decomposition}
(SVD) to construct a parsimonious representation
of a set of rates, probabilities, means, or
other values.
}
\details{
`scaled_svd_comp() typically proceeds as follows:
\itemize{
\item transform values in matrix \code{x} (eg take logs)
\item carry out a SVD on the transformed version of \code{x}
\item centre and scale the results from the SVD to produce
matrix \code{transform} and vector \code{translate}.
}

If
\itemize{
\item \eqn{X} is a the matrix of transformed values
of \code{x}
\item \eqn{A} is the matrix called \code{transform},
\item \eqn{b} is the vector called \code{translate}, and
\item \eqn{\beta} is a vector of standard normal variates,
}

and

\deqn{v = A \beta + b}

then \eqn{v} should look like a randomly-selected
column from \eqn{X}.

Matrix \code{x} typically has age along the rows,
and some combination of classification variables,
such as country and time, along the columns.
One exception is when the SVD is used to capture
the relationship between female and male rates,
in which case rows are formed by interacting
sex and age. See below for an example.

The number of components used by the SVD
(and hence the number of columns of the
\code{tranform} matrix) is governed by
argument \code{n}. The default is 5.

When \code{scale} is \code{"log"} or \code{"logit"},
\code{scaled_svd_comp()} converts any \code{0}s in
\code{x} to values just above \code{0} before
applying the log or logit function.
When \code{scale} is \code{"logit"},
\verb{scaled_svd_comp() also converts any }1\verb{s to values just below }1`.
}
\examples{
x <- matrix(rgamma(n = 50, shape = 1), nrow = 5, ncol = 10)
x
scaled_svd_comp(x, n = 3)
scaled_svd_comp(x, n = 3, transform = "none")
## TODO - AN EXAMPLE COMBINING FEMALE AND MALE
}
\seealso{
\itemize{
\item \code{\link[poputils:to_matrix]{poputils::to_matrix()}} can create matrices to
supply to \code{scaled_svd_comp}
\item \if{html}{\out{<name of vignette>}} has mathematical details
\item example of use of \code{scaled_svd_comp}:
href{https://github.com/bayesiandemography/svd_mortality}{svd_mortality}
}
}
