% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_datamod_exposure}
\alias{set_datamod_exposure}
\title{Specify Exposure Data Model}
\usage{
set_datamod_exposure(mod, disp)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod_pois"},
created with \code{\link[=mod_pois]{mod_pois()}}.}

\item{disp}{Dispersion of measurement errors.
A single number, or a data frame
with a variable called \code{"disp"},
and one or more 'by' variables.}
}
\value{
A modified version of \code{mod}.
}
\description{
Specify a data model for the exposure
term in a Poisson model. The data model assumes
that, within each cell, observed exposure is drawn
from an Inverse-Gamma distribution, with

E[expected exposure] = true exposure

and

var[expected exposure] = dispersion \eqn{\times}
(true exposure)^2
}
\details{
The model assumes that the exposure variable
is unbiased. If there is in fact evidence
of biases, then this evidence should be
used to create a de-biased version of the
variable (eg one where estimated biases
have been subtracted), to supply to
\code{\link[=mod_pois]{mod_pois()}}.

With the Inverse-Gamma distribution,
the coefficient of variation for
observed exposure (i.e. the standard
deviation divided by the mean) is \eqn{\sqrt{\text{dispersion}}}).
This implies that the average size of errors
scales with the mean. It also implies that
values for dispersion will typically be very small.

\code{set_datamod_exposure()} can only be used
with model where dispersion in Poisson rates
has been set to zero, using function \code{\link[=set_disp]{set_disp()}}.
In models that use \code{set_datamod_exposure()},
"dispersion" refers to variation in
observed exposures, not variation in Poisson rates.
}
\section{The \code{disp} argument}{


\code{disp} can be a single number, in which
case the same dispersion is used for all cells.
\code{sd} can also be a data frame with a
with a variable called \code{"disp"} and
one or more columns with 'by' variables.
For instance, a  \code{disp} of

\if{html}{\out{<div class="sourceCode">}}\preformatted{data.frame(sex = c("Female", "Male"),
           disp = c(0.15, 0.12))
}\if{html}{\out{</div>}}

implies that dispersion is 0.15 for
females and 0.12 for males.
}

\section{Mathematical details}{


The model for observed exposure is

\deqn{w_i^{\text{obs}} \sim \text{InvGamma}(2 + d_{g \lbrack i \rbrack }^{-1}, (1 + d_{g \lbrack i\rbrack }^{-1}) w_i^{\text{true}})}

where
\itemize{
\item \eqn{w_i^{\text{obs}}} is observed exposure for cell \eqn{i}
(the \code{exposure} argument to \code{\link[=mod_pois]{mod_pois()}});
\item \eqn{w_i^{\text{true}}} is true exposure for cell \eqn{i}; and
\item \eqn{d_{g\lbrack i\rbrack }} is the value for dispersion
that is applied to cell \eqn{i}.
}

The number of distinct values of \eqn{d_g}
is typically much less than the number
of cells. For instance, there might
only be two distinct values for \eqn{d_g},
with everyone aged 0-49 receiving the
first value, and everyone age 50+ receiving
the second. See below for an example.
}

\examples{
## specify model, including setting
## dispersion for rates to zero
mod <- mod_pois(injuries ~ age * sex + year,
                data = nzl_injuries,
                exposure = popn) |>
  set_disp(mean = 0) |>
  set_datamod_exposure(disp = 0.00002)

## fit the model
mod <- mod |>
  fit()
mod

## examine results - note the new variable
## '.popn' with estimates of the true
## population
aug <- mod |>
  augment()

## allow different dispersions for each sex
disp_sex <- data.frame(sex = c("Female", "Male"),
                       disp = c(0.00001, 0.00003))
mod <- mod |>
  set_datamod_exposure(disp = disp_sex)
mod

## our outcome variable is confidentialized,
## so we recognize that in the model too
mod <- mod |>
  set_confidential_rr3()
mod

## now a model where everyone aged 0-49
## receives one value for dispersion,
## and everyone aged 50+ receives another
library(poputils) ## for 'age_upper()'
nzl_injuries_age <- nzl_injuries |>
  mutate(age_group = if_else(age_upper(age) < 50,
                             "0-49",
                             "50+"))
disp_age <- data.frame(age_group = c("0-49", "50+"),
                       disp = c(0.000025, 0.00001))
mod <- mod_pois(injuries ~ age * sex + year,
                data = nzl_injuries_age, ## new
                exposure = popn) |>
  set_disp(mean = 0) |>
  set_datamod_exposure(disp = disp_age)
}
\seealso{
\itemize{
\item \code{\link[=mod_pois]{mod_pois()}} Specify a Poisson model
\item \code{\link[=set_disp]{set_disp()}} Specify dispersion of rates, probabilities, or means
\item \code{\link[=augment]{augment()}} Original data plus estimated values,
including estimates of true value for exposure
\item \link{datamods} Data models implemented in \code{bage}
\item \href{https://bayesiandemography.github.io/bage/articles/vig2_math.html}{Mathematical Details}
vignette
}
}
