% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_datamod_exposure}
\alias{set_datamod_exposure}
\title{Specify Exposure Data Model}
\usage{
set_datamod_exposure(mod, disp)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod_pois"},
created with \code{\link[=mod_pois]{mod_pois()}}.}

\item{disp}{Dispersion. A data frame
with a variable called \code{"mean"},
and, oprtionally, one or more 'by' variables.}
}
\value{
A modified version of \code{mod}.
}
\description{
Specify a data model for the exposure
term in a Poisson model. The data model assumes
that, within each cell, observed exposure is drawn
from an Inverse-Gamma distribution, with

E[expected exposure] = true exposure

and

var[expected exposure] = dispersion \eqn{\times}
(true exposure)^2
}
\details{
With the Inverse-Gamma distribution,
the coefficient of variation for
observed exposure (i.e. the standard
deviation divided by the mean) is \eqn{\sqrt{\text{dispersion}}}).
This implies that the average size of errors
scales with the mean.

\code{set_datamod_exposure()} can only be used
with model where dispersion in Poisson rates
has been set to zero, using function \code{\link[=set_disp]{set_disp()}}.
In models that use \code{set_datamod_exposure()},
"dispersion" refers to variation in
observed exposures, not variation in Poisson rates.
}
\section{The \code{disp} argument}{


\code{disp} can be a single number, in which
case the same dispersion is used for all cells.
It can also be a data frame with a
with a variable called \code{"disp"} and
one or more columns with 'by' variables.
For instance, a  \code{disp} of

\if{html}{\out{<div class="sourceCode">}}\preformatted{data.frame(sex = c("Female", "Male"),
           disp = c(0.15, 0.12))
}\if{html}{\out{</div>}}

implies that dispersion is 0.15 for
females and 0.12 for males.
}

\section{Mathematical details}{


The model for observed exposure is

\deqn{w_i^{\text{obs}} \sim \text{InvGamma}(2 + d_{g \lbrack i \rbrack }^{-1}, (1 + d_{g \lbrack i\rbrack }^{-1}) r_{h \lbrack i \rbrack } w_i^{\text{true}})}

where
\itemize{
\item \eqn{w_i^{\text{obs}}} is observed exposure for cell \eqn{i}
(the \code{exposure} argument to \code{\link[=mod_pois]{mod_pois()}});
\item \eqn{w_i^{\text{true}}} is true exposure for cell \eqn{i}; and
\item \eqn{d_{g\lbrack i\rbrack }} is dispersion for cell \eqn{i}.
}

The same value \eqn{d_g} for dispersion can
be used for multiple cells \eqn{i}.
}

\examples{
## specify model, including setting
## dispersion for rates to zero
mod <- mod_pois(injuries ~ age * sex + year,
                data = nzl_injuries,
                exposure = popn) |>
  set_disp(mean = 0) |>
  set_datamod_exposure(disp = 0.00002)

## fit the model
mod <- mod |>
  fit()
mod

## examine results - note the new variable '.popn'
## with imputed true values for population
aug <- mod |>
  augment()

## allow different dispersions for each sex
disp_sex <- data.frame(sex = c("Female", "Male"),
                       disp = c(0.00001, 0.00003))
mod <- mod |>
  set_datamod_exposure(disp = disp_sex)
mod

## our outcome variable is in fact confidentialized
## confidentialized, so we allow for that too
mod <- mod |>
  set_confidential_rr3()
mod
}
\seealso{
\itemize{
\item \code{\link[=mod_pois]{mod_pois()}} Specify a Poisson model
\item \code{\link[=set_disp]{set_disp()}} Specify dispersion of rates, probabilities, or means
\item \code{\link[=components]{components()}} Values for parameters estimated in the
model, included the dispersion parameter in the data
model for exposure.
\item \code{\link[=set_datamod_miscount]{set_datamod_miscount()}} Specify a data model for a
Poisson model where the outcome variable
is subject to undercount and overcount
\item \code{\link[=set_datamod_overcount]{set_datamod_overcount()}} Specify a data model for a
Poisson model where the outcome variable is
subject to overcount, but not undercount
\item \code{\link[=set_datamod_undercount]{set_datamod_undercount()}} Specify a data model for a
Poisson or binomial model where the
outcome variable is subject to undercount,
but not overcount
\item \code{\link[=set_datamod_noise]{set_datamod_noise()}} Specify a data model for a
normal model where the  outcome variable
has normally-distributed measurement errors
}
}
