% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_datamod_miscount}
\alias{set_datamod_miscount}
\title{Specify Miscount Data Model}
\usage{
set_datamod_miscount(mod, prob, rate)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod"},
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}

\item{prob}{A data frame with 'by' variables,
a variable called \code{"mean"}, and a variable
called \code{"disp"}}

\item{rate}{A data frame with 'by' variables,
a variable called \code{"mean"}, and a variable
called \code{"disp"}}
}
\value{
A revised version of \code{mod}.
}
\description{
Specify a data model for the outcome variable
in a Poisson model, where the outcome variable
is subject simultaneously to undercount and
to overcount.
}
\details{
The miscount data model assumes that
the reported value for an outcome includes
(1) people or events who belong in the
target population, and (2) people
or events who do not belong in the targe population.
Coverage of (1) is incomplete, so that
the count (1)
is typically less (and never higher) than
the true population. people or events who belong in
the target population is incomplete, so
the
the reported value for an outcome as the sum
of two quantities:
\enumerate{
\item An incomplete count of people or events who
belong in the target population.
\item A count
belong in the target population.
\item CouPeople or events who belong in the target
population, but who are
}

The miscount data model

reported outcome = i

The data model assumes that observed counts
of people or events are the sum
of two quantities:
\enumerate{
\item People are events that are
are the sum of two quantities:
\item made up of two things:
}

1

observed outcome = true (but incomplet
sum of two terms:
\enumerate{
\item True
}

observed outcome =

We have observed outcome variable takes the form
}
