% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_datamod_outcome_rr3}
\alias{set_datamod_outcome_rr3}
\title{Specify RR3 Data Model}
\usage{
set_datamod_outcome_rr3(mod)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod"},
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}
}
\value{
A modified version of \code{mod}.
}
\description{
Specify a data model where the outcome variable
has been randomly rounded to base 3.
}
\details{
\code{set_datamod_outcome_rr3()} can only be used with
Poisson and binomial models (created with
\code{\link[=mod_pois]{mod_pois()}} and \code{\link[=mod_binom]{mod_binom()}}.)

Random rounding to base 3 (RR3) is a confidentialization
technique that is sometimes applied by statistical
agencies. RR3 is applied to integer data. The
procedure for rounding value \eqn{n} is as follows:
\itemize{
\item If \eqn{n} is divisible by 3, leave it unchanged
\item If dividing \eqn{n} by 3 leaves a remainder of 1, then
round down (subtract 1) with probability 2/3,
and round up (add 2) with probability 1/3.
\item If dividing \eqn{n} by 3 leaves a remainder of 1,
then round down (subtract 2)
with probability 1/3, and round up (add 1)
with probability 2/3.
}

If \code{set_datamod_outcome_rr3()} is applied to
a fitted model, it 'unfits'
the model, deleting existing estimates.
}
\examples{
## 'injuries' variable in 'injuries' dataset
## has been randomly rounded to base 3
mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
                data = injuries,
                exposure = popn) |>
  set_datamod_outcome_rr3() |>
  fit()
}
\seealso{
\itemize{
\item \link{datamods} Overview of data models implemented in \strong{bage}
\item \code{\link[=mod_pois]{mod_pois()}}, \code{\link[=mod_binom]{mod_binom()}}, \code{\link[=mod_norm]{mod_norm()}} Specify a
model for rates, probabilities, or means
}
}
