% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_datamod_overcount}
\alias{set_datamod_overcount}
\title{Specify Overcount Data Model}
\usage{
set_datamod_overcount(mod, rate)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod"},
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}

\item{rate}{A data frame with 'by' variables,
a variable called \code{"mean"}, and a variable
called \code{"disp"}}
}
\value{
A modified version of \code{mod}.
}
\description{
Specify Overcount Data Model
}
