% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_datamod_undercount}
\alias{set_datamod_undercount}
\title{Specify Undercount Data Model}
\usage{
set_datamod_undercount(mod, prob)
}
\arguments{
\item{mod}{An object of class \code{"bage_mod"},
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}

\item{prob}{A data frame with 'by' variables,
a variable called \code{"mean"}, and a variable
called \code{"disp"}}
}
\value{
A revised version of \code{mod}.
}
\description{
Specify Undercount Data Model
}
