% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-functions.R
\name{set_season}
\alias{set_season}
\title{Add a seasonal effect}
\usage{
set_season(mod, n, s = 1)
}
\arguments{
\item{mod}{A \code{bage_mod} object, typically
created with \code{\link[=mod_pois]{mod_pois()}},
\code{\link[=mod_binom]{mod_binom()}}, or \code{\link[=mod_norm]{mod_norm()}}.}

\item{n}{Number of seasons.}

\item{s}{Scale of half-normal prior for
standard deviation (\eqn{\sigma}).
Defaults to 1.}
}
\value{
A modified \code{bage_mod} object.
}
\description{
Add a seasonal effect to a model.
}
\details{
TODO - description
}
\examples{
mod <- mod_pois(deaths ~ month,
                data = us_acc_deaths,
                exposure = 1)
mod
mod |> set_season(n = 12)
}
