% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{Spline}
\alias{Spline}
\title{P-spline prior}
\usage{
Spline(n = NULL, scale = 1)
}
\arguments{
\item{n}{Number of spline vectors.}

\item{scale}{Scale for error term.
The default is 1.}
}
\value{
An object of class \code{"bage_prior_spline"}.
}
\description{
Specify a P-spline (penalised spline) prior.
A P-spline is flexible, but
favours profiles that are relatively smooth.
}
\section{Mathematical details}{


The model for the effect, on the log scale, is

\deqn{\beta = X \gamma}

where
\itemize{
\item \eqn{\beta} is a main effect,
\item \eqn{X} is a matrix holding the basis functions
for the spline, with \code{n_spline} columns, and
\item \eqn{\alpha} is a vector of coefficients,
with \code{n_spline} elements.
}

The elements of \eqn{\gamma} are assumed to follow
a second order random walk,

\deqn{(\gamma_j - \gamma_{j-1}) - (\gamma_{j-1} - \gamma_{j-2}) \sim \text{N}(0, \tau^2)}

Parameter \eqn{\tau} has prior

\deqn{\tau \sim \text{N}^+(0, A^2)}

where \eqn{\text{N}^+} denotes a half-normal distribution,
and a value for \eqn{A} is supplied by the user.
}

\examples{
Spline()
Spline(df = 10)
}
\seealso{
\itemize{
\item \code{\link[=RW2]{RW2()}} etc
}
}
