% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{SVD}
\alias{SVD}
\title{SVD prior}
\usage{
SVD(scaled_svd, n = 5, indep = TRUE)
}
\arguments{
\item{scaled_svd}{An object created by \code{\link[=scaled_svd]{scaled_svd()}}
holding scaled values from a SVD of
age-specific rates, probabilities, or means.}

\item{n}{Number of components from scaled SVD
to use in modelling. Must be between 1 and 10.
Default is 10.}

\item{indep}{Whether, in an interaction,
age profiles for different
sexes/genders are modelled independently.
See description in Details. Default is \code{TRUE}.}
}
\value{
An object of class \code{"bage_prior_svd"}.
}
\description{
Specify a scaled SVD (singular value decomposition)
prior for an age variable, or a combination
of age and sex/gender variables.
}
