% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{SVD}
\alias{SVD}
\title{SVD Prior for Age}
\usage{
SVD(ssvd, n_comp = NULL)
}
\arguments{
\item{ssvd}{Object of class \code{"bage_ssvd"}
holding a scaled SVD. See below for scaled SVDs
of databases currently available in \strong{bage}.}

\item{n_comp}{Number of components from scaled SVD
to use in modelling. The default is half
the number of components of \code{ssvd}.}
}
\value{
An object of class \code{"bage_prior_svd"}.
}
\description{
A Singular Value Decomposition (SVD) prior
for a main effect or interaction involving age.
}
\details{
A \code{SVD()} prior assumes that the age profile for the quantity
being modelled looks like it was drawn at random
from an external demographic database. For instance,
the \code{SVD()} prior obtained via

\if{html}{\out{<div class="sourceCode">}}\preformatted{SVD(HMD)
}\if{html}{\out{</div>}}

favours age profiles that look like
they were obtained from the
\href{https://www.mortality.org}{Human Mortality Database}.

If \code{SVD()} is used with an interaction,
then separate profiles are constructed along
within each combination of the non-age variables.

Age-sex or age-gender profiles should be
modelled using \code{\link[=SVDS]{SVDS()}} rather than \code{SVD()}.
}
\section{Mathematical details}{


When \code{SVD()} is used with a main effect,

\deqn{\pmb{\beta} = \pmb{F} \pmb{\alpha} + \pmb{g},}

and when it is used with an interaction,

\deqn{\pmb{\beta}_u = \pmb{F} \pmb{\alpha}_u + \pmb{g},}

where
\itemize{
\item \eqn{\pmb{\beta}} is a main effect or interaction involving age;
\item \eqn{\pmb{\beta}_u} is a subvector of \eqn{\pmb{\beta}} holding
values for the  \eqn{u}th combination of the non-age variables;
\item \eqn{J} is the number of elements of \eqn{\pmb{\beta}};
\item \eqn{V} is the number of elements of \eqn{\pmb{\beta}_u};
\item \eqn{n} is \code{n_comp};
\item \eqn{\pmb{F}} is a known matrix with dimension \eqn{J \times n}
if \eqn{\pmb{\beta}} is a main effect, and dimension \eqn{V \times n}
if \eqn{\pmb{\beta}} is an interaction; and
\item \eqn{\pmb{g}} is a known vector with \eqn{J} elements if
if \eqn{\pmb{\beta}} is a main effect and \eqn{V}
elements if \eqn{\pmb{\beta}} is an interaction.
}

\eqn{\pmb{F}} and \eqn{\pmb{g}} are constructed from
a large database of age-specific demographic estimates
by performing an SVD and standardizing.

The elements of \eqn{\pmb{\alpha}} have prior
\deqn{\alpha_k \sim \text{N}(0, 1), \quad k = 1, \cdots, K.}
}

\section{Scaled SVDs of demographic databases in bage}{

\itemize{
\item \code{\link{HMD}} Mortality rates from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item \code{\link{LFP}} Labour forcce participation
rates from the \href{https://data-explorer.oecd.org}{OECD}.
}
}

\examples{
SVD(HMD)
SVD(HMD, n_comp = 3)
}
\references{
\itemize{
\item For details of the construction of
scaled SVDS see the vignette
\href{https://bayesiandemography.github.io/bage/articles/vig4_svd.html}{here}.
}
}
\seealso{
\itemize{
\item \code{\link[=SVDS]{SVDS()}} SVD prior for age-sex or age-gender profile
\item \code{\link[=SVDS_RW]{SVDS_RW()}} SVD prior for age with time-varying coeffients
\item \code{\link[=RW]{RW()}} Smoothing via random walk
\item \code{\link[=RW2]{RW2()}} Smoothing via random walk with drift
\item \code{\link[=Sp]{Sp()}} Smoothing via splines
\item \link{priors} Overview of priors implemented in \strong{bage}
\item \code{\link[=set_prior]{set_prior()}} Specify prior for intercept,
main effect, or interaction
}
}
