% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{SVD}
\alias{SVD}
\title{SVD Prior for Age Main Effect}
\usage{
SVD(ssvd, n = NULL)
}
\arguments{
\item{ssvd}{Object of class \code{"bage_ssvd"}
holding results from a scaled SVD.
See below for current options.}

\item{n}{Number of components from scaled SVD
to use in modelling. The default is half
the number of components of \code{ssvd}.}
}
\value{
An object of class \code{"bage_prior_svd"}.
}
\description{
Specify a Singular Value Decomposition (SVD) prior
for an age main effect.
}
\details{
An SVD prior assumes that the age profile for the quantity
being modelled looks like an age profile drawn at random
from an external demographic database. For instance,

\if{html}{\out{<div class="sourceCode">}}\preformatted{set_prior(mod, age ~ SVD(HMD))
}\if{html}{\out{</div>}}

specifies that the age profile looks like it was drawn from
\href{https://www.mortality.org}{Human Mortality Database}.

A \link[=svd]{Singular Value Decomposition} is closely related
to a Principal Components analysis, and is a way of
extracting patterns from large volumes of data.
}
\section{Mathematical details}{


With an SVD prior,

\deqn{\pmb{\beta} = \pmb{F} \pmb{\alpha} + \pmb{g}}

where
\itemize{
\item \eqn{\pmb{\beta}} is an age effect with \eqn{J} elements;
\item \eqn{\pmb{X}} is a known \eqn{J \times K} matrix, where \eqn{J}
is the number of age groups and \eqn{K} is the number of
vectors from the SVD to use (with \eqn{K} typically much
less than than \eqn{J});
\item \eqn{\pmb{\alpha}} is a vector with \eqn{K} elements; and
\item \eqn{\pmb{g}} is a known vector with \eqn{J} elements.
}

The elements of \eqn{\pmb{\alpha}} have prior

\deqn{\alpha_k \sim \text{N}(0, 1), \quad k = 1, \cdots, K.}

\eqn{\pmb{F}} and \eqn{\pmb{g}} are constructed from
a large database of age-specific demographic estimates
by preforming an SVD and then standardizing.
For details see TODO - REFERENCE TO VIGNETTE.
}

\section{Scaled SVDs in bage}{

\itemize{
\item \code{\link{HMD}} Mortality rates from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item \code{\link{LFP}} Labour forcce participation
rates from the \href{https://data-explorer.oecd.org}{OECD}.
}
}

\examples{
SVD(HMD)
SVD(HMD, n = 3)
}
\seealso{
\itemize{
\item \code{\link[=set_prior]{set_prior()}} Specify the prior for a model
\item \code{\link[=SVDS]{SVDS()}} SVD prior for interaction between
age and sex/gender.
\item \code{\link[=ESVD]{ESVD()}} Exchangeable SVD prior for interaction involving
age and other dimensions, but not sex/gender.
\item \code{\link[=ESVDS]{ESVDS()}} Exchangeable SVD prior for interaction involving
age, sex/gender, and other dimension(s).
}
}
