% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{SVDS}
\alias{SVDS}
\title{SVD Prior for Age-Sex or Age-Gender Interaction}
\usage{
SVDS(ssvd, n_comp = NULL, joint = FALSE)
}
\arguments{
\item{ssvd}{Object of class \code{"bage_ssvd"}
holding a scaled SVD. See below for scaled SVDs
of databases currently available in \strong{bage}.}

\item{n_comp}{Number of components from scaled SVD
to use in modelling. The default is half
the number of components of \code{ssvd}.}

\item{joint}{Whether to use combined or
separate SVDs. Default is \code{FALSE}.
See below for details.}
}
\value{
An object of class \code{"bage_prior_svd"}.
}
\description{
A Singular Value Decomposition (SVD) prior
for an interaction involving age-sex
profiles or age-gender profiles.
}
\details{
A \code{SVDS()} prior assumes that the age-sex profile
or age-gender profile for the quantity being
modelled looks like it was drawn at random
from an external demographic database. For instance,
the prior obtained via

\if{html}{\out{<div class="sourceCode">}}\preformatted{SVDS(HMD)
}\if{html}{\out{</div>}}

assumes that age-sex profiles look like
they were obtained from the
\href{https://www.mortality.org}{Human Mortality Database}.

If \code{SVDS()} is used with an interaction
between age, sex/gender, and one or more
other variables, then separate age-sex or
age-gender profiles are constructed within
each combination of the other variables.

Age profiles with no sex or gender dimension
should be modelled using \code{\link[=SVD]{SVD()}} rather than
\code{SVDS()}.
}
\section{Joint or independent SVDs}{


Two possible ways of extracting patterns
from age-sex-specific data are
\enumerate{
\item carry out separate SVDs on separate datasets for
each sex/gender; or
\item carry out a single SVD on dataset that has separate
entries for each sex/gender.
}

Option 1 is more flexible. Option 2 is
more robust to sampling or measurement errors.
Option 1 is obtained by setting the \code{joint}
argument  to \code{FALSE}. Option 2
is obtained by setting the \code{joint} argument to \code{TRUE}.
The default is \code{FALSE}.
}

\section{Mathematical details}{


\strong{Case 1: Interaction involving age and sex or gender, but no other variables}

When \code{SVDS()} is used with \code{joint = FALSE},

\deqn{\pmb{\beta}_{s} = \pmb{F}_s \pmb{\alpha}_{s} + \pmb{g}_s,}

and when \code{SVDS()} is used with \code{joint = TRUE},

\deqn{\pmb{\beta} = \pmb{F} \pmb{\alpha} + \pmb{g},}

where
\itemize{
\item \eqn{\pmb{\beta}} is an interaction involving age and sex/gender;
\item \eqn{\pmb{\beta}_{s}} is a subvector of \eqn{\pmb{\beta}},
holding values for sex/gender \eqn{s};
\item \eqn{J} is the number of elements in \eqn{\pmb{\beta}};
\item \eqn{S} is the number of sexes/genders;
\item \eqn{n} is \code{n_comp};
\item \eqn{\pmb{F}_s} is a known \eqn{(J/S) \times n} matrix, specific
to sex/gender \eqn{s};
\item \eqn{\pmb{g}_s} is a known vector with \eqn{J/S} elements,
specific to sex/gender \eqn{s};
\item \eqn{\pmb{F}} is a known \eqn{J \times n} matrix, with values
for all sexes/genders; and
\item \eqn{\pmb{g}} is a known vector with \eqn{J} elements, with values
for all sexes/genders.
}

The elements of \eqn{\pmb{\alpha}_s} and \eqn{\pmb{\alpha}} have prior
\deqn{\alpha_k \sim \text{N}(0, 1).}

\strong{Case 2: Interaction involving age and sex/gender, and one or more other variables}

When \code{SVDS()} is used with \code{joint = FALSE},

\deqn{\pmb{\beta}_{u,s} = \pmb{F}_s \pmb{\alpha}_{u,s} + \pmb{g}_s,}

and when \code{SVDS()} is used with \code{joint = TRUE},

\deqn{\pmb{\beta}_u = \pmb{F} \pmb{\alpha}_u + \pmb{g},}

where
\itemize{
\item \eqn{\pmb{\beta}} is an interaction involving sex/gender;
\item \eqn{\pmb{\beta}_{u,s}} is a subvector of \eqn{\pmb{\beta}},
holding values for sex/gender \eqn{s} for the \eqn{u}th
combination of the other variables;
\item \eqn{V} is the number of elements in \eqn{\pmb{\beta}_u};
\item \eqn{S} is the number of sexes/genders;
\item \eqn{n} is \code{n_comp};
\item \eqn{\pmb{F}_s} is a known \eqn{(V/S) \times n} matrix, specific
to sex/gender \eqn{s};
\item \eqn{\pmb{g}_s} is a known vector with \eqn{V/S} elements,
specific to sex/gender \eqn{s};
\item \eqn{\pmb{F}} is a known \eqn{V \times n} matrix, with values
for all sexes/genders; and
\item \eqn{\pmb{g}} is a known vector with \eqn{V} elements, with values
for all sexes/genders.
}

The elements of \eqn{\pmb{\alpha}_{u,s}} and \eqn{\pmb{\alpha}_u} have prior
\deqn{\alpha_k \sim \text{N}(0, 1).}
}

\section{Scaled SVDs of demographic databases in bage}{

\itemize{
\item \code{\link{HMD}} Mortality rates from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item \code{\link{LFP}} Labour forcce participation
rates from the \href{https://data-explorer.oecd.org}{OECD}.
}
}

\examples{
SVDS(HMD)
SVDS(HMD, joint = TRUE)
}
\references{
\itemize{
\item For details of the construction of
scaled SVDS see the vignette
\href{https://bayesiandemography.github.io/bage/articles/vig4_svd.html}{here}.
}
}
\seealso{
\itemize{
\item \code{\link[=SVD]{SVD()}} SVD prior for age, with no sex or gender
\item \code{\link[=SVDS_RW]{SVDS_RW()}} SVD prior for age and sex/gender with time-varying coeffients
\item \code{\link[=RW]{RW()}} Smoothing via random walk
\item \code{\link[=RW2]{RW2()}} Smoothing via random walk with drift
\item \code{\link[=Sp]{Sp()}} Smoothing via splines
\item \link{priors} Overview of priors implemented in \strong{bage}
\item \code{\link[=set_prior]{set_prior()}} Specify prior for intercept,
main effect, or interaction
}
}
