% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_prior-constructors.R
\name{SVDS}
\alias{SVDS}
\title{SVD Prior for Interaction Between Age and Sex/Gender}
\usage{
SVDS(ssvd, n = 5, joint = FALSE)
}
\arguments{
\item{ssvd}{Object of class \code{"bage_ssvd"}
holding results from a scaled SVD.
See below for current options.}

\item{n}{Number of vectors from scaled SVD
to use in modelling.
Default is half the number of components of \code{ssvd}.}

\item{joint}{Whether to use combined or
separate SVDs. Default is \code{FALSE}.
See below for details.}
}
\value{
An object of class \code{"bage_prior_svd"}.
}
\description{
Specify a Singular Value Decomposition (SVD) prior
for an interaction between age and sex/gender.
The age-sex profile is assumed to look like it
was drawn at random
from an external demographic database. For instance,

\if{html}{\out{<div class="sourceCode">}}\preformatted{set_prior(mod, age:sex ~ SVDS(HMD))
}\if{html}{\out{</div>}}

specifies that age-sex profile looks like it was drawn from the
\href{https://www.mortality.org}{Human Mortality Database}.
}
\section{Joint or independent SVDs}{


To possible ways of extracting patterns
from age-sex-specific data are
\enumerate{
\item carry out separate SVDs on separate datasets for
each sex/gender; or
\item carry out a single SVD on dataset that has separate
entries for each sex/gender.
}

Option 1 is more flexible. Option 2 is
more robust to sampling or measurement errors.
Option 1 is obtained by setting the \code{joint}
argument  to \code{FALSE}. Option 2
is obtained by setting the \code{joint} argument to \code{TRUE}.
The default is \code{FALSE}.
}

\section{Mathematical details}{


Let \eqn{A} denote the number of age groups, and \eqn{S} the
number of sexes/genders. When \code{joint} is \code{FALSE},

\deqn{\pmb{\beta}_s = \pmb{F}_s \pmb{\alpha}_s + \pmb{g}_s, \quad s = 1, \cdots S}

where \eqn{\pmb{\beta}_s}, \eqn{\pmb{\alpha}_s},
and \eqn{\pmb{g}_s} each have \eqn{A}
elements, and \eqn{\pmb{F}} has \eqn{A} rows.

When \code{joint} is \code{TRUE},

\deqn{\pmb{\beta} = \pmb{F} \pmb{\alpha} + \pmb{g}}

where

where \eqn{\pmb{\beta}}, \eqn{\pmb{\alpha}},
and \eqn{\pmb{g}} each have \eqn{A \times S}
elements, and \eqn{\pmb{F}_s} has \eqn{A \times S} rows.

\eqn{\pmb{F}_s}, \eqn{\pmb{g}_s}, \eqn{\pmb{F}}, and \eqn{\pmb{g}},
are all estimated externally, using estimates from a
demographic database, and are treated as fixed and known.

The elements of \eqn{\pmb{\alpha}_s} and \eqn{\pmb{\alpha}} have prior

\deqn{\alpha_k \sim \text{N}(0, 1)}

For a description of the construction of \eqn{\pmb{F}_s},
\eqn{\pmb{F}}, \eqn{\pmb{g}_s}, and \eqn{\pmb{g}},
see TODO - REFERENCE TO VIGNETTE.
}

\section{Scaled SVDs currently implemented in bage}{

\itemize{
\item \code{\link{HMD}} Values from the
\href{https://www.mortality.org}{Human Mortality Database}.
\item TODO - add others
}
}

\examples{
SVDS(HMD)
SVDS(HMD, joint = TRUE)
}
\seealso{
\itemize{
\item \code{\link[=set_prior]{set_prior()}} Specify the prior for a model.
\item \code{\link[=SVD]{SVD()}} SVD prior for age main effect.
\item \code{\link[=ESVD]{ESVD()}} Exchangeable SVD prior for interaction involving
age and other dimensions, but not sex/gender.
\item \code{\link[=ESVDS]{ESVDS()}} Exchangeable SVD prior for interaction involving
age, sex/gender, and other dimension(s).
}
}
