% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bage_mod-methods.R
\name{tidy.bage_mod}
\alias{tidy.bage_mod}
\title{Components from a fitted model}
\usage{
\method{tidy}{bage_mod}(x, ...)
}
\arguments{
\item{x}{A fitted \code{bage_mod} object.}

\item{...}{Unused. Included for generic consistency only.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Components from a fitted model
}
\examples{
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)
mod <- fit(mod)
tidy(mod)
}
\seealso{
\code{\link[=glimpse]{glimpse()}} provides less detailed information,
and \code{\link[=augment]{augment()}} provides more detailed.
}
