% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidier.R
\name{tidy.bage_mod}
\alias{tidy.bage_mod}
\title{Summarise model output via 'tidy'}
\usage{
tidy.bage_mod(x, ...)
}
\arguments{
\item{x}{A \code{bage_sysmod} object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Provide information on a model, including
overall performance, and individual-level
terms.
}
\examples{
mod <- mod_pois(injuries ~ age + sex + time,
                data = injuries,
                exposure = popn) \%>\%
  fit()

tidy(mod)
}
\seealso{
\code{\link[=glimpse]{glimpse()}} provides less detailed information,
and \code{\link[=augment]{augment()}} provides more detailed.
}
